% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{abcd_import}
\alias{abcd_import}
\title{Import a raw ABCD dataframe for selected subjects at baseline}
\usage{
abcd_import(abcd_df, subjects)
}
\arguments{
\item{abcd_df}{An ABCD dataframe}

\item{subjects}{Dataframe containing list of required subjects}
}
\value{
abcd_clean_df The subsetted dataframe
}
\description{
Import a raw ABCD dataframe for selected subjects at baseline
}
