% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_char_vec_as_cols}
\alias{add_char_vec_as_cols}
\title{Add character vector as columns}
\usage{
add_char_vec_as_cols(df, char_vector, filler)
}
\arguments{
\item{df}{The dataframe to extend}

\item{char_vector}{The vector containing new column names}

\item{filler}{The values of the elements of the newly added columns}
}
\value{
extended_df The dataframe containing the added columns
}
\description{
Extend a dataframe with columns from a character vector
}
