% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_columns}
\alias{add_columns}
\title{Add columns to a dataframe}
\usage{
add_columns(df, cols, value = NA)
}
\arguments{
\item{df}{The dataframe to extend.}

\item{cols}{The vector containing new column names.}

\item{value}{The values stored in the newly added columns. NA by default.}
}
\value{
A data frame containing with the same columns as the \code{df} argument
as well as the new columns specified in the \code{cols} argument.
}
\description{
Add new columns to a dataframe by specifying their names and a value to
initialize them with.
}
\keyword{internal}
