% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_matrix.R
\name{add_design_matrix_rows}
\alias{add_design_matrix_rows}
\title{Add design matrix rows}
\usage{
add_design_matrix_rows(
  design_matrix,
  nrows,
  min_removed = NULL,
  max_removed = NULL,
  retry_limit = 10
)
}
\arguments{
\item{design_matrix}{The existing design matrix}

\item{nrows}{The number of rows to be added to the design matrix}

\item{min_removed}{The smallest number of elements that may be removed}

\item{max_removed}{The largest number of elements that may be removed}

\item{retry_limit}{The maximum number of attempts to generate a novel row}
}
\value{
design_matrix New design matrix containing additional rows
}
\description{
Add design matrix rows
}
