% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{affinity_matrix_path}
\alias{affinity_matrix_path}
\title{Generate a complete path and filename to store an affinity matrix}
\usage{
affinity_matrix_path(affinity_matrix_dir, i)
}
\arguments{
\item{affinity_matrix_dir}{Directory to store affinity matrices}

\item{i}{Corresponding settings matrix row}
}
\value{
path Complete path and filename to store an affinity matrix
}
\description{
Generate a complete path and filename to store an affinity matrix
}
