% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_df}
\alias{age_df}
\title{Mock age data}
\format{
\subsection{\code{age_df}}{

A data frame with 200 rows and 2 columns:
\describe{
\item{patient_id}{Random three-digit number uniquely identifying the patient}
\item{age}{Mock age feature}
}
}
}
\source{
This data came from the SNFtool package, with slight modifications.
}
\usage{
age_df
}
\description{
Mock age data
}
\keyword{datasets}
