% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ak_heatmap}
\alias{ak_heatmap}
\title{Heatmap alpha and k hyperparameter results by minimum and mean p-values}
\usage{
ak_heatmap(ak_scan_om, min_or_mean = "min", save_path = NULL)
}
\arguments{
\item{ak_scan_om}{output matrix containing alpha, K, and min/mean p-values}

\item{min_or_mean}{string specifying whether min or mean p-value should be
visualized}

\item{save_path}{optional path to save figure to}
}
\description{
Heatmap alpha and k hyperparameter results by minimum and mean p-values
}
