% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ak_scan_plot}
\alias{ak_scan_plot}
\title{Scatter plot alpha and k hyperparameter results by minimum and mean p-values}
\usage{
ak_scan_plot(ak_scan_om, a_or_k = "k")
}
\arguments{
\item{ak_scan_om}{output matrix containing alpha, K, and min/mean p-values}

\item{a_or_k}{string specifying whether alpha or K should be visualized}
}
\description{
Scatter plot alpha and k hyperparameter results by minimum and mean p-values
}
