% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_cluster.R
\name{ari_heatmap}
\alias{ari_heatmap}
\title{Heatmap meta-clustering results}
\usage{
ari_heatmap(output_matrix_aris, save = NULL)
}
\arguments{
\item{output_matrix_aris}{results from meta_cluster function}

\item{save}{optional path to save figure to}
}
\description{
Heatmap meta-clustering results
}
