% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{arrange_dl}
\alias{arrange_dl}
\title{Sort data frames in a data list by their unique ID values.}
\usage{
arrange_dl(dl)
}
\arguments{
\item{dl}{A data list-like object.}
}
\value{
arranged_dl The data list-like object with all data frames sorted
by uid.
}
\description{
Sort data frames in a data list by their unique ID values.
}
\keyword{internal}
