% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{arrange_dll}
\alias{arrange_dll}
\title{Sort data frames in a data list by their unique ID values}
\usage{
arrange_dll(dll)
}
\arguments{
\item{dll}{A data list-like \code{list} class object.}
}
\value{
The data list-like object with all data frames sorted by uid.
}
\description{
Sort data frames in a data list by their unique ID values
}
\keyword{internal}
