% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data_frame.R
\name{as.data.frame.data_list}
\alias{as.data.frame.data_list}
\title{Coerce a \code{data_list} class object into a \code{data.frame} class object}
\usage{
\method{as.data.frame}{data_list}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A \code{data_list} class object.}

\item{row.names}{Additional parameter passed to \code{as.data.frame()}.}

\item{optional}{Additional parameter passed to \code{as.data.frame()}.}

\item{...}{Additional parameter passed to \code{as.data.frame()}.}
}
\value{
dl_df A \code{data.frame} class object with all the features and
observations of \code{dl}.
}
\description{
Horizontally joins data frames within a data list into a single data frame,
using the \code{uid} attribute as the joining key.
}
