% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data_frame.R
\name{as.data.frame.ext_solutions_df}
\alias{as.data.frame.ext_solutions_df}
\title{Coerce a \code{ext_solutions_df} class object into a \code{data.frame} class object}
\usage{
\method{as.data.frame}{ext_solutions_df}(
  x,
  row.names = NULL,
  optional = FALSE,
  keep_attributes = FALSE,
  ...
)
}
\arguments{
\item{x}{A \code{ext_solutions_df} class object.}

\item{row.names}{Additional parameter passed to \code{as.data.frame()}.}

\item{optional}{Additional parameter passed to \code{as.data.frame()}.}

\item{keep_attributes}{If TRUE, resulting data frame includes settings
data frame and weights matrix.}

\item{...}{Additional parameter passed to \code{as.data.frame()}.}
}
\value{
A \code{data.frame} class object with all the columns of x and its
contained solutions data frame.
}
\description{
Coerce a \code{ext_solutions_df} class object into a \code{data.frame} class object
}
