% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data_frame.R
\name{as.data.frame.snf_config}
\alias{as.data.frame.snf_config}
\title{Coerce a \code{settings_df} class object into a \code{data.frame} class object}
\usage{
\method{as.data.frame}{snf_config}(x, ...)
}
\arguments{
\item{x}{A \code{settings_df} class object.}

\item{...}{Additional parameter passed to \code{as.data.frame()}.}
}
\value{
A \code{data.frame} class object with all the columns of x and its
contained solutions data frame.
}
\description{
Coerce a \code{settings_df} class object into a \code{data.frame} class object
}
