% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_list.R
\name{as.list.data_list}
\alias{as.list.data_list}
\title{Coerce a \code{data_list} class object into a \code{list} class object}
\usage{
\method{as.list}{data_list}(x, ...)
}
\arguments{
\item{x}{A \code{data_list} class object.}

\item{...}{Additional parameter passed to \code{as.list()}.}
}
\value{
A \code{list} class object with all the objects of \code{x}.
}
\description{
Coerce a \code{data_list} class object into a \code{list} class object
}
