% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_list.R
\name{as.list.dist_fns_list}
\alias{as.list.dist_fns_list}
\title{Coerce a \code{dist_fns_list} class object into a \code{list} class object}
\usage{
\method{as.list}{dist_fns_list}(x, ...)
}
\arguments{
\item{x}{A \code{dist_fns_list} class object.}

\item{...}{Additional parameter passed to \code{as.list()}.}
}
\value{
A \code{list} class object with all the functions of \code{x}.
}
\description{
Coerce a \code{dist_fns_list} class object into a \code{list} class object
}
