% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_list.R
\name{as.list.snf_config}
\alias{as.list.snf_config}
\title{Coerce a \code{snf_config} class object into a \code{list} class object}
\usage{
\method{as.list}{snf_config}(x, ...)
}
\arguments{
\item{x}{A \code{snf_config} class object.}

\item{...}{Additional parameter passed to \code{as.list()}.}
}
\value{
A \code{list} class object with all the functions of \code{x}.
}
\description{
Coerce a \code{snf_config} class object into a \code{list} class object
}
