% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_matrix.R
\name{as.matrix.ari_matrix}
\alias{as.matrix.ari_matrix}
\title{Coerce a \code{ari_matrix} class object into a \code{matrix} class object}
\usage{
\method{as.matrix}{ari_matrix}(x, ...)
}
\arguments{
\item{x}{A \code{ari_matrix} class object.}

\item{...}{Additional parameter passed to \code{as.matrix()}.}
}
\value{
A \code{matrix} and \code{array} class object.
}
\description{
Coerce a \code{ari_matrix} class object into a \code{matrix} class object
}
