% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_ari_matrix.R
\name{as_ari_matrix}
\alias{as_ari_matrix}
\title{Convert an object to an ARI matrix}
\usage{
as_ari_matrix(x)
}
\arguments{
\item{x}{The object to convert into a weights matrix.}
}
\value{
An \code{ari_matrix} class object.
}
\description{
This function coerces non-\code{ari_matrix} class objects into
\code{ari_matrix} class objects.
}
