% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data_list.R
\name{as_data_list}
\alias{as_data_list}
\title{Convert an object to a data list}
\usage{
as_data_list(x)
}
\arguments{
\item{x}{The object to convert into a data list.}
}
\value{
A \code{data_list} class object.
}
\description{
This function coerces non-\code{data_list} class objects into \code{data_list} class
objects.
}
