% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_settings_df.R
\name{as_settings_df}
\alias{as_settings_df}
\title{Convert an object to a settings data frame}
\usage{
as_settings_df(x)
}
\arguments{
\item{x}{The object to convert into a data list.}
}
\value{
A \code{settings_df} class object.
}
\description{
This function coerces non-\code{settings_df} class objects into \code{settings_df} class
objects.
}
