% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_sim_mats_list.R
\name{as_sim_mats_list}
\alias{as_sim_mats_list}
\title{Convert an object to a similarity matrix list}
\usage{
as_sim_mats_list(x)
}
\arguments{
\item{x}{The object to convert into a \code{sim_mats_list}. Must be a list of
square matrices with identical column and row names.}
}
\value{
A \code{sim_mats_list} class object.
}
\description{
This function converts non-\code{sim_mats_list} class objects into
\code{sim_mats_list} class objects.
}
