% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_snf_config.R
\name{as_snf_config}
\alias{as_snf_config}
\title{Convert an object to a snf config}
\usage{
as_snf_config(x)
}
\arguments{
\item{x}{The object to convert into a snf config.}
}
\value{
A \code{snf_config} class object.
}
\description{
This function coerces non-\code{snf_config} class objects into \code{snf_config} class
objects.
}
