% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_type.R, R/weights_matrix.R
\name{as_weights_matrix}
\alias{as_weights_matrix}
\title{Convert an object to a weights matrix}
\usage{
as_weights_matrix(x)

as_weights_matrix(x)
}
\arguments{
\item{x}{The object to convert into a weights matrix.}
}
\value{
A \code{weights_matrix} class object.

A \code{weights_matrix} class object.
}
\description{
This function converts non-\code{weights_matrix} objects into \code{weights_matrix}
class objects.

This function coerces non-\code{weights_matrix} class objects into
\code{weights_matrix} class objects.
}
