% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{association_heatmap}
\alias{association_heatmap}
\title{(DEPRECATED) Association p-value heatmap}
\usage{
association_heatmap(...)
}
\arguments{
\item{...}{Function arguments}
}
\description{
(DEPRECATED) Association p-value heatmap
}
