% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_plots.R
\name{bar_plot}
\alias{bar_plot}
\title{Bar plot separating a feature by cluster}
\usage{
bar_plot(df, feature)
}
\arguments{
\item{df}{A data.frame containing cluster column and the feature to plot.}

\item{feature}{The feature to plot.}
}
\value{
A bar plot (class "gg", "ggplot") showing the distribution of a
feature across clusters.
}
\description{
Bar plot separating a feature by cluster
}
