% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_snf.R
\name{batch_snf}
\alias{batch_snf}
\title{Run variations of SNF as described by a settings matrix}
\usage{
batch_snf(
  data_list,
  settings_matrix,
  processes = 1,
  affinity_matrix_dir = NULL,
  run_clustering = TRUE
)
}
\arguments{
\item{data_list}{nested list of input data generated by the function
\code{get_data_list()}}

\item{settings_matrix}{matrix indicating parameters to iterate SNF through}

\item{processes}{Specify number of processes used to complete SNF iterations
* \code{1} (default) Sequential processing: function will iterate through the
\code{settings_matrix} one row at a time with a for loop. This option will
not make use of multiple CPU cores, but will show a progress bar.
* \code{2} or higher: Parallel processing will use the
\code{future.apply::future_apply} to distribute the SNF iterations across
the specified number of CPU cores. If higher than the number of
available cores, a warning will be printed and the maximum number of
cores will be used.
* \code{max}: All available cores will be used.}

\item{affinity_matrix_dir}{If specified, this directory will be used to save
all generated affinity matrices}

\item{run_clustering}{If TRUE (default), will apply default or custom
clustering algorithms to provide cluster solutions on every iteration of
SNF. If FALSE, parameter \code{affinity_matrix_dir} must be specified to proceed.}
}
\value{
populated_settings_matrix settings matrix with filled columns related to
subtype membership
}
\description{
Run variations of SNF as described by a settings matrix
}
