% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coclustering.R
\name{batch_snf_subsamples}
\alias{batch_snf_subsamples}
\title{Run SNF clustering pipeline on a list of subsampled data lists.}
\usage{
batch_snf_subsamples(
  data_list_subsamples,
  settings_matrix,
  processes = 1,
  return_similarity_matrices = FALSE,
  clust_algs_list = NULL,
  suppress_clustering = FALSE,
  distance_metrics_list = NULL,
  weights_matrix = NULL,
  automatic_standard_normalize = FALSE,
  return_solutions_matrices = FALSE
)
}
\arguments{
\item{data_list_subsamples}{A list of subsampled data lists. This object is
generated by the function \code{batch_snf_subsamples()}.}

\item{settings_matrix}{A settings matrix defining the parameters of the SNF
pipelines to be applied to the subsampled data lists.}

\item{processes}{See ?batch_snf.}

\item{return_similarity_matrices}{See ?batch_snf.}

\item{clust_algs_list}{See ?batch_snf.}

\item{suppress_clustering}{See ?batch_snf.}

\item{distance_metrics_list}{See ?batch_snf.}

\item{weights_matrix}{See ?batch_snf.}

\item{automatic_standard_normalize}{See ?batch_snf.}

\item{return_solutions_matrices}{If TRUE, includes the solutions matrices
corresponding to each subsample in the output.}
}
\description{
Run SNF clustering pipeline on a list of subsampled data lists.
}
