% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf.R
\name{build_design_matrix_ak}
\alias{build_design_matrix_ak}
\title{Build design matrix for scanning alpha and K}
\usage{
build_design_matrix_ak()
}
\value{
design_matrix Adds the standard grid expansion for SNF hyperparams
}
\description{
Build design matrix for scanning alpha and K
}
