% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf.R
\name{build_design_matrix_base}
\alias{build_design_matrix_base}
\title{Build a design matrix skeleton}
\usage{
build_design_matrix_base()
}
\value{
design_matrix A skeleton dataframe to build a design matrix out of
}
\description{
Construct the base of the design matrix, a dataframe that initially contains
rows that each provide information about the hyperparamters with which
patient subtyping should occur for a single iteration. After executing the
subtyping specified by a given row, that row is extended with detailed
information about the resulting subtype membership of each included subject
and the efficacy of the subtyping solution in distinguishing subjects by
PPCS.
}
