% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf.R
\name{build_design_matrix_base_t2}
\alias{build_design_matrix_base_t2}
\title{Build a design matrix skeleton FOR TESTING 2 - Contains psychosocial vars}
\usage{
build_design_matrix_base_t2()
}
\value{
design_matrix A skeleton dataframe to build a design matrix out of
}
\description{
Construct the base of the design matrix for small testing purposes
}
