% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf.R
\name{build_design_matrix_s}
\alias{build_design_matrix_s}
\title{Build standard design matrix}
\usage{
build_design_matrix_s()
}
\value{
design_matrix Adds the standard grid expansion for SNF hyperparams
}
\description{
Build standard design matrix
}
