% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf.R
\name{build_design_matrix_soft}
\alias{build_design_matrix_soft}
\title{Build a design matrix - but softcoded}
\usage{
build_design_matrix_soft(data_list, nrows = 0, seed = NULL, retry_limit = 10)
}
\arguments{
\item{data_list}{a data list object to determine variables for inclusion/exclusion}

\item{nrows}{number of design matrix rows}

\item{seed}{set a seed for the random matrix generation. Note that this}

\item{retry_limit}{The maximum number of attempts to generate a novel row
affects the global seed.}
}
\value{
design_matrix A design matrix
}
\description{
Build a design matrix - but softcoded
}
