% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_matrix.R
\name{build_output_matrix}
\alias{build_output_matrix}
\title{Build output matrix}
\usage{
build_output_matrix(data_list, design_matrix)
}
\arguments{
\item{data_list}{nested list of input data generated by the function
\code{get_data_list()}}

\item{design_matrix}{matrix indicating parameters to iterate SNF through}
}
\value{
output_matrix The output matrix
}
\description{
The matrix that will contain all the columns to be populated by SNF results
}
