% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{c.data_list}
\alias{c.data_list}
\title{Combine an arbitrary number of data lists into one data list}
\usage{
\method{c}{data_list}(...)
}
\arguments{
\item{...}{Data lists to be concatenated.}
}
\value{
A data list made of the concatenated inputs.
}
\description{
Horizontally joins data frames within a data list into a single data frame,
using the \code{uid} attribute as the joining key.
}
