% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cache_a_complete_example_ext_sol_df}
\alias{cache_a_complete_example_ext_sol_df}
\title{Cached example extended solutions data frame}
\format{
\subsection{\code{cache_a_complete_example_ext_sol_df}}{

Contains 20 cluster solutions, 87 observations, and p-values for 336 features.
}
}
\source{
This data came from the metasnf package.
}
\usage{
cache_a_complete_example_ext_sol_df
}
\description{
An extended solutions data frame used as a cached example in the "a_complete_example.Rmd" vignette.
}
\keyword{datasets}
