% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cache_a_complete_example_sol_df}
\alias{cache_a_complete_example_sol_df}
\title{Cached example solutions data frame}
\format{
\subsection{\code{cache_a_complete_example_sol_df}}{

A solutions data frame with 20 cluster solutions and 87 observations.
}
}
\source{
This data came from the metasnf package.
}
\usage{
cache_a_complete_example_sol_df
}
\description{
An solutions data frame used as a cached example in the "a_complete_example.Rmd" vignette.
}
\keyword{datasets}
