% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_cluster.R
\name{calc_ari}
\alias{calc_ari}
\title{Calculate adjusted rand index between two rows of solutions matrix}
\usage{
calc_ari(r1, r2, df)
}
\arguments{
\item{r1}{first row of cluster labels}

\item{r2}{second row of cluster labels}

\item{df}{dataframe containing cluster labels}
}
\value{
ari adjusted rand index
}
\description{
Output matrix must have no row_id or non-sub columns
}
