% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_cluster.R
\name{calc_ari_progressr}
\alias{calc_ari_progressr}
\title{Wrapper for mclust::adjustedRandIndex compatible with progressr}
\usage{
calc_ari_progressr(v1, v2, p)
}
\arguments{
\item{v1}{First vector of cluster memberships}

\item{v2}{Second vector of cluster memberships}

\item{p}{Progressr function to update parallel processing progress}
}
\description{
Wrapper for mclust::adjustedRandIndex compatible with progressr
}
