% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_characterization.R
\name{calc_nmi}
\alias{calc_nmi}
\title{Calculate NMI scores for SNF inputs}
\usage{
calc_nmi(om_row, data_list)
}
\arguments{
\item{om_row}{An output matrix row}

\item{data_list}{A data list}
}
\value{
nmi_df A dataframe containing input NMI scores in descending order
}
\description{
Calculate NMI scores for SNF inputs
}
