% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_cluster.R
\name{calc_om_aris}
\alias{calc_om_aris}
\title{Meta-cluster calculations}
\usage{
calc_om_aris(om, progress = FALSE)
}
\arguments{
\item{om}{solutions_matrix}

\item{progress}{If TRUE, wrapping the function with progressr (i.e.,
progressr::with_progress({calc_om_aris(om)}) will reveal a progress bar.
The progress bar slows down the computation considerably.}
}
\value{
om_aris ARIs between clustering solutions of an solutions matrix
}
\description{
Generate matrix of pairwise cluster-solution similarities by Adjusted Rand
index calculations.
}
