% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_measures.R
\name{calculate_dunn_indices}
\alias{calculate_dunn_indices}
\title{Calculate Dunn indices}
\usage{
calculate_dunn_indices(sol_df)
}
\arguments{
\item{sol_df}{A sol_df (see ?batch_snf)}
}
\value{
dunn_indices A vector of Dunn indices for each cluster solution
}
\description{
Given a sol_df and a list of similarity_matrices (or a single
similarity_matrix if the sol_df has only 1 row), return vector of
Dunn indices
}
