% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_measures.R
\name{calculate_silhouettes}
\alias{calculate_silhouettes}
\title{Calculate silhouette scores}
\usage{
calculate_silhouettes(sol_df)
}
\arguments{
\item{sol_df}{A \code{solutions_df} class object created by  \code{batch_snf()}.}
}
\value{
silhouette_scores A list of "silhouette" objects from the cluster
package.
}
\description{
Given a solutions data frame containing a similarity matrix list, return a
list of 'silhouette' objects from the cluster package
}
