% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cancer_diagnosis_df}
\alias{cancer_diagnosis_df}
\title{Mock diagnosis data}
\format{
\subsection{\code{cancer_diagnosis_df}}{

A data frame with 200 rows and 2 columns:
\describe{
\item{patient_id}{Random three-digit number uniquely identifying the patient}
\item{diagnosis}{Mock cancer diagnosis feature (1, 2, or 3)}
}
}
}
\source{
This data came from the SNFtool package, with slight modifications.
}
\usage{
cancer_diagnosis_df
}
\description{
This is the same data as \code{diagnosis_df}, with renamed features and columns.
}
\keyword{datasets}
