% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{cat_colours}
\alias{cat_colours}
\title{Helper function for generating categorical colour palette}
\usage{
cat_colours(vector, palette)
}
\arguments{
\item{vector}{Vector of categorical data to generate palette for.}

\item{palette}{Which RColorBrewer palette should be used.}
}
\value{
A named list of colours where the names correspond to the unique
values of vector and the values correspond to their colours.
}
\description{
Helper function for generating categorical colour palette
}
\keyword{internal}
