% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterization.R
\name{cbcl_anova}
\alias{cbcl_anova}
\title{Calculate anova p-values for a characterization_df}
\usage{
cbcl_anova(characterization_df, bonferroni = FALSE)
}
\arguments{
\item{characterization_df}{A merged list containing cluster, subjectkey, and
various CBCL outcomes}

\item{bonferroni}{boolean for reporting bonferroni corrected p-values}
}
\description{
Calculate anova p-values for a characterization_df
}
