% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_characterization.R
\name{cbcl_ord_reg}
\alias{cbcl_ord_reg}
\title{Calculate overall p-values for a characterization_df}
\usage{
cbcl_ord_reg(characterization_df)
}
\arguments{
\item{characterization_df}{A merged list containing cluster, subjectkey, and
various CBCL outcomes}
}
\description{
Calculate overall p-values for a characterization_df
}
