% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_characterization.R
\name{cbcl_ord_reg_from_om}
\alias{cbcl_ord_reg_from_om}
\title{Calculate overall p-values for an om_row}
\usage{
cbcl_ord_reg_from_om(om_row, cbcl_list, bonferroni = FALSE)
}
\arguments{
\item{om_row}{a row of an output matrix}

\item{cbcl_list}{a list of CBCL measures}

\item{bonferroni}{boolean for reporting bonferroni corrected p-values}
}
\description{
Calculate overall p-values for an om_row
}
