% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{cell_significance_fn}
\alias{cell_significance_fn}
\title{Place significance stars on ComplexHeatmap cells.}
\usage{
cell_significance_fn(data)
}
\arguments{
\item{data}{The matrix containing the cells to base the significance stars
on.}
}
\value{
cell_fn Another function that is well-formatted for usage as the
cell_fun argument in ComplexHeatmap::Heatmap.
}
\description{
This is an internal function meant to be used to by the
assoc_pval_heatmap function.
}
