% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{char_to_fac}
\alias{char_to_fac}
\title{Convert char columns to factors}
\usage{
char_to_fac(df)
}
\arguments{
\item{df}{The dataframe containing char columns to be converted}
}
\value{
df The dataframe with factor columns
}
\description{
Convert char columns to factors
}
