% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_affinity_matrices}
\alias{check_affinity_matrices}
\title{Check validity of affinity matrices}
\usage{
check_affinity_matrices(affinity_matrices)
}
\arguments{
\item{affinity_matrices}{A list of affinity matrices}
}
\value{
valid_matrices Boolean indicating if properties are met by all
affinity matrices
}
\description{
Check to see if affinity matrices in a list have the following properties:
\enumerate{
\item The maximum value in the entire matrix is 0.5
\item Every value in the diagonal is 0.5
}
}
