% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_fns_list.R
\name{check_cfll_fn_args}
\alias{check_cfll_fn_args}
\title{Check if functions in a distance metrics list-like have valid arguments}
\usage{
check_cfll_fn_args(cfll)
}
\value{
Doesn't return any value. Raises error if the functions in dfll
don't have valid arguments.
}
\description{
Check if functions in a distance metrics list-like have valid arguments
}
\keyword{internal}
