% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_fns_list.R
\name{check_cfll_fns}
\alias{check_cfll_fns}
\title{Check if items of a clustering functions list-like object are functions}
\usage{
check_cfll_fns(cfll)
}
\arguments{
\item{cfll}{A clust_fns_list-like \code{list} class object.}
}
\value{
Doesn't return any value. Raises error if the items of cfll are
not functions.
}
\description{
Check if items of a clustering functions list-like object are functions
}
\keyword{internal}
