% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_fns_list.R
\name{check_cfll_named}
\alias{check_cfll_named}
\title{Check if clustering functions list-like object has named algorithms}
\usage{
check_cfll_named(cfll)
}
\arguments{
\item{cfll}{A clust_fns_list-like \code{list} class object.}
}
\value{
Doesn't return any value. Raises error if there are unnamed
clustering functions in cfll.
}
\description{
Check if clustering functions list-like object has named algorithms
}
\keyword{internal}
