% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_fns_list.R
\name{check_cfll_unique_names}
\alias{check_cfll_unique_names}
\title{Check if names in a clustering functions list-like object are unique}
\usage{
check_cfll_unique_names(cfll)
}
\arguments{
\item{cfll}{A clust_fns_list-like \code{list} class object.}
}
\value{
Doesn't return any value. Raises error if the names in cfll aren't
unique.
}
\description{
Check if names in a clustering functions list-like object are unique
}
\keyword{internal}
