% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_snf.R
\name{check_compatible_sdf_cfl}
\alias{check_compatible_sdf_cfl}
\title{Check if settings_df exceeds bounds of clust_fns_list}
\usage{
check_compatible_sdf_cfl(sdf, cfl)
}
\arguments{
\item{sdf}{A \code{settings_df} class object.}

\item{cfl}{A \code{clust_fns_list} class object.}
}
\value{
Doesn't return any value. Raises error if sdf calls for a clustering
function outside the range of cfl.
}
\description{
Check if settings_df exceeds bounds of clust_fns_list
}
\keyword{internal}
