% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_snf.R
\name{check_compatible_sdf_dfl}
\alias{check_compatible_sdf_dfl}
\title{Check if settings_df exceeds bounds of dist_fns_list}
\usage{
check_compatible_sdf_dfl(sdf, dfl)
}
\arguments{
\item{sdf}{A \code{settings_df} class object.}

\item{dfl}{A \code{dist_fns_list} class object.}
}
\value{
Doesn't return any value. Raises error if sdf calls for a distance
function outside the range of dfl.
}
\description{
Check if settings_df exceeds bounds of dist_fns_list
}
\keyword{internal}
