% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_snf.R
\name{check_compatible_sdf_wm}
\alias{check_compatible_sdf_wm}
\title{Check if settings_df and weights_matrix have same number of rows}
\usage{
check_compatible_sdf_wm(sdf, wm)
}
\arguments{
\item{sdf}{A \code{settings_df} class object.}

\item{wm}{A \code{weights_matrix} class object.}
}
\value{
Doesn't return any value. Raises error if sdf and wm don't have the
same number of rows.
}
\description{
Check if settings_df and weights_matrix have same number of rows
}
\keyword{internal}
