% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{check_dataless_annotations}
\alias{check_dataless_annotations}
\title{Helper function to stop annotation building when no data was provided}
\usage{
check_dataless_annotations(annotation_requests, data)
}
\arguments{
\item{annotation_requests}{A list of requested annotations}

\item{data}{A data frame with data to build annotations}
}
\value{
Does not return any value. This function just raises an error when
annotations are requested without any provided data for a heatmap.
}
\description{
Helper function to stop annotation building when no data was provided
}
