% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_fns_list.R
\name{check_dfll_fn_args}
\alias{check_dfll_fn_args}
\title{Check if functions in a distance metrics list-like have valid arguments}
\usage{
check_dfll_fn_args(dfll)
}
\arguments{
\item{dfll}{A distance metrics list-like list object to be validated.}
}
\value{
Doesn't return any value. Raises error if the functions in dfll
don't have valid arguments.
}
\description{
Check if functions in a distance metrics list-like have valid arguments
}
\keyword{internal}
