% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_fns_list.R
\name{check_dfll_fn_names}
\alias{check_dfll_fn_names}
\title{Check if functions in a distance metrics list-like have names}
\usage{
check_dfll_fn_names(dfll)
}
\arguments{
\item{dfll}{A distance metrics list-like list object to be validated.}
}
\value{
Doesn't return any value. Raises error if the functions in dfll
don't have names.
}
\description{
Check if functions in a distance metrics list-like have names
}
\keyword{internal}
