% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_fns_list.R
\name{check_dfll_item_names}
\alias{check_dfll_item_names}
\title{Check if items of a distance metrics list-like object have valid names}
\usage{
check_dfll_item_names(dfll)
}
\arguments{
\item{dfll}{A distance metrics list-like list object to be validated.}
}
\value{
Doesn't return any value. Raises error if the items of dfll don't
have valid formatted names.
}
\description{
Check if items of a distance metrics list-like object have valid names
}
\keyword{internal}
