% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_fns_list.R
\name{check_dfll_unique_names}
\alias{check_dfll_unique_names}
\title{Check if names in a distance metrics list-like object are unique}
\usage{
check_dfll_unique_names(dfll)
}
\arguments{
\item{dfll}{A distance metrics list-like list object to be validated.}
}
\value{
Doesn't return any value. Raises error if the items of dfll aren't
unique across layer 1 or within each item of layer 2.
}
\description{
Check if names in a distance metrics list-like object are unique
}
\keyword{internal}
