% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{check_dl_duplicate_features}
\alias{check_dl_duplicate_features}
\title{Check if data list contains any duplicate features}
\usage{
check_dl_duplicate_features(x)
}
\arguments{
\item{x}{A nested list of input data from \code{data_list()}.}
}
\value{
Doesn't return any value. Raises warning if there are features
with duplicate names in a generated data list.
}
\description{
Check if data list contains any duplicate features
}
\keyword{internal}
