% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{check_dl_types}
\alias{check_dl_types}
\title{Error if data list-like structure has invalid feature types}
\usage{
check_dl_types(x)
}
\arguments{
\item{x}{Object (list) in the process of being converted to a data list.}
}
\value{
Raises an error if the loaded types are not among continuous,
discrete, ordinal, categorical, or mixed.
}
\description{
Error if data list-like structure has invalid feature types
}
\keyword{internal}
