% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{check_dl_uid}
\alias{check_dl_uid}
\title{Check if UID columns in a nested list have valid structure for a data list}
\usage{
check_dl_uid(dl)
}
\arguments{
\item{dl}{A data list or data list-like list object to be checked.}
}
\value{
Raises an error if the UID columns do not have a valid structure.
}
\description{
Check if UID columns in a nested list have valid structure for a data list
}
\keyword{internal}
