% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{check_dll_duplicate_components}
\alias{check_dll_duplicate_components}
\title{Check if data list contains any duplicate names}
\usage{
check_dll_duplicate_components(dll)
}
\arguments{
\item{dll}{A data list-like \code{list} class object.}
}
\value{
Doesn't return any value. Raises error if there are features with
duplicate names in a generated data list.
}
\description{
Check if data list contains any duplicate names
}
\keyword{internal}
