% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{check_dll_duplicate_features}
\alias{check_dll_duplicate_features}
\title{Check if data list contains any duplicate features}
\usage{
check_dll_duplicate_features(dll)
}
\arguments{
\item{dll}{A data list-like \code{list} class object.}
}
\value{
Doesn't return any value. Raises error if there are features with
duplicate names in a generated data list.
}
\description{
Check if data list contains any duplicate features
}
\keyword{internal}
