% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{check_dll_empty_input}
\alias{check_dll_empty_input}
\title{Error if empty input provided during data list initalization}
\usage{
check_dll_empty_input(data_list_input)
}
\arguments{
\item{data_list_input}{Input data provided for data list initialization.}
}
\value{
Raises an error if data_list_input has 0 length.
}
\description{
Error if empty input provided during data list initalization
}
\keyword{internal}
