% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{check_dll_four_subitems}
\alias{check_dll_four_subitems}
\title{Error if data list-like list doesn't have only 4-item nested lists}
\usage{
check_dll_four_subitems(dll)
}
\arguments{
\item{dll}{A data list-like \code{list} class object.}
}
\value{
Raises error if dll doesn't have only 4-item nested lists
}
\description{
Error if data list-like list doesn't have only 4-item nested lists
}
\keyword{internal}
