% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{check_dll_inherits_list}
\alias{check_dll_inherits_list}
\title{Error if data list-like structure isn't a list}
\usage{
check_dll_inherits_list(dll)
}
\arguments{
\item{dll}{A data list-like \code{list} class object.}
}
\value{
Raises error if data list-like structure isn't a list
}
\description{
Error if data list-like structure isn't a list
}
\keyword{internal}
