% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{check_dll_subitem_names}
\alias{check_dll_subitem_names}
\title{Check valid subitem names for a data list-like list}
\usage{
check_dll_subitem_names(dll)
}
\arguments{
\item{dll}{A data list-like \code{list} class object.}
}
\value{
Raises error if dll doesn't have only 4-item nested lists
}
\description{
Error if data list-like structure doesn't have nested names of "data",
"name", "domain", and "type".
}
\keyword{internal}
