% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{check_dll_types}
\alias{check_dll_types}
\title{Error if data list-like structure has invalid feature types}
\usage{
check_dll_types(dll)
}
\arguments{
\item{dll}{A data list-like \code{list} class object.}
}
\value{
Raises an error if the loaded types are not among continuous,
discrete, ordinal, categorical, or mixed.
}
\description{
Error if data list-like structure has invalid feature types
}
\keyword{internal}
