% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{check_dll_uid}
\alias{check_dll_uid}
\title{Check if UID columns in a nested list have valid structure for a data list}
\usage{
check_dll_uid(dll)
}
\arguments{
\item{dll}{A data list-like \code{list} class object.}
}
\value{
Raises an error if the UID columns do not have a valid structure.
}
\description{
Check if UID columns in a nested list have valid structure for a data list
}
\keyword{internal}
