% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_metrics_list.R
\name{check_dmll_fn_args}
\alias{check_dmll_fn_args}
\title{Check if functions in a distance metrics list-like have valid arguments}
\usage{
check_dmll_fn_args(dmll)
}
\arguments{
\item{dmll}{A distance metrics list-like list object to be validated.}
}
\value{
Doesn't return any value. Raises error if the functions in dmll
don't have valid arguments.
}
\description{
Check if functions in a distance metrics list-like have valid arguments
}
\keyword{internal}
