% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_metrics_list.R
\name{check_dmll_fn_names}
\alias{check_dmll_fn_names}
\title{Check if functions in a distance metrics list-like have names}
\usage{
check_dmll_fn_names(dmll)
}
\arguments{
\item{dmll}{A distance metrics list-like list object to be validated.}
}
\value{
Doesn't return any value. Raises error if the functions in dmll
don't have names.
}
\description{
Check if functions in a distance metrics list-like have names
}
\keyword{internal}
