% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_metrics_list.R
\name{check_dmll_item_names}
\alias{check_dmll_item_names}
\title{Check if items of a distance metrics list-like object have valid names}
\usage{
check_dmll_item_names(dmll)
}
\arguments{
\item{dmll}{A distance metrics list-like list object to be validated.}
}
\value{
Doesn't return any value. Raises error if the items of dmll don't
have valid formatted names.
}
\description{
Check if items of a distance metrics list-like object have valid names
}
\keyword{internal}
