% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_metrics_list.R
\name{check_dmll_subitems_are_fns}
\alias{check_dmll_subitems_are_fns}
\title{Check if subitems of a distance metrics list-like object are functions}
\usage{
check_dmll_subitems_are_fns(dmll)
}
\arguments{
\item{dmll}{A distance metrics list-like list object to be validated.}
}
\value{
Doesn't return any value. Raises error if the subitems of dmll are
not functions.
}
\description{
Check if subitems of a distance metrics list-like object are functions
}
\keyword{internal}
