% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_metrics_list.R
\name{check_dmll_unique_names}
\alias{check_dmll_unique_names}
\title{Check if names in a distance metrics list-like object are unique}
\usage{
check_dmll_unique_names(dmll)
}
\arguments{
\item{dmll}{A distance metrics list-like list object to be validated.}
}
\value{
Doesn't return any value. Raises error if the items of dmll aren't
unique across layer 1 or within each item of layer 2.
}
\description{
Check if names in a distance metrics list-like object are unique
}
\keyword{internal}
