% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{check_hm_dependencies}
\alias{check_hm_dependencies}
\title{Check for ComplexHeatmap and circlize dependencies}
\usage{
check_hm_dependencies()
}
\description{
Check for ComplexHeatmap and circlize dependencies
}
