% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_df.R
\name{check_sdfl_numeric}
\alias{check_sdfl_numeric}
\title{Check if settings data frame is numeric}
\usage{
check_sdfl_numeric(sdfl)
}
\arguments{
\item{sdfl}{A settings data frame-like matrix object to be validated.}
}
\value{
Doesn't return any value. Raises error if there are features with
duplicate names in a generated data list.
}
\description{
Check if settings data frame is numeric
}
\keyword{internal}
