% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_similarity_matrices}
\alias{check_similarity_matrices}
\title{Check validity of similarity matrices}
\usage{
check_similarity_matrices(similarity_matrices)
}
\arguments{
\item{similarity_matrices}{A list of similarity matrices}
}
\value{
valid_matrices Boolean indicating if properties are met by all
similarity matrices
}
\description{
Check to see if similarity matrices in a list have the following properties:
\enumerate{
\item The maximum value in the entire matrix is 0.5
\item Every value in the diagonal is 0.5
}
}
