% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_characterization.R
\name{check_subj_orders_for_lp}
\alias{check_subj_orders_for_lp}
\title{Check subject orders for label propagation}
\usage{
check_subj_orders_for_lp(data_list, om_row, n_train, n_test)
}
\arguments{
\item{data_list}{A data list}

\item{om_row}{An output matrix row}

\item{n_train}{number of training subjects}

\item{n_test}{number of testing subjects}
}
\value{
all_checks_passed boolean indicating if all subs are in order
}
\description{
Prior to label propagation, it is essential that the subject orders of the
full fused network and the supplied clustering information are consistent.
}
