% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_snf.R
\name{check_valid_k}
\alias{check_valid_k}
\title{Check if max K exceeds the number of observations}
\usage{
check_valid_k(sdf, dl)
}
\arguments{
\item{sdf}{A \code{settings_df} class object.}

\item{dl}{A nested list of input data from \code{data_list()}.}
}
\value{
Doesn't return any value. Raises error if max K exceeds the number
of observations.
}
\description{
Check if max K exceeds the number of observations
}
\keyword{internal}
