% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_snf.R
\name{check_valid_sc}
\alias{check_valid_sc}
\title{Check if SNF config has valid structure}
\usage{
check_valid_sc(sc)
}
\arguments{
\item{sc}{An \code{snf_config} class object.}
}
\value{
Doesn't return any value. Raises error if snf_config is not an
\code{snf_config} class object.
}
\description{
Check if SNF config has valid structure
}
\keyword{internal}
