% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_matrix.R
\name{chi_sq_p}
\alias{chi_sq_p}
\title{Chi-squared test p-value}
\usage{
chi_sq_p(clust_membership, outcome_df, outcome_var)
}
\arguments{
\item{clust_membership}{Dataframe of cluster membership (get_clustered_subs)}

\item{outcome_df}{Dataframe containing outcome feature}

\item{outcome_var}{Outcome feature as a string}
}
\value{
p_val The chi-squared test p-value
}
\description{
Returns the p-value following a chi-squared test (without Yates' continuity
correction) on the distribution of a categorical variable by cluster.
}
