% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions_df.R
\name{chi_squared_pval}
\alias{chi_squared_pval}
\title{Chi-squared test p-value (generic)}
\usage{
chi_squared_pval(cat_var1, cat_var2)
}
\arguments{
\item{cat_var1}{A categorical feature.}

\item{cat_var2}{A categorical feature.}
}
\value{
pval A p-value (class "numeric").
}
\description{
Return p-value for chi-squared test for any two features
}
\keyword{internal}
