% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{clean_plot}
\alias{clean_plot}
\title{Clean a plot}
\usage{
clean_plot(plot, removables = c())
}
\arguments{
\item{plot}{A ggplot object}

\item{removables}{A character vector of items to be removed. Can contain "x"
for x-axis label, "y" for y-axis label, and "legend" to remove the legend.}
}
\value{
plot A cleaned ggplot object
}
\description{
Given a ggplot object and a list of elements to be removed, return a cleaned
version of the plot.
}
