% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_plotting.r
\name{clusterToOutcomeCorr}
\alias{clusterToOutcomeCorr}
\title{Functions to calculate correlation between cluster assignment to outcome variables and visualize to find meaningful clusters with Manhattan plot}
\usage{
clusterToOutcomeCorr(df, outcomes, method, datatype)
}
\arguments{
\item{df}{a dataframe of samples with cluster_column and outcomes columns.}

\item{outcomes}{one or more outcomes of interest from df}

\item{method}{correlation test method ("chi-squared" or "kruskal")}

\item{datatype}{name of the SNF integration datatype}
}
\description{
Calculate correlation of clusters to each outcome using chi-squared (categorical outcome) and/or kruskal-wallis test (continuous outcome) in each data set that were integrated using SNF, and then generates long format data input for ClustersToOutcomeManhattan
}
