% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_plotting.r
\name{clusterToOutcomeManhattan}
\alias{clusterToOutcomeManhattan}
\title{Display cluster assignment to outcome correlation as Manhattan plot}
\usage{
clusterToOutcomeManhattan(cco, levels = NULL)
}
\arguments{
\item{cco}{cco #!}

\item{levels}{optional argument to re-arrange outcome display on x-axis}
}
\description{
Manhattan plot outputs the Correlation of Clusters from spectral clustering with the Outcomes (cco), colored by data types, dot size represents sample size.
}
