% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coclustering.R
\name{cocluster_heatmap}
\alias{cocluster_heatmap}
\title{Heatmap of patient co-clustering across resampled data}
\usage{
cocluster_heatmap(cocluster_data)
}
\arguments{
\item{cocluster_data}{A named list containing two matrices:
\itemize{
\item "same_solution": A patient x patient matrix where each cell is the number
of subsamples that contained both of those patients
\item "same_cluster": A patient x patient matrix where each cell is the number
of subsamples where those patients were clustered together
}}
}
\description{
Uses the output of generate_cocluster_data (see ?generate_cocluster_data)
and returns a well-formatted ComplexHeatmap for visualizing clustering
structure across resamplings of the data.
}
