% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coclustering.R
\name{coclustering_coverage_check}
\alias{coclustering_coverage_check}
\title{Co-clustering coverage check}
\usage{
coclustering_coverage_check(cocluster_df, action = "warn")
}
\arguments{
\item{cocluster_df}{data frame containing co-clustering data.}

\item{action}{Control if parent function should warn or stop.}
}
\value{
This function does not return any value. It checks a \code{cocluster_df}
for complete coverage (all pairs occur in the same solution at least once).
Will raise a warning or error if coverage is incomplete depending on the
value of the action parameter.
}
\description{
Check if co-clustered data has at least one subsample in which every pair
of observations were a part of simultaneously.
}
\keyword{internal}
