% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{col_to_num}
\alias{col_to_num}
\title{Convert specific columns to numeric}
\usage{
col_to_num(df, col_indices)
}
\arguments{
\item{df}{The dataframe containing columns to be converted}

\item{col_indices}{The positions of the columns to be converted}
}
\value{
df The dataframe with numeric columns
}
\description{
Convert specific columns to numeric
}
