% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{col_to_num_all_possible}
\alias{col_to_num_all_possible}
\title{Convert all possible columns to numeric}
\usage{
col_to_num_all_possible(df)
}
\arguments{
\item{df}{A dataframe or tibble}
}
\value{
df The same dataframe with all possible columns made numeric
}
\description{
Convert all possible columns to numeric
}
