% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{collapse_dl}
\alias{collapse_dl}
\title{Convert a data list into a data frame}
\usage{
collapse_dl(data_list)
}
\arguments{
\item{data_list}{A nested list of input data from \code{generate_data_list()}.}
}
\value{
A "data.frame"-formatted version of the provided data list.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}
Defunct function for converting a data list into a data frame. Please
use \code{as.data.frame()} instead.
}
