% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colour_scale}
\alias{colour_scale}
\title{Return a colour ramp for a given vector}
\usage{
colour_scale(data, min_colour, max_colour)
}
\arguments{
\item{data}{Vector of numeric values.}

\item{min_colour}{Minimum colour value.}

\item{max_colour}{Maximum colour value.}
}
\value{
A "function" class object that can build a circlize-style colour
ramp.
}
\description{
Given a numeric vector and min and max colour values, return a colour ramp
that assigns a colour to each element in the vector. This function is a
wrapper for \code{circlize::colorRamp2}.'
}
