% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_plotting.R
\name{correlation_pval_heatmap}
\alias{correlation_pval_heatmap}
\title{Heatmap of pairwise associations between variables}
\usage{
correlation_pval_heatmap(
  correlation_matrix,
  scale_diag = "max",
  cluster_rows = TRUE,
  cluster_columns = TRUE,
  show_row_names = TRUE,
  show_column_names = TRUE,
  show_heatmap_legend = FALSE,
  confounders = NULL,
  out_of_models = NULL,
  annotation_colours = NULL,
  labels_colour = NULL,
  split_by_domain = NULL,
  ...
)
}
\arguments{
\item{correlation_matrix}{Matrix containing all pairwise association
p-values. The recommended way to obtain this matrix is through the
calculate_associations function.}

\item{scale_diag}{Parameter that controls how the diagonals of the
correlation_matrix are adjusted in the heatmap. For best viewing, this is
set to "max", which will match the diagonals to whichever pairwise
association has the highest p-value.}

\item{cluster_rows}{Parameter for ComplexHeatmap::Heatmap. Will be ignored
if split_by_domain is also provided.}

\item{cluster_columns}{Parameter for ComplexHeatmap::Heatmap. Will be
ignored if split_by_domain is also provided.}

\item{show_row_names}{Parameter for ComplexHeatmap::Heatmap.}

\item{show_column_names}{Parameter for ComplexHeatmap::Heatmap.}

\item{show_heatmap_legend}{Parameter for ComplexHeatmap::Heatmap.}

\item{confounders}{A named list where the elements are columns in the
correlation_matrix and the names are the corresponding display names.}

\item{out_of_models}{Like confounders, but a named list of out of model
measures (who are also present as columns in the correlation_matrix).}

\item{annotation_colours}{Named list of heatmap annotations and their
colours.}

\item{labels_colour}{Vector of colours to use for the columns and rows
of the heatmap.}

\item{split_by_domain}{The results of \code{dl_var_summar} - a dataframe that has
the domain of every variable in the plotted data.
columns of the correlation_matrix. Will be used to "slice" the heatmap into
visually separated sections.}

\item{...}{Additional parameters passed into ComplexHeatmap::Heatmap.}
}
\description{
Heatmap of pairwise associations between variables
}
