% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_plotting.r
\name{corrHeatmap}
\alias{corrHeatmap}
\title{Generate correlation heatmap}
\usage{
corrHeatmap(corr, outcome_label_color = NULL)
}
\arguments{
\item{corr}{matrix of outcomes-outcomes correlation p_values}

\item{outcome_label_color}{optional argument to specify outcome color labels}
}
\description{
Generate correlation heatmap (need more generalization. Only tested outcomes-outcomes correlation)
}
