% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_plotting.r
\name{corrHeatmap_legend}
\alias{corrHeatmap_legend}
\title{Generate correlation heatmap legend}
\usage{
corrHeatmap_legend(legend_name, outcome_labels, outcome_labels_color)
}
\arguments{
\item{legend_name}{graph path to be saved to}

\item{outcome_labels}{optional argument to specify outcome label names}

\item{outcome_labels_color}{optional argument to specify outcome label name colors}
}
\description{
Generate legend for correlation heatmap.
}
