% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbi.R
\name{detail_mtbi}
\alias{detail_mtbi}
\title{Chain several TBI annotation functions to add multiple helpful mTBI columns}
\usage{
detail_mtbi(otbi01, subjects)
}
\arguments{
\item{otbi01}{The baseline TBI dataframe}

\item{subjects}{Dataframe containing list of required subjects}
}
\value{
detailed_otbi01 The modified dataframe
}
\description{
Chain several TBI annotation functions to add multiple helpful mTBI columns
}
