% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nclust_estimation.R
\name{discretisation_evec_data}
\alias{discretisation_evec_data}
\title{Internal function for \code{estimate_nclust_given_graph}}
\usage{
discretisation_evec_data(eigenvector)
}
\arguments{
\item{eigenvector}{Matrix of eigenvectors}
}
\value{
"Matrix" class object of discretized provided eigenvector to values 0
or 1.
}
\description{
Internal function taken from \code{SNFtool} to use for number of cluster
estimation.
}
\keyword{internal}
