% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_plotting.R
\name{displayClustersHeatmap}
\alias{displayClustersHeatmap}
\title{Display SNF cluster output in Heatmap}
\usage{
displayClustersHeatmap(
  similarity_matrix,
  group_cluster,
  top_annotation = NULL,
  left_annotation = NULL,
  scale_diag = "none",
  log_graph = FALSE
)
}
\arguments{
\item{similarity_matrix}{A similarity matrix}

\item{group_cluster}{cluster assignment from spectral clustering}

\item{top_annotation}{annotation to be displayed above the heatmap output}

\item{left_annotation}{annotation to be displayed on the left of the heatmap output}

\item{scale_diag}{Method of rescaling matrix diagonals. Can be "none"
(default - don't change diagonals), "mean" (replace diagonals with average
value of off-diagonals), or "zero" (replace diagonals with 0).}

\item{log_graph}{If TRUE, log transforms the graph.}
}
\description{
Normalize SNF matrix and plot heatmap.
}
