% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_metrics_list.R
\name{distance_metrics_list}
\alias{distance_metrics_list}
\title{Build a distance metrics list}
\usage{
distance_metrics_list(
  cnt_dist_fns = NULL,
  dsc_dist_fns = NULL,
  ord_dist_fns = NULL,
  cat_dist_fns = NULL,
  mix_dist_fns = NULL,
  use_default_dist_fns = FALSE
)
}
\arguments{
\item{cnt_dist_fns}{A named list of continuous distance metric functions.}

\item{dsc_dist_fns}{A named list of discrete distance metric functions.}

\item{ord_dist_fns}{A named list of ordinal distance metric functions.}

\item{cat_dist_fns}{A named list of categorical distance metric functions.}

\item{mix_dist_fns}{A named list of mixed distance metric functions.}

\item{use_default_dist_fns}{If TRUE, prepend the base distance metrics (euclidean
distance for continuous, discrete, and ordinal data and gower distance for
categorical and mixed data) to the resulting distance metrics list.}
}
\value{
A distance metrics list object.
}
\description{
The distance metrics list object (inherits classes \code{distance_metrics_list}
and \code{list}) is a list that stores R functions which can convert a data
frame of features into a matrix of pairwise distances. The list is a nested
one, where the first layer of the list can hold up to 5 items (one for each
of the metasnf recognized feature types, continuous, discrete, ordinal,
categorical, and mixed), and the second layer can hold an arbitrary number
of distance functions for each of those types.
}
\examples{
# Using just the base distance metrics  ------------------------------------
distance_metrics_list <- distance_metrics_list()

# Adding your own metrics --------------------------------------------------
# This will contain only the and user-provided distance function:
my_distance_metric <- function(df) {
    # your code that converts a dataframe to a distance metric here...
    # return(distance_metric)
}

distance_metrics_list <- distance_metrics_list(
    cnt_dist_fns = list(
         "my_distance_metric" = my_distance_metric
    )
)

# Using default base metrics------------------------------------------------
# This will contain user-provided and default distance functions:
distance_metrics_list <- distance_metrics_list(
    cnt_dist_fns = list(
         "my_distance_metric" = my_distance_metric
    ),
    dsc_dist_fns = list(
         "my_distance_metric" = my_distance_metric
    ),
    ord_dist_fns = list(
         "my_distance_metric" = my_distance_metric
    ),
    cat_dist_fns = list(
         "my_distance_metric" = my_distance_metric
    ),
    mix_dist_fns = list(
         "my_distance_metric" = my_distance_metric
    ),
    use_default_dist_fns = TRUE
)
}
