% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{dl_uid_first_col}
\alias{dl_uid_first_col}
\title{Make the uid UID columns of a data list first}
\usage{
dl_uid_first_col(dl)
}
\arguments{
\item{dl}{A nested list of input data from \code{data_list()}.}
}
\value{
A data list ("list"-class object) in which each data-subcomponent
has "uid" positioned as its first column.
}
\description{
Make the uid UID columns of a data list first
}
