% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{dll_uid_first_col}
\alias{dll_uid_first_col}
\title{Make the uid UID columns of a data list first}
\usage{
dll_uid_first_col(dll)
}
\arguments{
\item{dll}{A data list-like \code{list} class object.}
}
\value{
The object with "uid" positioned as the first of each data frame
column.
}
\description{
Make the uid UID columns of a data list first
}
\keyword{internal}
