% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{dm_heatmap}
\alias{dm_heatmap}
\title{Heatmap design matrix based on meta-clustering results}
\usage{
dm_heatmap(design_matrix, save = NULL)
}
\arguments{
\item{design_matrix}{matrix indicating parameters to iterate SNF through}

\item{save}{optional path to save figure to}
}
\description{
Normalizes the design matrix and plots as a heatmap. Rows are reordered to
match the row-clustering present within a provided meta-clustering result.
}
