% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_snf.R
\name{dm_row_fn}
\alias{dm_row_fn}
\title{Apply-based function for batch_snf}
\usage{
dm_row_fn(dm_row, dl)
}
\arguments{
\item{dm_row}{a row of a design matrix}

\item{dl}{a data list}
}
\value{
the corresponding OM row
}
\description{
Apply-based function for batch_snf
}
