% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf.R
\name{domain_merge}
\alias{domain_merge}
\title{Domain merge}
\usage{
domain_merge(data_list)
}
\arguments{
\item{data_list}{nested list of input data generated by the function
\code{get_data_list()}}
}
\value{
domain_dl
}
\description{
Domain merge
}
