% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{domains}
\alias{domains}
\title{Domains}
\usage{
domains(data_list)
}
\arguments{
\item{data_list}{A nested list of input data from \code{generate_data_list()}.}
}
\value{
domain_list list of domains
}
\description{
Domains
}
