% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{dplyr_row_slice.solutions_df}
\alias{dplyr_row_slice.solutions_df}
\title{Function to extend dplyr to solutions data frame objects}
\usage{
dplyr_row_slice.solutions_df(data, i, ...)
}
\arguments{
\item{data}{A solutions data frame.}

\item{i}{A vector of row indices.}

\item{...}{Additional arguments.}
}
\value{
Row sliced object with appropriately preserved attributes.
}
\description{
Function to extend dplyr to solutions data frame objects
}
