% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_cols}
\alias{drop_cols}
\title{Helper function to remove columns from a data frame}
\usage{
drop_cols(x, cols)
}
\arguments{
\item{x}{A data frame}

\item{cols}{Vector of column names to be removed}
}
\value{
x without columns in cols
}
\description{
Helper function to remove columns from a data frame
}
\keyword{internal}
