% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_snf.R
\name{drop_inputs}
\alias{drop_inputs}
\title{Execute inclusion}
\usage{
drop_inputs(settings_matrix_row, data_list)
}
\arguments{
\item{settings_matrix_row}{Row of a settings matrix.}

\item{data_list}{A nested list of input data from \code{generate_data_list()}.}
}
\value{
A data list (class "list") in which any component with a
corresponding 0 value in the provided settings matrix row has been removed.
}
\description{
Given a data list and a settings matrix row, returns a data list of selected
inputs
}
