% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_measures.R
\name{dunn_manual}
\alias{dunn_manual}
\title{Calculate Dunn index}
\usage{
dunn_manual(dist_mat, clusters)
}
\arguments{
\item{dist_mat}{A dissimilarity matrix.}

\item{clusters}{An integer vector of cluster assignments for that matrix.}
}
\value{
Dunn index (numeric).
}
\description{
Calculate Dunn index
}
\keyword{internal}
