% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{edm_high}
\alias{edm_high}
\title{Execute design matrix function based on apply rather than for loop}
\usage{
edm_high(data_list, design_matrix)
}
\arguments{
\item{data_list}{nested list of input data generated by the function
\code{get_data_list()}}

\item{design_matrix}{matrix indicating parameters to iterate SNF through}
}
\value{
populated_design_matrix design matrix with filled columns related to
subtype membership
}
\description{
Execute design matrix function based on apply rather than for loop
}
