% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{ensure_dll_df}
\alias{ensure_dll_df}
\title{Ensure the data item of each component is a \code{data.frame} class object}
\usage{
ensure_dll_df(dll)
}
\arguments{
\item{dll}{A data list-like \code{list} class object.}
}
\value{
The provided dll with the data item of each component as a
data frame.
}
\description{
Ensure the data item of each component is a \code{data.frame} class object
}
\keyword{internal}
