% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_metrics.R
\name{euclidean_distance}
\alias{euclidean_distance}
\title{Distance metric: Euclidean distance}
\usage{
euclidean_distance(df)
}
\arguments{
\item{df}{Dataframe containing at least 1 data column}
}
\value{
distance_matrix A distance matrix.
}
\description{
Distance metric: Euclidean distance
}
