% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf.R
\name{execute_design_matrix}
\alias{execute_design_matrix}
\title{Execute variations of SNF as described by a design matrix}
\usage{
execute_design_matrix(design_matrix)
}
\arguments{
\item{design_matrix}{Design matrix
* as built by fns \code{build_design_matrix_base} and \code{add_design_matrix_rows}}
}
\value{
design_matrix_executed Design matrix with populated subtype
membership and target variable columns
}
\description{
Execute variations of SNF as described by a design matrix
}
