% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{execute_design_matrix_hard}
\alias{execute_design_matrix_hard}
\title{Execute variations of SNF as described by a design matrix}
\usage{
execute_design_matrix_hard(data_list, design_matrix, outcome_list)
}
\arguments{
\item{data_list}{nested list of input data generated by the function
\code{get_data_list()}}

\item{design_matrix}{matrix indicating parameters to iterate SNF through}

\item{outcome_list}{nested list of outcome data}
}
\value{
populated_design_matrix design matrix with filled columns related to
subtype membership and cluster performance
}
\description{
Execute variations of SNF as described by a design matrix
}
