% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{execute_design_matrix_soft}
\alias{execute_design_matrix_soft}
\title{Execute variations of SNF as described by a design matrix - but softcoded}
\usage{
execute_design_matrix_soft(data_list, design_matrix)
}
\arguments{
\item{data_list}{nested list of input data generated by the function
\code{get_data_list()}}

\item{design_matrix}{matrix indicating parameters to iterate SNF through}
}
\value{
populated_design_matrix design matrix with filled columns related to
subtype membership
}
\description{
Execute variations of SNF as described by a design matrix - but softcoded
}
