% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expression_df}
\alias{expression_df}
\title{Modification of SNFtool mock data frame "Data1"}
\format{
\subsection{\code{expression_df}}{

A data frame with 200 rows and 3 columns:
\describe{
\item{gene_1_expression}{Mock gene expression feature}
\item{gene_2_expression}{Mock gene expression feature}
\item{patient_id}{Random three-digit number uniquely identifying the patient}
}
}
}
\source{
This data came from the SNFtool package, with slight modifications.
}
\usage{
expression_df
}
\description{
Modification of SNFtool mock data frame "Data1"
}
\keyword{datasets}
