% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_matrix.R
\name{extend_om}
\alias{extend_om}
\title{Extend an solutions matrix to include outcome evaluations}
\usage{
extend_om(solutions_matrix, outcome_list)
}
\arguments{
\item{solutions_matrix}{an solutions_matrix}

\item{outcome_list}{an outcome_list}
}
\value{
extended_solutions_matrix an extended solutions matrix that contains
p-value columns for each outcome in the provided outcome_list
}
\description{
Extend an solutions matrix to include outcome evaluations
}
