% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions_matrix.R
\name{extend_solutions}
\alias{extend_solutions}
\title{Extend an solutions matrix to include outcome evaluations}
\usage{
extend_solutions(solutions_matrix, target_list, cat_test = "chi_squared")
}
\arguments{
\item{solutions_matrix}{A solutions_matrix.}

\item{target_list}{A target_list.}

\item{cat_test}{String indicating which statistical test will be used to
associate cluster with a categorical variable. Options are "chi_squared" for
the Chi-squared test and "fisher_exact" for Fisher's exact test.}
}
\value{
extended_solutions_matrix an extended solutions matrix that contains
p-value columns for each outcome in the provided target_list
}
\description{
Extend an solutions matrix to include outcome evaluations
}
