% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{features}
\alias{features}
\title{Return character vector of features stored in an object}
\usage{
features(x)
}
\arguments{
\item{x}{The object to pull features from.}
}
\value{
A character vector of features in x.
}
\description{
Return character vector of features stored in an object
}
