% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_baseline}
\alias{filter_baseline}
\title{Remove the data dictionary from a dataset}
\usage{
filter_baseline(abcd_df)
}
\arguments{
\item{abcd_df}{An ABCD dataframe containing a data dictionary}
}
\value{
baseline_abcd_df Baseline data only
}
\description{
Remove the data dictionary from a dataset
}
