% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_subjects}
\alias{filter_subjects}
\title{Subset a dataframe to a given subject list}
\usage{
filter_subjects(abcd_df, subject_list)
}
\arguments{
\item{abcd_df}{An ABCD dataframe}

\item{subject_list}{A dataframe containing subjects in the 'subjectkey' col}
}
\value{
filtered_df The subsetted dataframe
}
\description{
Subset a dataframe to a given subject list
}
