% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions_matrix.R
\name{fisher_exact_pval}
\alias{fisher_exact_pval}
\title{Fisher exact test p-value}
\usage{
fisher_exact_pval(cat_var1, cat_var2)
}
\arguments{
\item{cat_var1}{A categorical feature.}

\item{cat_var2}{A categorical feature.}
}
\value{
pval A p-value (class "numeric").
}
\description{
Return p-value for Fisher exact test for any two features
}
