% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{fraction_clustered_together}
\alias{fraction_clustered_together}
\title{(DEPRECATED) Average fraction of clustering together}
\usage{
fraction_clustered_together(...)
}
\arguments{
\item{...}{Function arguments}
}
\description{
(DEPRECATED) Average fraction of clustering together
}
