% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gender_df}
\alias{gender_df}
\title{Mock gender data}
\format{
\subsection{\code{gender_df}}{

A data frame with 200 rows and 2 columns:
\describe{
\item{patient_id}{Random three-digit number uniquely identifying the patient}
\item{gender_df}{Mock gene methylation feature}
}
}
}
\source{
This data came from the SNFtool package, with slight modifications.
}
\usage{
gender_df
}
\description{
Mock gender data
}
\keyword{datasets}
