% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf.R
\name{generate_data_list}
\alias{generate_data_list}
\title{Generate data_list object}
\usage{
generate_data_list(
  mtbi_loc = NULL,
  mtbi_mechanism = NULL,
  mtbi_mem_daze = NULL,
  income = NULL,
  interview_age = NULL,
  mtbi_age = NULL,
  pubertal_status = NULL,
  race = NULL,
  sex = NULL,
  wmndf = NULL,
  headaches = NULL,
  mtbi_count = NULL,
  gord_cor = NULL,
  gord_var = NULL,
  subc_cor = NULL,
  subc_var = NULL,
  cort_sa = NULL,
  cort_t = NULL,
  subc_v = NULL
)
}
\arguments{
\item{mtbi_loc}{mtbi_loc dataframe generated by the \code{get_mtbi_loc}
function}

\item{mtbi_mechanism}{mtbi_mechanisc dataframe generated by the
\code{get_mtbi_mechanism} function}

\item{mtbi_mem_daze}{mtbi_mem_dazc dataframe generated by the
\code{get_mtbi_mem_daze} function}

\item{income}{incomc dataframe generated by the \code{get_income} function}

\item{interview_age}{interview_agc dataframe generated by the
\code{get_interview_age} function}

\item{mtbi_age}{mtbi_agc dataframe generated by the \code{get_mtbi_age} function}

\item{pubertal_status}{pubertal_statuc dataframe generated by the
\code{get_pubertal_status} function}

\item{race}{race_cc dataframe generated by the \code{get_race} function}

\item{sex}{sec dataframe generated by the \code{get_sex} function}

\item{wmndf}{wmndc dataframe generated by the \code{get_wmndf} function}

\item{headaches}{headachec dataframe generated by the \code{get_headaches}
function}

\item{mtbi_count}{mtbi_counc dataframe generated by the \code{get_mtbi_count}
function}

\item{gord_cor}{gord_coc dataframe generated by the \code{get_gord_cor} function}

\item{gord_var}{gord_vac dataframe generated by the \code{get_gord_var} function}

\item{subc_cor}{subc_coc dataframe generated by the \code{get_subc_cor} function}

\item{subc_var}{subc_vac dataframe generated by the \code{get_subc_var} function}

\item{cort_sa}{cort_sc dataframe generated by the \code{get_cort_sa} function}

\item{cort_t}{cort_c dataframe generated by the \code{get_cort_t} function}

\item{subc_v}{subc_c dataframe generated by the \code{get_subc_v} function}
}
\description{
This is the major data object that will be processed when iterating through
the design matrix. The full list contains one list per measurement type.
Within each measurement type's list, elements include the actual data
structure, the name, the subdomain (if applicable), the domain, and the
data 'type' (i.e, numerical or categorical).
}
