% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf.R
\name{generate_data_list_soft}
\alias{generate_data_list_soft}
\title{Generate data_list object - but softcoded}
\usage{
generate_data_list_soft(...)
}
\arguments{
\item{...}{Lists formatted as (df, "df_name", "df_domain", "df_type")}
}
\description{
This is the major data object that will be processed when iterating through
the design matrix. The full list contains one list per measurement type.
Within each measurement type's list, elements include the actual data
structure, the name, the domain, and the data 'type' (i.e, numeric or
categorical).
}
\details{
To-do: include checks to make sure format of data list is correct
}
