% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{generate_demographic}
\alias{generate_demographic}
\title{Returns race information}
\usage{
generate_demographic(puberty, age, sex, income, race)
}
\arguments{
\item{puberty}{x}

\item{age}{x}

\item{sex}{x}

\item{income}{x}

\item{race}{x}
}
\value{
race_df x
}
\description{
Returns race information
}
