% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_matrix.R
\name{generate_design_matrix}
\alias{generate_design_matrix}
\title{Build a design matrix - but softcoded}
\usage{
generate_design_matrix(
  data_list,
  nrows = 0,
  seed = NULL,
  min_removed = NULL,
  max_removed = NULL,
  retry_limit = 10
)
}
\arguments{
\item{data_list}{a data list object to determine variables for inclusion and
exclusion}

\item{nrows}{number of design matrix rows}

\item{seed}{set a seed for the random matrix generation. Note that this}

\item{min_removed}{The smallest number of elements that may be removed}

\item{max_removed}{The largest number of elements that may be removed}

\item{retry_limit}{The maximum number of attempts to generate a novel row
affects the global seed.}
}
\value{
design_matrix A design matrix
}
\description{
Build a design matrix - but softcoded
}
