% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_matrix.R
\name{generate_design_matrix_ak}
\alias{generate_design_matrix_ak}
\title{Build design matrix for scanning alpha and K}
\usage{
generate_design_matrix_ak()
}
\value{
design_matrix Adds the standard grid expansion for SNF hyperparams
}
\description{
Build design matrix for scanning alpha and K
}
