% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_list.R
\name{generate_outcome_list}
\alias{generate_outcome_list}
\title{Generate outcome_list object - but softcoded}
\usage{
generate_outcome_list(...)
}
\arguments{
\item{...}{Lists of outcomes formatted as (dataframe, "name", "type")}
}
\value{
outcome_list structure containing all outcome measure data
}
\description{
The major object containing all outcome variables
}
