% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_matrix.R
\name{generate_settings_matrix}
\alias{generate_settings_matrix}
\title{Build a settings matrix}
\usage{
generate_settings_matrix(
  data_list,
  nrows = 0,
  seed = NULL,
  min_removed = NULL,
  max_removed = NULL,
  possible_snf_schemes = c(1, 2, 3),
  retry_limit = 10
)
}
\arguments{
\item{data_list}{a data list object to determine variables for inclusion and
exclusion}

\item{nrows}{number of settings matrix rows}

\item{seed}{set a seed for the random matrix generation. Note that this}

\item{min_removed}{The smallest number of elements that may be removed}

\item{max_removed}{The largest number of elements that may be removed}

\item{possible_snf_schemes}{A vector containing the possible snf_schemes to
uniformly randomly select from. By default, the vector contains all
possible schemes: c(1, 2, 3)}

\item{retry_limit}{The maximum number of attempts to generate a novel row
affects the global seed.}
}
\value{
settings_matrix A settings matrix
}
\description{
Build a settings matrix
}
