% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions_matrix.R
\name{generate_solutions_matrix}
\alias{generate_solutions_matrix}
\title{Build solutions matrix}
\usage{
generate_solutions_matrix(data_list, settings_matrix)
}
\arguments{
\item{data_list}{nested list of input data generated by the function
\code{get_data_list()}}

\item{settings_matrix}{matrix indicating parameters to iterate SNF through}
}
\value{
solutions_matrix The solutions matrix
}
\description{
The matrix that will contain all the columns to be populated by SNF results
}
