% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_list.R
\name{generate_target_list}
\alias{generate_target_list}
\title{Generate target_list object}
\usage{
generate_target_list(..., uid = NULL)
}
\arguments{
\item{...}{Lists of outcomes formatted as (dataframe, "name", "type")}

\item{uid}{(string) the name of the uid column currently used data}
}
\value{
target_list structure containing all outcome measure data
}
\description{
The major object containing all outcome variables
}
