% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_resilience.R
\name{get_activities}
\alias{get_activities}
\title{Extract activities}
\usage{
get_activities(abcd_saiq02, subjects = NULL)
}
\arguments{
\item{abcd_saiq02}{ABCD Parent Sports and Activities Involvement
Questionnaire}

\item{subjects}{Dataframe containing list of required subjects}
}
\value{
activities
}
\description{
Extract activities
}
