% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acute_symptoms.R
\name{get_acute_symptoms}
\alias{get_acute_symptoms}
\title{Select columns relevant for acute symptoms}
\usage{
get_acute_symptoms(tbi_df)
}
\arguments{
\item{tbi_df}{A TBI dataframe}
}
\value{
acute_symptoms Dataframe containing acute symptom information
}
\description{
Select columns relevant for acute symptoms
}
