% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbcl.R
\name{get_cbcl_aggressive_r}
\alias{get_cbcl_aggressive_r}
\title{Get CBCL aggressive data}
\usage{
get_cbcl_aggressive_r(abcd_cbcls01, subjects = NULL)
}
\arguments{
\item{subjects}{list of subjects to receive data for}

\item{abcd_cbcl01}{NDA cbcl dataframe}
}
\value{
cbcl_aggressive_r aggressive data
}
\description{
Get CBCL aggressive data
}
