% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbcl.R
\name{get_cbcl_anxiety_r}
\alias{get_cbcl_anxiety_r}
\title{Get CBCL anxiety data}
\usage{
get_cbcl_anxiety_r(abcd_cbcls01, subjects = NULL)
}
\arguments{
\item{subjects}{list of subjects to receive data for}

\item{abcd_cbcl01}{NDA cbcl dataframe}
}
\value{
cbcl_anxiety_r anxiety data
}
\description{
Get CBCL anxiety data
}
