% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbcl.R
\name{get_cbcl_attention_r}
\alias{get_cbcl_attention_r}
\title{Get CBCL attention data}
\usage{
get_cbcl_attention_r(abcd_cbcls01, subjects = NULL)
}
\arguments{
\item{subjects}{list of subjects to receive data for}

\item{abcd_cbcl01}{NDA cbcl dataframe}
}
\value{
cbcl_attention_r attention data
}
\description{
Get CBCL attention data
}
