% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbcl.R
\name{get_cbcl_headaches}
\alias{get_cbcl_headaches}
\title{Get CBCL headache data}
\usage{
get_cbcl_headaches(abcd_cbcl01, subjects)
}
\arguments{
\item{abcd_cbcl01}{NDA cbcl dataframe}

\item{subjects}{list of subjects to receive data for}
}
\value{
cbcl_headaches headache data
}
\description{
Get CBCL headache data
}
