% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbcl.R
\name{get_cbcl_nausea}
\alias{get_cbcl_nausea}
\title{Get CBCL nausea data}
\usage{
get_cbcl_nausea(abcd_cbcl01, subjects = NULL)
}
\arguments{
\item{abcd_cbcl01}{NDA cbcl dataframe}

\item{subjects}{list of subjects to receive data for}
}
\value{
cbcl_nausea nausea data
}
\description{
Get CBCL nausea data
}
