% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbcl.R
\name{get_cbcl_overtired}
\alias{get_cbcl_overtired}
\title{Get CBCL overtired data}
\usage{
get_cbcl_overtired(abcd_cbcl01, subjects = NULL)
}
\arguments{
\item{abcd_cbcl01}{NDA cbcl dataframe}

\item{subjects}{list of subjects to receive data for}
}
\value{
cbcl_overtired overtired data
}
\description{
Get CBCL overtired data
}
