% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbcl.R
\name{get_cbcl_sleeping_less}
\alias{get_cbcl_sleeping_less}
\title{Get CBCL sleeping_less data}
\usage{
get_cbcl_sleeping_less(abcd_cbcl01, subjects = NULL)
}
\arguments{
\item{abcd_cbcl01}{NDA cbcl dataframe}

\item{subjects}{list of subjects to receive data for}
}
\value{
cbcl_sleeping_less sleeping_less data
}
\description{
Get CBCL sleeping_less data
}
