% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbcl.R
\name{get_cbcl_sleeping_more}
\alias{get_cbcl_sleeping_more}
\title{Get CBCL sleeping_more data}
\usage{
get_cbcl_sleeping_more(abcd_cbcl01, subjects = NULL)
}
\arguments{
\item{abcd_cbcl01}{NDA cbcl dataframe}

\item{subjects}{list of subjects to receive data for}
}
\value{
cbcl_sleeping_more sleeping_more data
}
\description{
Get CBCL sleeping_more data
}
