% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbcl.R
\name{get_cbcl_vomiting}
\alias{get_cbcl_vomiting}
\title{Get CBCL vomiting data}
\usage{
get_cbcl_vomiting(abcd_cbcl01, subjects = NULL)
}
\arguments{
\item{abcd_cbcl01}{NDA cbcl dataframe}

\item{subjects}{list of subjects to receive data for}
}
\value{
cbcl_vomiting vomiting data
}
\description{
Get CBCL vomiting data
}
