% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_characterization.R
\name{get_cluster_df}
\alias{get_cluster_df}
\title{Extract dataframe of cluster and subject key from output matrix row}
\usage{
get_cluster_df(om_row)
}
\arguments{
\item{om_row}{Output matrix row}
}
\value{
cluster_df dataframe of cluster and subjectkey
}
\description{
Extract dataframe of cluster and subject key from output matrix row
}
