% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{get_cluster_pval}
\alias{get_cluster_pval}
\title{(DEPRECATED) Get p-value}
\usage{
get_cluster_pval(...)
}
\arguments{
\item{...}{Function arguments}
}
\description{
(DEPRECATED) Get p-value
}
