% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_characterization.R
\name{get_clusters}
\alias{get_clusters}
\title{Extract list of assigned clusters}
\usage{
get_clusters(om_row)
}
\arguments{
\item{om_row}{Output matrix row}
}
\value{
clusters list of assigned clusters
}
\description{
Extract list of assigned clusters
}
