% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_complete_uids}
\alias{get_complete_uids}
\title{Pull complete-data UIDs from a list of dataframes}
\usage{
get_complete_uids(list_of_dfs, uid)
}
\arguments{
\item{list_of_dfs}{List of dataframes.}

\item{uid}{Name of column across dataframes containing UIDs}
}
\value{
A character vector of the UIDs of observations that have complete
data across the provided list of dataframes.
}
\description{
Pull complete-data UIDs from a list of dataframes
}
