% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroimaging.R
\name{get_cort_t}
\alias{get_cort_t}
\title{Extract cortical thicknesses}
\usage{
get_cort_t(mrisdp10201, subjects)
}
\arguments{
\item{mrisdp10201}{Data file containing cortical data}

\item{subjects}{Optional dataframe containing needed subjectkeys}
}
\value{
cort_t_df Dataframe of cortical thicknesses
}
\description{
Extract cortical thicknesses
}
