% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf.R
\name{get_dist_matrix}
\alias{get_dist_matrix}
\title{Calculate distance matrices}
\usage{
get_dist_matrix(df, input_type)
}
\arguments{
\item{df}{Raw dataframe with subject IDs in column 1}

\item{input_type}{Either "numeric" (resulting in euclidean distances),
"categorical" (resulting in binary distances), or "mixed" (resulting in
gower distances)}
}
\value{
dist_matrix Matrix of inter-observation distances
}
\description{
Given a dataframe of numerical variables, return a euclidean distance matrix
}
