% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{get_dl_subjects}
\alias{get_dl_subjects}
\title{Extract subjects from a data_list}
\usage{
get_dl_subjects(data_list, prefix = FALSE)
}
\arguments{
\item{data_list}{A nested list of input data from \code{generate_data_list()}.}

\item{prefix}{If TRUE, preserves the "subject_" prefix added to UIDs when
creating a data_list.}
}
\value{
A character vector of the UID labels contained in a data list.
}
\description{
Extract subjects from a data_list
}
