% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_resilience.R
\name{get_exercise}
\alias{get_exercise}
\title{Extract healthy behaviours: exercise questionnaire}
\usage{
get_exercise(abcd_yrb01, subjects = NULL)
}
\arguments{
\item{abcd_yrb01}{ABCD Youth Youth Risk Behavior Survey Exercise Physical
Activity}

\item{subjects}{Dataframe containing list of required subjects}
}
\value{
exercise
}
\description{
Extract healthy behaviours: exercise questionnaire
}
