% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{get_family_function}
\alias{get_family_function}
\title{Extract family function}
\usage{
get_family_function(fes02, abcd_fes01, subjects)
}
\arguments{
\item{fes02}{ABCD Parent Family Environment Scale-Family Conflict Subscale
Modified from PhenX}

\item{abcd_fes01}{ABCD Parent Family Environment Scale-Family Conflict
Subscale Modified from PhenX}

\item{subjects}{Dataframe containing list of required subjects}
}
\value{
family_function
}
\description{
Extract family function
}
