% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroimaging.R
\name{get_gord_cor}
\alias{get_gord_cor}
\title{Extract cortical network correlations}
\usage{
get_gord_cor(betnet02, subjects)
}
\arguments{
\item{betnet02}{Data file containing neurite density data}

\item{subjects}{Dataframe containing list of required subjects}
}
\value{
gord_cor Dataframe of white matter neurite densities
}
\description{
Extract cortical network correlations
}
