% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroimaging.R
\name{get_gord_var}
\alias{get_gord_var}
\title{Extract cortical temporal variances}
\usage{
get_gord_var(mrirstv02, subjects)
}
\arguments{
\item{mrirstv02}{Data file containing neurite density data}

\item{subjects}{Optional dataframe containing needed subjectkeys}
}
\value{
subc_cor Dataframe of white matter neurite densities
}
\description{
Extract cortical temporal variances
}
