% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_history.R
\name{get_headaches}
\alias{get_headaches}
\title{Get subject headache history}
\usage{
get_headaches(mx01, subjects = NULL)
}
\arguments{
\item{mx01}{Dataframe containing medical history}

\item{subjects}{Dataframe containing list of required subjects}
}
\value{
headaches Dataframe containing headache history
}
\description{
Get subject headache history
}
