% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_cluster.R
\name{get_heatmap_order}
\alias{get_heatmap_order}
\title{Return the row ordering of a meta-clustering solution}
\usage{
get_heatmap_order(matrix)
}
\arguments{
\item{matrix}{matrix used as pheatmap input}
}
\value{
pheatmap_order Row orders of the clustered pheatmap
}
\description{
Pheatmap reorders meta clustering results to enable meta-cluster
visualization. This function extracts the new row orders to apply to other
matrices.
}
