% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{get_income}
\alias{get_income}
\title{Returns combined household incomes split into low, medium, and high groups}
\usage{
get_income(pdem02, subjects)
}
\arguments{
\item{pdem02}{Dataframe containing parent demographic information}

\item{subjects}{Dataframe containing list of required subjects}
}
\value{
income_df Dataframe containing household incomes
}
\description{
Low: $0 - $50k, Medium: $50k - $100k, High: > $100k
}
