% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{get_interview_age}
\alias{get_interview_age}
\title{Return dataframe containing interview age of specified subjects}
\usage{
get_interview_age(abcd_df, subjects = NULL)
}
\arguments{
\item{abcd_df}{Any ABCD dataframe containing interview age}

\item{subjects}{Dataframe containing list of required subjects}
}
\value{
interview_age Dataframe containing interview age
}
\description{
Return dataframe containing interview age of specified subjects
}
