% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_resilience.R
\name{get_loneliness}
\alias{get_loneliness}
\title{Extract loneliness}
\usage{
get_loneliness(abcd_ysr01, subjects = NULL)
}
\arguments{
\item{abcd_ysr01}{ABCD Other Resilience}

\item{subjects}{Dataframe containing list of required subjects}
}
\value{
loneliness
}
\description{
Extract loneliness
}
