% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{get_matrix_order}
\alias{get_matrix_order}
\title{Return the hierarchical clustering order of a matrix}
\usage{
get_matrix_order(matrix)
}
\arguments{
\item{matrix}{Matrix to cluster.}
}
\description{
Return the hierarchical clustering order of a matrix
}
