% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf.R
\name{get_mean_p}
\alias{get_mean_p}
\title{Get mean p-value}
\usage{
get_mean_p(output_matrix_row)
}
\arguments{
\item{output_matrix_row}{row of output_matrix object}
}
\value{
mean_p mean p-value
}
\description{
Given an output matrix row containing evaluated p-values, returns mean
}
