% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions_matrix.R
\name{get_mean_pval}
\alias{get_mean_pval}
\title{Get mean p-value}
\usage{
get_mean_pval(solutions_matrix_row)
}
\arguments{
\item{solutions_matrix_row}{row of solutions_matrix object}
}
\value{
mean_pval mean p-value
}
\description{
Given an solutions matrix row containing evaluated p-values, returns mean.
}
