% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_history.R
\name{get_med_history}
\alias{get_med_history}
\title{Returns race information}
\usage{
get_med_history(med_df, detailed_tbi_df)
}
\arguments{
\item{med_df}{Dataframe containing medical history information}

\item{detailed_tbi_df}{Dataframe containing tbi information}
}
\value{
med_history_df Dataframe containing important medical history vars
}
\description{
Returns race information
}
