% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf.R
\name{get_min_p}
\alias{get_min_p}
\title{Get minimum p-value}
\usage{
get_min_p(output_matrix_row)
}
\arguments{
\item{output_matrix_row}{row of output_matrix object}
}
\value{
min_p minimum p-value
}
\description{
Given an output matrix row containing evaluated p-values, returns minimum
}
