% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions_df.R
\name{get_min_pval}
\alias{get_min_pval}
\title{Get minimum p-value}
\usage{
get_min_pval(sol_df_row)
}
\arguments{
\item{sol_df_row}{row of sol_df object}
}
\value{
min_pval minimum p-value
}
\description{
Given an solutions data frame row containing evaluated p-values, returns min.
}
\keyword{internal}
