% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{get_mtbi_age}
\alias{get_mtbi_age}
\title{Get acute symptom input variable 'latest_mtbi_age'}
\usage{
get_mtbi_age(otbi01, subjects = NULL)
}
\arguments{
\item{otbi01}{The baseline TBI dataframe}

\item{subjects}{Dataframe containing list of required subjects}
}
\value{
mtbi_age Dataframe containing latest_mtbi_age
}
\description{
Get acute symptom input variable 'latest_mtbi_age'
}
