% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_history.R
\name{get_mtbi_count}
\alias{get_mtbi_count}
\title{Get number of mtbis sustained by subject}
\usage{
get_mtbi_count(otbi01, subjects)
}
\arguments{
\item{otbi01}{The baseline TBI dataframe}

\item{subjects}{Dataframe containing list of required subjects}
}
\value{
mtbi_count Dataframe containing number of previous mTBIs
}
\description{
Get number of mtbis sustained by subject
}
