% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acute_symptoms.R
\name{get_mtbi_loc}
\alias{get_mtbi_loc}
\title{Get acute symptom input variable 'latest_mtbi_loc'}
\usage{
get_mtbi_loc(otbi01, subjects)
}
\arguments{
\item{otbi01}{The baseline TBI dataframe}

\item{subjects}{Dataframe containing list of required subjects}
}
\value{
mtbi_loc Dataframe containing latest_mtbi_loc
}
\description{
Get acute symptom input variable 'latest_mtbi_loc'
}
