% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acute_symptoms.R
\name{get_mtbi_mem_daze}
\alias{get_mtbi_mem_daze}
\title{Get acute symptom input variable 'latest_mtbi_mem_daze'}
\usage{
get_mtbi_mem_daze(otbi01, subjects)
}
\arguments{
\item{otbi01}{The baseline TBI dataframe}

\item{subjects}{Dataframe containing list of required subjects}
}
\value{
mtbi_mem_daze Dataframe containing latest_mtbi_mem_daze
}
\description{
Get acute symptom input variable 'latest_mtbi_mem_daze'
}
