% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nihtb.R
\name{get_nihtbx_cardsort_fc}
\alias{get_nihtbx_cardsort_fc}
\title{Get nihtbx cardsort data}
\usage{
get_nihtbx_cardsort_fc(abcd_tbss01, subjects = NULL)
}
\arguments{
\item{abcd_tbss01}{NDA nihtbx dataframe}

\item{subjects}{cardsort of subjects to receive data for}
}
\value{
nihtbx_cardsort_fc cardsort data
}
\description{
Get nihtbx cardsort data
}
