% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nihtb.R
\name{get_nihtbx_list_fc}
\alias{get_nihtbx_list_fc}
\title{Get nihtbx list sorting data}
\usage{
get_nihtbx_list_fc(abcd_tbss01, subjects = NULL)
}
\arguments{
\item{abcd_tbss01}{NDA nihtbx dataframe}

\item{subjects}{list of subjects to receive data for}
}
\value{
nihtbx_list_fc list sorting data
}
\description{
Get nihtbx list sorting data
}
