% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nihtb.R
\name{get_nihtbx_pattern_fc}
\alias{get_nihtbx_pattern_fc}
\title{Get nihtbx pattern data}
\usage{
get_nihtbx_pattern_fc(abcd_tbss01, subjects = NULL)
}
\arguments{
\item{abcd_tbss01}{NDA nihtbx dataframe}

\item{subjects}{pattern of subjects to receive data for}
}
\value{
nihtbx_pattern_fc pattern data
}
\description{
Get nihtbx pattern data
}
