% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf.R
\name{get_p}
\alias{get_p}
\title{Get p-value}
\usage{
get_p(assigned_subs, outcome_component)
}
\arguments{
\item{assigned_subs}{dataframe of subjects who were assigned to a cluster and
the cluster they were assigned to}

\item{outcome_component}{the outcome_list element of interest}
}
\value{
p_val the smallest p-value of interest
}
\description{
Depending on outcome measure, perform ordinal regression or linear regression
and return p-value as a benchmark measure of how well-separated clusters
are by the outcome measure
}
