% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_p_vals_row_df}
\alias{get_p_vals_row_df}
\title{Return the p-values of a single settings_matrix row from a target_pvals df}
\usage{
get_p_vals_row_df(target_pvals, row)
}
\arguments{
\item{target_pvals}{Output of "p_val_select" function.}

\item{row}{Row ID of solutions_matrix to extract p-values for.}
}
\description{
Return the p-values of a single settings_matrix row from a target_pvals df
}
