% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_resilience.R
\name{get_parent_psychopathology}
\alias{get_parent_psychopathology}
\title{Extract parent psychopathology}
\usage{
get_parent_psychopathology(abcd_asrs01, subjects)
}
\arguments{
\item{abcd_asrs01}{ABCD Parent Adult Self Report Scores Aseba}

\item{subjects}{Dataframe containing list of required subjects}
}
\value{
parent_psychopathology
}
\description{
Extract parent psychopathology
}
