% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{get_prosocial_behaviour}
\alias{get_prosocial_behaviour}
\title{Extract prosocial behaviour}
\usage{
get_prosocial_behaviour(psb01, abcd_psb01, subjects)
}
\arguments{
\item{psb01}{Parent Prosocial Behavior Survey}

\item{abcd_psb01}{Youth Prosocial Behavior Survey}

\item{subjects}{Dataframe containing list of required subjects}
}
\value{
prosocial_behaviour
}
\description{
Extract prosocial behaviour
}
