% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{get_pubertal_status}
\alias{get_pubertal_status}
\title{Select columns relevant for demographic inputs}
\usage{
get_pubertal_status(youth_pubertal_df, parent_pubertal_df)
}
\arguments{
\item{youth_pubertal_df}{x}

\item{parent_pubertal_df}{x}
}
\value{
pubertal_status x
}
\description{
Select columns relevant for demographic inputs
}
