% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{get_race}
\alias{get_race}
\title{Returns race information}
\usage{
get_race(parent_demographics)
}
\arguments{
\item{parent_demographics}{x}
}
\value{
race_df x
}
\description{
Returns race information
}
