% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_resilience.R
\name{get_screen_time}
\alias{get_screen_time}
\title{Extract healthy behaviours: screen time questionnaire}
\usage{
get_screen_time(stq01, subjects = NULL)
}
\arguments{
\item{stq01}{ABCD Parent Screen Time Survey}

\item{subjects}{Dataframe containing list of required subjects}
}
\value{
screen_time
}
\description{
Extract healthy behaviours: screen time questionnaire
}
