% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{get_sex}
\alias{get_sex}
\title{Return dataframe containing sex of specified subjects}
\usage{
get_sex(abcd_df, subjects)
}
\arguments{
\item{abcd_df}{Any ABCD dataframe containing sex}

\item{subjects}{Dataframe containing list of required subjects}
}
\value{
sex Dataframe containing sex
}
\description{
Return dataframe containing sex of specified subjects
}
