% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_resilience.R
\name{get_sports_and_activities}
\alias{get_sports_and_activities}
\title{Extract healthy behaviours: spots and activities questionnaire}
\usage{
get_sports_and_activities(abcd_saiq02, subjects = NULL)
}
\arguments{
\item{abcd_saiq02}{ABCD Parent Sports and Activities Involvement
Questionnaire}

\item{subjects}{Dataframe containing list of required subjects}
}
\value{
activities
}
\description{
Extract healthy behaviours: spots and activities questionnaire
}
