% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroimaging.R
\name{get_subc_cor}
\alias{get_subc_cor}
\title{Extract subcortical network correlations}
\usage{
get_subc_cor(mrirscor02, subjects)
}
\arguments{
\item{mrirscor02}{Data file containing neurite density data}

\item{subjects}{Dataframe containing list of required subjects}
}
\value{
subc_cor Dataframe of white matter neurite densities
}
\description{
Extract subcortical network correlations
}
