% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroimaging.R
\name{get_subc_v}
\alias{get_subc_v}
\title{Extract subcortical volumes}
\usage{
get_subc_v(smrip10201, subjects)
}
\arguments{
\item{smrip10201}{Data file containing subcortical data}

\item{subjects}{Dataframe containing list of required subjects}
}
\value{
subc_v_df Dataframe of subcortical volumes
}
\description{
Extract subcortical volumes
}
