% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gexclude}
\alias{gexclude}
\title{Helper function to drop columns from a data frame by grepl search}
\usage{
gexclude(x, pattern)
}
\arguments{
\item{x}{Data frame to drop columns from.}

\item{pattern}{Pattern used to match columns to drop.}
}
\value{
x without columns matching pattern.
}
\description{
Helper function to drop columns from a data frame by grepl search
}
\keyword{internal}
