% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_metrics.R
\name{gower_distance}
\alias{gower_distance}
\title{Distance metric: Gower distance}
\usage{
gower_distance(df)
}
\arguments{
\item{df}{Dataframe containing one subjectkey column and at least 1 data
column}
}
\value{
distance_matrix A distance matrix.
}
\description{
Distance metric: Gower distance
}
