% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gselect}
\alias{gselect}
\title{Helper function to pick columns from a data frame by grepl search}
\usage{
gselect(x, pattern)
}
\arguments{
\item{x}{Data frame to select columns from.}

\item{pattern}{Pattern used to match columns to select.}
}
\value{
x with only columns matching pattern.
}
\description{
Helper function to pick columns from a data frame by grepl search
}
\keyword{internal}
