% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_metrics_list.R
\name{hamming_distance}
\alias{hamming_distance}
\title{Distance metric: Hamming distance}
\usage{
hamming_distance(df, weights_row)
}
\arguments{
\item{df}{Dataframe containing one subjectkey column in the first column and
at least 1 categorical data column. All feature data should be categorical.}

\item{weights_row}{Single-row dataframe where the column names contain the
column names in df and the row contains the corresponding weights.}
}
\value{
distance_matrix A distance matrix.
}
\description{
Distance metric: Hamming distance
}
