% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{heatmap_pvals}
\alias{heatmap_pvals}
\title{Heatmap p-value matrix}
\usage{
heatmap_pvals(p_val_matrix, file_path = NA)
}
\arguments{
\item{p_val_matrix}{matrix of p-values}

\item{file_path}{where to store heatmap}
}
\description{
Heatmap p-value matrix
}
