% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbi.R
\name{identify_latest_mtbi_mechanism}
\alias{identify_latest_mtbi_mechanism}
\title{Add columns to a TBI dataframe indicating mechanism of the latest mTBI}
\usage{
identify_latest_mtbi_mechanism(tbi_df)
}
\arguments{
\item{tbi_df}{A TBI dataframe}
}
\value{
df_mech The modified dataframe
}
\description{
Add columns to a TBI dataframe indicating mechanism of the latest mTBI
}
