% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbi.R
\name{identify_latest_mtbi_mem_daze}
\alias{identify_latest_mtbi_mem_daze}
\title{Add columns to a TBI dataframe indicating if their latest mTBI had mem/daze}
\usage{
identify_latest_mtbi_mem_daze(tbi_df)
}
\arguments{
\item{tbi_df}{A TBI dataframe}
}
\value{
dfmd The modified dataframe
}
\description{
Add columns to a TBI dataframe indicating if their latest mTBI had mem/daze
}
