% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbi.R
\name{identify_mtbi}
\alias{identify_mtbi}
\title{Generate columns indicating which injury types were mTBIs}
\usage{
identify_mtbi(tbi_df)
}
\arguments{
\item{tbi_df}{A TBI dataframe}
}
\value{
df_mtbi The modified dataframe
}
\description{
Generate columns indicating which injury types were mTBIs
}
