% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbi.R
\name{identify_num_mtbi}
\alias{identify_num_mtbi}
\title{Add columns to a TBI dataframe indicating a subject's estimated mTBI count}
\usage{
identify_num_mtbi(tbi_df)
}
\arguments{
\item{tbi_df}{A TBI dataframe}
}
\value{
df_num_mtbi The modified dataframe
}
\description{
Add columns to a TBI dataframe indicating a subject's estimated mTBI count
}
