% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{is_data_list}
\alias{is_data_list}
\title{Test if the object is a data list}
\usage{
is_data_list(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if the object inherits from the \code{data_list} class.
}
\description{
Given an object, returns \code{TRUE} if that object inherits from the \code{data_list}
class.
}
