% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_type.R
\name{is_valid}
\alias{is_valid}
\title{Check if an object is properly formatted}
\usage{
is_valid(x, validator_fn)
}
\arguments{
\item{x}{The object to check for validity}

\item{validator_fn}{Helper function containing tryCatch statement}
}
\value{
TRUE if the object is a valid member of its main class, FALSE
otherwise.
}
\description{
This function coerces non-\code{data_list} class objects into \code{data_list} class
objects.
}
