% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_plots.R
\name{jitter_plot}
\alias{jitter_plot}
\title{Jitter plot separating a feature by cluster}
\usage{
jitter_plot(df, feature)
}
\arguments{
\item{df}{A data.frame containing cluster column and the feature to plot.}

\item{feature}{The feature to plot.}
}
\value{
A jitter+violin plot (class "gg", "ggplot") showing the
distribution of a feature across clusters.
}
\description{
Jitter plot separating a feature by cluster
}
