% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{keep_split}
\alias{keep_split}
\title{Filter data to training or testing subjects only}
\usage{
keep_split(df, assigned_df, split)
}
\arguments{
\item{df}{Dataframe to be subsetted into training or testing split}

\item{assigned_df}{Dataframe containing "subjectkey" and "split" cols from
\code{train_test_assign()}}

\item{split}{String indicating which split to keep ("train" or "test")}
}
\value{
split_df Dataframe subsetted to specified split
}
\description{
Given a dataframe and the results of \code{train_test_split()}, return just the
data for subjects that were assigned the specified split
}
