% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_meta_clusters.R
\name{label_meta_clusters}
\alias{label_meta_clusters}
\title{Assign meta cluster labels to rows of a solutions data frame or extended
solutions data frame}
\usage{
label_meta_clusters(sol_df, split_vector, order = NULL)
}
\arguments{
\item{sol_df}{A solutions data frame or extended solutions data frame to
assign meta clusters to.}

\item{split_vector}{A numeric vector indicating which rows of sol_df should
be the split points for meta cluster labeling.}

\item{order}{An optional numeric vector indicating how the solutions data
frame should be reordered prior to meta cluster labeling. This vector can
be obtained by running \code{get_matrix_order()} on an ARI matrix, which itself
can be obtained by calling \code{calc_aris()} on a solutions data frame.}
}
\value{
A solutions data frame with a populated "meta_clusters" attribute.
}
\description{
Given a solutions data frame or extended solutions data frame class object
and a numeric vector indicating which rows correspond to which meta
clusters, assigns meta clustering information to the "meta_clusters"
attribute of the data frame.
}
