% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_propagation.R
\name{label_propagate}
\alias{label_propagate}
\title{Label propagate cluster solutions to unclustered subjects}
\usage{
label_propagate(train_sol_df, full_dl, verbose = FALSE)
}
\arguments{
\item{train_sol_df}{A solutions data frame derived from the training set.}

\item{full_dl}{A data list containing subjects from both the training
and testing sets.}

\item{verbose}{If TRUE, output progress to console.}
}
\value{
A data frame with one row per observation containing a column for
UIDs, a column for whether the subject was in the train (original) or test
(held out) set, and one column per row of the solutions data frame
indicating the original and propagated clusters.
}
\description{
Given a sol_df derived from training subjects and a full_data_list
containing both training and test subjects, re-run SNF to generate a total
affinity matrix of both train and subjects and use the label propagation
algorithm to assigned predicted clusters to test subjects.
}
