% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{label_splits}
\alias{label_splits}
\title{Convert a vector of partition indices into meta cluster labels}
\usage{
label_splits(split_vector, nrow)
}
\arguments{
\item{split_vector}{A vector of partition indices.}

\item{nrow}{The number of rows in the data being partitioned.}
}
\description{
Convert a vector of partition indices into meta cluster labels
}
