% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions_matrix.R
\name{linear_model_pval}
\alias{linear_model_pval}
\title{Linear model p-value (generic)}
\usage{
linear_model_pval(predictor, response)
}
\arguments{
\item{predictor}{A categorical or numeric feature.}

\item{response}{A numeric feature.}
}
\value{
pval A p-value (class "numeric").
}
\description{
Return p-value of F-test for a linear model of any two features
}
