% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_propagation.R
\name{lp_row}
\alias{lp_row}
\title{Label propagation over an solutions matrix}
\usage{
lp_row(om, full_data_list, distance_metrics_list = NULL)
}
\arguments{
\item{om}{An solutions matrix}

\item{full_data_list}{A data_list object made by rbinding(train, test) data}

\item{distance_metrics_list}{A distance_metrics_list.
See ?generate_distance_metrics_list.}
}
\value{
labeled_df A dataframe of the label propagated results of all om rows
}
\description{
Given an solutions matrix and a data_list object containing all subjects, return
a dataframe of the label propagated results of all om rows
}
