% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{manhattan_plot}
\alias{manhattan_plot}
\title{(DEPRECATED) Manhattan plot creation}
\usage{
manhattan_plot(...)
}
\arguments{
\item{...}{Function arguments}
}
\description{
(DEPRECATED) Manhattan plot creation
}
