% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{mc_heatmap}
\alias{mc_heatmap}
\title{Heatmap meta-clustering results}
\usage{
mc_heatmap(mc_results, save = NULL)
}
\arguments{
\item{mc_results}{results from meta_cluster function}

\item{save}{optional path to save figure to}
}
\description{
Heatmap meta-clustering results
}
