% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_plot.R
\name{mc_manhattan_plot}
\alias{mc_manhattan_plot}
\title{Manhattan plot of variable-meta cluster associaiton p-values}
\usage{
mc_manhattan_plot(
  rep_solution,
  threshold = NULL,
  data_list = NULL,
  target_list = NULL,
  variable_order = NULL,
  xints = NULL,
  text_size = 20,
  point_size = 5,
  neg_log_pval_thresh = 5,
  hide_x_labels = FALSE,
  domain_colours = NULL
)
}
\arguments{
\item{rep_solution}{The dataframe of representative solutions from the
\code{get_representative_solutions()} function.}

\item{threshold}{p-value threshold to plot horizontal dashed line at.}

\item{data_list}{List of dataframes containing data information.}

\item{target_list}{List of dataframes containing target information.}

\item{variable_order}{Order of variables to be displayed in the plot.}

\item{xints}{Either "outcomes" or a vector of numeric values to plot
vertical lines at.}

\item{text_size}{Size of text in the plot.}

\item{point_size}{Size of points in the plot.}

\item{neg_log_pval_thresh}{Threshold for negative log p-values.}

\item{hide_x_labels}{If TRUE, hides x-axis labels.}

\item{domain_colours}{Named vector of colours for domains.}
}
\description{
Given a dataframe of representative meta cluster solutions (see
\code{get_representative_solutions()}, returns a Manhattan plot for showing
variable separation across all variables in provided data/target_lists.
}
