% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge.data_list}
\alias{merge.data_list}
\title{Merge observations between two compatible data lists}
\usage{
\method{merge}{data_list}(x, y, ...)
}
\arguments{
\item{x}{The first data list to merge.}

\item{y}{The second data list to merge.}

\item{...}{Additional arguments passed into merge function.}
}
\value{
A data list ("list"-class object) containing the observations of
both provided data lists.
}
\description{
Join two data lists with the same components (data frames) but separate
observations. To instead merge two data lists that have the same
observations but different components, simply use \code{c()}.
}
