% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge.dist_fns_list}
\alias{merge.dist_fns_list}
\title{Merge \code{dist_fns_list} objects}
\usage{
\method{merge}{dist_fns_list}(x, y, ...)
}
\arguments{
\item{x}{The first \code{clust_fns_list} object to merge.}

\item{y}{The second \code{clust_fns_list} object to merge.}

\item{...}{Additional arguments (not used).}
}
\value{
A new \code{clust_fns_list} object containing the merged clustering
functions.
}
\description{
Merge \code{dist_fns_list} objects
}
