% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge.sim_mats_list}
\alias{merge.sim_mats_list}
\title{Merge \code{sim_mats_list} objects}
\usage{
\method{merge}{sim_mats_list}(x, y, ...)
}
\arguments{
\item{x}{The first \code{sim_mats_list} object to merge.}

\item{y}{The second \code{sim_mats_list} object to merge.}

\item{...}{Additional arguments (not used).}
}
\value{
A merged \code{sim_mats_list} object containing the similarity matrices
from both input objects.
}
\description{
Merge \code{sim_mats_list} objects
}
