% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge.snf_config}
\alias{merge.snf_config}
\title{Merge method for SNF config objects}
\usage{
\method{merge}{snf_config}(x, y, reset_indices = TRUE, ...)
}
\arguments{
\item{x}{SNF config to merge.}

\item{y}{SNF config to merge.}

\item{reset_indices}{If TRUE (default), re-labels the "solutions" indices in
the config from 1 to the number of defined settings.}

\item{...}{Additional arguments passed into merge function.}
}
\value{
An SNF config combining the rows of both prior configurations.
}
\description{
Merge method for SNF config objects
}
