% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge.solutions_df}
\alias{merge.solutions_df}
\title{Merge \code{solutions_df} objects}
\usage{
\method{merge}{solutions_df}(x, y, ...)
}
\arguments{
\item{x}{The first \code{solutions_df} object to merge.}

\item{y}{The second \code{solutions_df} object to merge.}

\item{...}{Additional arguments (not used).}
}
\value{
Error message indicating that the merge function is not applicable
to \code{solutions_df} objects.
}
\description{
Merge \code{solutions_df} objects
}
