% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge.weights_matrix}
\alias{merge.weights_matrix}
\title{Merge \code{weights_matrix} objects}
\usage{
\method{merge}{weights_matrix}(x, y, ...)
}
\arguments{
\item{x}{The first \code{weights_matrix} object to merge.}

\item{y}{The second \code{weights_matrix} object to merge.}

\item{...}{Additional arguments (not used).}
}
\value{
Error message indicating that the merge function is not applicable
to \code{weights_matrix} objects.
}
\description{
Merge \code{weights_matrix} objects
}
