% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{merge_data_lists}
\alias{merge_data_lists}
\title{Horizontally merge compatible data lists}
\usage{
merge_data_lists(data_list1, data_list2)
}
\arguments{
\item{data_list1}{The first data_list to merge.}

\item{data_list2}{The second data_list to merge.}
}
\value{
A data list ("list"-class object) containing the observations of
both provided data lists.
}
\description{
Join two data_lists with the same components (dataframes) but separate
observations. To instead merge two data_lists that have the same
observations but different components, simply use \code{c()}.
}
