% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_df_list}
\alias{merge_df_list}
\title{Merge list of dataframes}
\usage{
merge_df_list(df_list, join = "inner", uid = "subjectkey", no_na = FALSE)
}
\arguments{
\item{df_list}{list of dataframes}

\item{join}{String indicating if join should be "inner" or "full"}

\item{uid}{Column name to join on. Default is "subjectkey"}

\item{no_na}{Whether to remove NA values from the merged dataframe}
}
\value{
merged_df inner join of all dataframes in list
}
\description{
Merge list of dataframes
}
