% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{merge_dls}
\alias{merge_dls}
\title{Horizontally merge compatible data lists}
\usage{
merge_dls(dl_1, dl_2)
}
\arguments{
\item{dl_1}{The first data_list to merge.}

\item{dl_2}{The second data_list to merge.}
}
\value{
A data list ("list"-class object) containing the observations of
both provided data lists.
}
\description{
Join two data_lists with the same components (dataframes) but separate
observations. To instead merge two data_lists that have the same
observations but different components, simply use \code{c()}.
}
