% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_cluster.R
\name{meta_cluster}
\alias{meta_cluster}
\title{Meta-cluster calculations}
\usage{
meta_cluster(mc_om)
}
\arguments{
\item{mc_om}{output_matrix}
}
\value{
mc_sm meta-clustering similarity matrix
}
\description{
Generate matrix of pairwise cluster-solution similarities by Adjusted Rand
index calculations
}
