% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{metasnf_alert}
\alias{metasnf_alert}
\title{Helper function for raising alerts}
\usage{
metasnf_alert(..., env = 1)
}
\arguments{
\item{...}{Arbitrary number of strings to be pasted together into alert
message.}

\item{env}{}
}
\value{
Returns no value. Raises an error through cli::cli_abort.
}
\description{
Helper function for raising alerts
}
\keyword{internal}
