% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{metasnf_deprecated}
\alias{metasnf_deprecated}
\title{Helper function for deprecated function warnings}
\usage{
metasnf_deprecated(version, alternative, env = 1)
}
\arguments{
\item{version}{Version of \code{metasnf} in which function has been deprecated.}

\item{alternative}{Recommended alternative approach.}

\item{env}{Environment to evaluate expressions in.}
}
\value{
Returns no value. Raises a warning through cli::cli_warn.
}
\description{
Helper function for deprecated function warnings
}
\keyword{internal}
