% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{metasnf_error}
\alias{metasnf_error}
\title{Helper function for raising errors}
\usage{
metasnf_error(..., env = 1)
}
\arguments{
\item{...}{Arbitrary number of strings to be pasted together into error
message.}

\item{env}{}
}
\value{
Returns no value. Raises an error through cli::cli_abort.
}
\description{
Helper function for raising errors
}
\keyword{internal}
