% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{methylation_df}
\alias{methylation_df}
\title{Modification of SNFtool mock data frame "Data2"}
\format{
\subsection{\code{methylation_df}}{

A data frame with 200 rows and 3 columns:
\describe{
\item{gene_1_expression}{Mock gene methylation feature}
\item{gene_2_expression}{Mock gene methylation feature}
\item{patient_id}{Random three-digit number uniquely identifying the patient}
}
}
}
\source{
This data came from the SNFtool package, with slight modifications.
}
\usage{
methylation_df
}
\description{
Modification of SNFtool mock data frame "Data2"
}
\keyword{datasets}
