% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mock_ari_matrix}
\alias{mock_ari_matrix}
\title{Mock example of an \code{ari_matrix} metasnf object}
\format{
\subsection{\code{mock_ari_matrix}}{

A 20 by 20 ARI matrix.
}
}
\source{
This data comes from the metasnf package.
}
\usage{
mock_ari_matrix
}
\description{
An \code{ari_matrix} class object containing adjusted Rand indices (ARIs) between 20 cluster solutions.
Used as an example of an \code{ari_matrix} metasnf object.
}
\keyword{datasets}
