% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mock_clust_fns_list}
\alias{mock_clust_fns_list}
\title{Mock example of a \code{clust_fns_list} metasnf object}
\format{
\subsection{\code{mock_clust_fns_list}}{

A \code{clust_fns_list} object containing two clustering functions covering 2 and 5 five cluster solution versions of spectral clustering.
Extracted from \code{mock_snf_config}.
}
}
\source{
This data comes from the metasnf package.
}
\usage{
mock_clust_fns_list
}
\description{
Mock example of a \code{clust_fns_list} metasnf object
}
\keyword{datasets}
