% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mock_data_list}
\alias{mock_data_list}
\title{Mock example of a \code{data_list} metasnf object}
\format{
\subsection{\code{mock_data_list}}{

A data list containing 4 data frames with 100 observations each:
- subcortical volume (30 features)
- cortical surface area (151 features)
- household income (1 feature)
- pubertal status (1 feature)
Used as an example of an \code{data_list} metasnf object.
}
}
\source{
This data comes from the metasnf package.
}
\usage{
mock_data_list
}
\description{
Mock example of a \code{data_list} metasnf object
}
\keyword{datasets}
