% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mock_dist_fns_list}
\alias{mock_dist_fns_list}
\title{Mock example of a \code{dist_fns_list} metasnf object}
\format{
\subsection{\code{mock_dist_fns_list}}{

A \code{dist_fns_list} object containing a variety of distance metrics.
Extracted from \code{mock_snf_config}.
}
}
\source{
This data comes from the metasnf package.
}
\usage{
mock_dist_fns_list
}
\description{
Mock example of a \code{dist_fns_list} metasnf object
}
\keyword{datasets}
