% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mock_ext_solutions_df}
\alias{mock_ext_solutions_df}
\title{Mock example of a \code{ext_solutions_df} metasnf object}
\format{
\subsection{\code{mock_ext_solutions_df}}{

Contains 20 cluster solutions.
}
}
\source{
This data comes from the metasnf package.
}
\usage{
mock_ext_solutions_df
}
\description{
An \code{ext_solutions_df} class object generated by extending the \code{mock_rep_solutions_df} object against \code{mock_data_list} as the target data list.
}
\keyword{datasets}
