% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mock_mc_solutions_df}
\alias{mock_mc_solutions_df}
\title{Mock example of a \code{mc_solutions_df} metasnf object}
\format{
\subsection{\code{mock_mc_solutions_df}}{

A meta cluster labeled solutions data frame derived from \code{mock_solutions_df}.
Contains 20 cluster solutions.
}
}
\source{
This data comes from the metasnf package.
}
\usage{
mock_mc_solutions_df
}
\description{
Mock example of a \code{mc_solutions_df} metasnf object
}
\keyword{datasets}
