% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mock_rep_solutions_df}
\alias{mock_rep_solutions_df}
\title{Mock example of a \code{rep_solutions_df} metasnf object}
\format{
\subsection{\code{mock_rep_solutions_df}}{

Contains 4 cluster solutions.
}
}
\source{
This data comes from the metasnf package.
}
\usage{
mock_rep_solutions_df
}
\description{
A \code{solutions_df} class object derived by filtering the \code{mock_mc_solutions_df} to its representative solutions.
}
\keyword{datasets}
