% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mock_settings_df}
\alias{mock_settings_df}
\title{Mock example of a \code{settings_df} metasnf object}
\format{
\subsection{\code{mock_settings_df}}{

Settings for 20 cluster solutions.
}
}
\source{
This data comes from the metasnf package.
}
\usage{
mock_settings_df
}
\description{
Mock example of a \code{settings_df} metasnf object
}
\keyword{datasets}
