% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mock_sim_mats_list}
\alias{mock_sim_mats_list}
\title{Mock example of an \code{sim_mats_list} metasnf object}
\format{
\subsection{\code{mock_sim_mats_list}}{

A \code{sim_mats_list} class object containing pairwise similarities between 100
observations across 20 cluster solutions.
}
}
\source{
This data comes from the metasnf package.
}
\usage{
mock_sim_mats_list
}
\description{
Used as an example of an \code{sim_mats_list} metasnf object.
}
\keyword{datasets}
