% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mock_snf_config}
\alias{mock_snf_config}
\title{Mock example of a \code{snf_config} metasnf object}
\format{
\subsection{\code{mock_snf_config}}{

An SNF config containing hyperparameters and functions defined for generating 20 cluster solutions from a data list.
The config has been specified to:
- limit the \code{k} hyperparameter to 40
- make use of uniformly distributed random weights
- randomly select between using spectral clustering where the number of clusters can be 2, 5, decided by the eigen-gap heuristic, or decided by the rotation cost heuristic
- use Gower distance for categorical and mixed data, Euclidean distance for ordinal data, and randomly select from Euclidean distance or standard/normalized Euclidean distance for continuous and discrete data
The config was built using the \code{mock_data_list} loaded into the namespace after calling \code{library("metasnf")}.
Used as an example of an \code{snf_config} metasnf object.
}
}
\source{
This data comes from the metasnf package.
}
\usage{
mock_snf_config
}
\description{
Mock example of a \code{snf_config} metasnf object
}
\keyword{datasets}
