% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mock_solutions_df}
\alias{mock_solutions_df}
\title{Mock example of a \code{solutions_df} metasnf object}
\format{
\subsection{\code{mock_solutions_df}}{

A solutions data frame containing 20 cluster solutions generated from \code{mock_snf_config} and \code{mock_data_list}.
Used as an example of an \code{solutions_df} metasnf object.
}
}
\source{
This data comes from the metasnf package.
}
\usage{
mock_solutions_df
}
\description{
Mock example of a \code{solutions_df} metasnf object
}
\keyword{datasets}
