% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mock_weights_matrix}
\alias{mock_weights_matrix}
\title{Mock example of a \code{weights_matrix} metasnf object}
\format{
\subsection{\code{mock_weights_matrix}}{

A \code{weights_matrix} class object containing 20 sets of weights for 183 features.
}
}
\source{
This data comes from the metasnf package.
}
\usage{
mock_weights_matrix
}
\description{
Mock example of a \code{weights_matrix} metasnf object
}
\keyword{datasets}
