% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_features.R
\name{n_features}
\alias{n_features}
\title{Extract number of features stored in an object}
\usage{
n_features(x)
}
\arguments{
\item{x}{The object to extract number of features from.}
}
\value{
The number of features in x.
}
\description{
Extract number of features stored in an object
}
