% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_observations.R
\name{n_observations}
\alias{n_observations}
\title{Extract number of observations stored in an object}
\usage{
n_observations(x)
}
\arguments{
\item{x}{The object to extract number of observations from.}
}
\value{
The number of observations in x.
}
\description{
Extract number of observations stored in an object
}
\keyword{internal}
