% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ari_matrix.R
\name{new_ari_matrix}
\alias{new_ari_matrix}
\title{Constructor for \code{ari_matrix} class object}
\usage{
new_ari_matrix(aml, dist_method, hclust_method)
}
\arguments{
\item{aml}{An ari_matrix-like matrix object to be validated.}
}
\value{
A \code{ari_matrix} object.
}
\description{
Constructor for \code{ari_matrix} class object
}
\keyword{internal}
