% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{new_data_list}
\alias{new_data_list}
\title{Constructor for data_list class object}
\usage{
new_data_list(x)
}
\arguments{
\item{x}{A nested list to be converted into a \code{data_list} object.}
}
\value{
A \code{data_list} object, which is a nested list with class \code{data_list}.
}
\description{
Constructor for data_list class object
}
\keyword{internal}
