% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_fns_list.R
\name{new_dist_fns_list}
\alias{new_dist_fns_list}
\title{Constructor for \code{dist_fns_list} class object}
\usage{
new_dist_fns_list(dfll)
}
\arguments{
\item{dfll}{A distance metrics list-like list object to be validated.}
}
\value{
A \code{dist_fns_list} object.
}
\description{
Constructor for \code{dist_fns_list} class object
}
\keyword{internal}
