% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_solutions_df.R
\name{new_ext_solutions_df}
\alias{new_ext_solutions_df}
\title{Constructor for \code{ext_solutions_df} class object}
\usage{
new_ext_solutions_df(ext_sol_dfl)
}
\arguments{
\item{ext_sol_dfl}{An extended solutions data frame-like object.}
}
\value{
An \code{ext_solutions_df} object, which is a data frame with class
\code{ext_solutions_df}.
}
\description{
Constructor for \code{ext_solutions_df} class object
}
\keyword{internal}
