% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_df.R
\name{new_settings_df}
\alias{new_settings_df}
\title{Constructor for \code{settings_df} class object}
\usage{
new_settings_df(sdfl)
}
\arguments{
\item{sdfl}{A settings data frame-like matrix object to be validated.}
}
\value{
A \code{settings_df} object.
}
\description{
Constructor for \code{settings_df} class object
}
\keyword{internal}
