% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mats_list.R
\name{new_sim_mats_list}
\alias{new_sim_mats_list}
\title{Constructor for \code{similarity_matrix_list} class object}
\usage{
new_sim_mats_list(smll)
}
\value{
RETURN
}
\description{
Constructor for \code{similarity_matrix_list} class object
}
