% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf_config.R
\name{new_snf_config}
\alias{new_snf_config}
\title{Constructor for \code{snf_config} class object}
\usage{
new_snf_config(scl)
}
\arguments{
\item{scl}{An SNF config-like \code{list} class object.}
}
\value{
An \code{snf_config} object.
}
\description{
Constructor for \code{snf_config} class object
}
\keyword{internal}
