% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf_results.R
\name{new_snf_results}
\alias{new_snf_results}
\title{Constructor for \code{snf_results} class object}
\usage{
new_snf_results(srl)
}
\arguments{
\item{srl}{A snf_results-like list object to be validated.}
}
\value{
An \code{snf_results} object.
}
\description{
Constructor for \code{snf_results} class object
}
\keyword{internal}
