% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions_df.R
\name{new_solutions_df}
\alias{new_solutions_df}
\title{Constructor for \code{solutions_df} class object}
\usage{
new_solutions_df(sol_dfl)
}
\value{
RETURN
}
\description{
Constructor for \code{solutions_df} class object
}
