% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_matrix.R
\name{new_weights_matrix}
\alias{new_weights_matrix}
\title{Constructor for \code{new_weights_matrix} class object}
\usage{
new_weights_matrix(wml)
}
\value{
A \code{new_weights_matrix} object.
}
\description{
Constructor for \code{new_weights_matrix} class object
}
\keyword{internal}
