% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{numcol_to_numeric}
\alias{numcol_to_numeric}
\title{Convert columns of a data frame to numeric type (if possible)}
\usage{
numcol_to_numeric(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
The data frame coercible columns converted to type numeric.
}
\description{
Converts all columns in a data frame that can be converted to numeric type to
numeric type.
}
\keyword{internal}
