% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{om_plot_all_cbcl}
\alias{om_plot_all_cbcl}
\title{Make a full CBCL grid plot for an output matrix}
\usage{
om_plot_all_cbcl(om, cbcl_list, fig_path_fn)
}
\arguments{
\item{om}{An output matrix (or dataframe with om-like structure)}

\item{cbcl_list}{A list of all cbcl dataframes}

\item{fig_path_fn}{A function that generates a figure path}
}
\description{
Make a full CBCL grid plot for an output matrix
}
