% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{om_scatter}
\alias{om_scatter}
\title{Scatter plot alpha and k hyperparameter results by minimum and mean p-values}
\usage{
om_scatter(om)
}
\arguments{
\item{om}{output matrix}
}
\description{
Scatter plot alpha and k hyperparameter results by minimum and mean p-values
}
