% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmi.R
\name{om_to_nmi_df}
\alias{om_to_nmi_df}
\title{Calculate NMI scores for an output matrix}
\usage{
om_to_nmi_df(om, data_list)
}
\arguments{
\item{om}{Output matrix}

\item{data_list}{A data list}
}
\value{
nmi_df Merged dataframe of all NMI values
}
\description{
Given an output matrix in dataframe form with "significance" columns,
calculate NMIs for all inputs and combine results into a single dataframe
}
