% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf.R
\name{ord_reg_min_p}
\alias{ord_reg_min_p}
\title{Ordinal regression minimum p-value}
\usage{
ord_reg_min_p(clust_membership, outcome_df, outcome_var)
}
\arguments{
\item{clust_membership}{Dataframe of cluster membership (get_clustered_subs)}

\item{outcome_df}{Dataframe containing DV}

\item{outcome_var}{DV as a string}
}
\value{
min_p_val The smallest p-value distinguishing clusters by the DV
}
\description{
Returns the smallest p-value following an ordinal regression in which cluster
is the IV and a provided ordinal variable is the DV
}
