% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_tbi.R
\name{original_otbi_names}
\alias{original_otbi_names}
\title{Return conversion table of original and new otbi names}
\usage{
original_otbi_names()
}
\description{
Return conversion table of original and new otbi names
}
