% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_plotting.R
\name{output_cluster_file}
\alias{output_cluster_file}
\title{Spectral clustering from c2 to user-defined max-number of cluster. Used by plotClustersAlluvial_wOutcome function}
\usage{
output_cluster_file(similarity_matrix, numc = 8)
}
\arguments{
\item{similarity_matrix}{A similarity matrix.}

\item{numc}{max number of cluster to do spectral clustering.}
}
\value{
result_df dataframe of cluster assignments
}
\description{
Do spectral clustering through cluster 2 to user defined cluster number and outputs clustering result
}
