% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pick_cols}
\alias{pick_cols}
\title{Helper function to pick columns from a data frame}
\usage{
pick_cols(x, cols)
}
\arguments{
\item{x}{A data frame}

\item{cols}{Vector of column names to be picked}
}
\value{
x with only columns in cols
}
\description{
Helper function to pick columns from a data frame
}
\keyword{internal}
