% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pl}
\alias{pl}
\title{Helper function to pluralize a string}
\usage{
pl(x)
}
\arguments{
\item{x}{A vector of length 1 or greater.}
}
\value{
A string "s" if the length of x is greater than 1, otherwise an
empty string.
}
\description{
Helper function to pluralize a string
}
\keyword{internal}
