% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_all_cbcl}
\alias{plot_all_cbcl}
\title{Plot clusters across all CBCL measures}
\usage{
plot_all_cbcl(om_row, outcome_list, save = NULL)
}
\arguments{
\item{om_row}{Row of an output matrix to be plotted}

\item{outcome_list}{List containing all CBCL dataframes}

\item{save}{optional path to save figure to}
}
\description{
Arrange and optionally save a grid of bar charts with cluster on the x-axis
and each CBCL measure on the y-axis.
}
