% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_nmis}
\alias{plot_nmis}
\title{Scatter plot the NMI values for each row of an nmi_df}
\usage{
plot_nmis(nmi_df, fig_path_fn = NULL)
}
\arguments{
\item{nmi_df}{A dataframe an input column and any number of NMI columns}

\item{fig_path_fn}{A function that returns a full file path location}
}
\description{
Scatter plot the NMI values for each row of an nmi_df
}
