% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_plotting.r
\name{plotClustersAlluvial_wDiagnosis}
\alias{plotClustersAlluvial_wDiagnosis}
\title{Plot alluvial plot}
\usage{
plotClustersAlluvial_wDiagnosis(W, df_env, numc, outcome, key_outcome = NULL)
}
\arguments{
\item{W}{a similarity matrix with sample ids in rownames and colnames}

\item{df_env}{dataframe of environmental or outcome variables to color the alluvium. With "Subject_Number" column filled with sample ids}

\item{numc}{number of clusters}

\item{outcome}{column name from df_env to fill the alluvium with}

\item{key_outcome}{optional key outcome such as "Disease Diagnosis" that is the ultimate outcome. To be displayed as the last x-axis alluvia}
}
\description{
Plot alluvial plot from 2 clusters to user-defined number of clusters along with outcome/environmental variables.
}
