% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coclustering.R
\name{pooled_cocluster_heatmap}
\alias{pooled_cocluster_heatmap}
\title{Plot a co-clustering heatmap across multiple settings_matrix rows}
\usage{
pooled_cocluster_heatmap(cocluster_list)
}
\arguments{
\item{cocluster_list}{A list of cocluster_data objects (from
?generate_cocluster_data) to pool together for the heatmap.}
}
\description{
Generate a heatmap using data from multiple cocluster_data objects. See
?generate_cocluster_data for more information. The resulting heatmap shows
how often two patients clustered together off of several resamplings of the
data.
}
