% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{prefix_dl_uid}
\alias{prefix_dl_uid}
\title{Add "uid_" prefix to all UID values in uid column}
\usage{
prefix_dl_uid(dl)
}
\arguments{
\item{dl}{A nested list of input data from \code{data_list()}.}
}
\value{
dl A data list with UIDs prefixed with the string "uid_"
}
\description{
Add "uid_" prefix to all UID values in uid column
}
