% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{prefix_dll_uid}
\alias{prefix_dll_uid}
\title{Add "uid_" prefix to all UID values in uid column}
\usage{
prefix_dll_uid(dll)
}
\arguments{
\item{dll}{A data list-like \code{list} class object.}
}
\value{
A data list with UIDs prefixed with the string "uid_".
}
\description{
Add "uid_" prefix to all UID values in uid column
}
\keyword{internal}
