% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_plotting.R
\name{prepare_for_alluvial_wDiagnosis}
\alias{prepare_for_alluvial_wDiagnosis}
\title{Prepare dataframe for multi-cluster alluvial plot}
\usage{
prepare_for_alluvial_wDiagnosis(
  cluster_env_df,
  fill_alluvium_by,
  numc,
  key_outcome = NULL
)
}
\arguments{
\item{cluster_env_df}{is a dataframe with clustering assignments and environmental variables}

\item{fill_alluvium_by}{variable to color alluvium. This is an outcome variable in cluster_env_df}

\item{numc}{user-defined max number of clusters that went through spectral clustering}

\item{key_outcome}{optional key outcome such as "Disease Diagnosis" that is the ultimate outcome. To be displayed on the last x-axis alluvia}
}
\value{
df wide format dataframe
}
\description{
Calculates the frequency of different data combinations and prepares dataframe. To be used by cluster_alluvial function
}
