% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ari_matrix}
\alias{print.ari_matrix}
\title{Print method for class \code{ari_matrix}}
\usage{
\method{print}{ari_matrix}(x, ...)
}
\arguments{
\item{x}{A \code{ari_matrix} class object.}

\item{...}{Other arguments passed to \code{print} (not used in this function)}
}
\value{
Function prints to console but does not return any value.
}
\description{
Custom formatted print for weights matrices that outputs
information about feature weights functions to the console.
}
