% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.clust_fns_list}
\alias{print.clust_fns_list}
\title{Print method for class \code{clust_fns_list}}
\usage{
\method{print}{clust_fns_list}(x, ...)
}
\arguments{
\item{x}{A \code{clust_fns_list} class object.}

\item{...}{Other arguments passed to \code{print} (not used in this function)}
}
\value{
Function prints to console but does not return any value.
}
\description{
Custom formatted print for clustering functions list objects that outputs
information about the contained clustering functions to the console.
}
