% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.data_list}
\alias{print.data_list}
\title{Print method for class \code{data_list}}
\usage{
\method{print}{data_list}(x, ...)
}
\arguments{
\item{x}{A \code{data_list} class object.}

\item{...}{Other arguments passed to \code{print} (not used in this function)}
}
\value{
Function prints to console but does not return any value.
}
\description{
Custom formatted print for data list objects that outputs information about
the contained observations and components to the console.
}
