% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.distance_metrics_list}
\alias{print.distance_metrics_list}
\title{Print method for class \code{distance_metrics_list}}
\usage{
\method{print}{distance_metrics_list}(x, ...)
}
\arguments{
\item{x}{A \code{distance_metrics_list} class object.}

\item{...}{Other arguments passed to \code{print} (not used in this function)}
}
\value{
Function prints to console but does not return any value.
}
\description{
Custom formatted print for distance metrics list objects that outputs
information about the contained distance metrics to the console.
}
