% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.sim_mats_list}
\alias{print.sim_mats_list}
\title{Print method for class \code{sim_mats_list}}
\usage{
\method{print}{sim_mats_list}(x, ...)
}
\arguments{
\item{x}{A \code{sim_mats_list} class object.}

\item{...}{Other arguments passed to \code{print} (not used in this function).}
}
\description{
Custom formatted print for similarity matrix list
}
