% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.snf_results}
\alias{print.snf_results}
\title{Print method for class \code{snf_results}}
\usage{
\method{print}{snf_results}(x, ...)
}
\arguments{
\item{x}{A \code{snf_results} class object.}

\item{...}{Other arguments passed to \code{print} (not used in this function)}
}
\value{
Function prints to console but does not return any value.
}
\description{
Custom formatted print for SNF results
}
