% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.t_ext_solutions_df}
\alias{print.t_ext_solutions_df}
\title{Print method for class \code{t_ext_solutions_df}}
\usage{
\method{print}{t_ext_solutions_df}(x, ...)
}
\arguments{
\item{x}{A \code{t_solutions_df} class object.}

\item{...}{Other arguments passed to \code{print} (not used in this function)}
}
\value{
Function prints to console but does not return any value.
}
\description{
Custom formatted print for transposed solutions data frame class objects.
}
