% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf.R
\name{print_data_list_dims}
\alias{print_data_list_dims}
\title{Given a \code{data_list} object, print out the dimensions of each nested dataframe}
\usage{
print_data_list_dims(data_list)
}
\arguments{
\item{data_list}{The data_list object to be reduced}
}
\description{
Given a \code{data_list} object, print out the dimensions of each nested dataframe
}
