% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_with_t_message}
\alias{print_with_t_message}
\title{Helper function for transposing solutions_df message}
\usage{
print_with_t_message()
}
\value{
Output a message to use print with \code{n} to change displayed rows.
}
\description{
Helper function for transposing solutions_df message
}
\keyword{internal}
