% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{pval_heatmap}
\alias{pval_heatmap}
\title{Heatmap of p-values}
\usage{
pval_heatmap(
  pvals,
  order = NULL,
  cluster_columns = TRUE,
  cluster_rows = FALSE,
  show_column_names = FALSE,
  show_row_names = FALSE,
  colour_breaks = c(0, 0.5, 1),
  colours = c("red2", "lightyellow", "slateblue4")
)
}
\arguments{
\item{pvals}{A matrix of p-values.}

\item{order}{Numeric vector containing row order of the heatmap.}

\item{cluster_columns}{Whether columns should be sorted by hierarchical
clustering.}

\item{cluster_rows}{Whether rows should be sorted by hierarchical
clustering.}

\item{show_column_names}{Whether column names should be shown.}

\item{show_row_names}{Whether row names should be shown.}

\item{colour_breaks}{Numeric vector of breaks for the legend.}

\item{colours}{Vector of colours to use for the heatmap. Should match the
length of colour_breaks.}
}
\description{
Heatmap of p-values
}
