% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{pval_summaries}
\alias{pval_summaries}
\title{(DEPRECATED) Add minimum and mean p-values to an extended solutions matrix}
\usage{
pval_summaries(solutions_matrix, na_rm = TRUE)
}
\arguments{
\item{solutions_matrix}{A solutions_matrix object that already has some
p-value columns included.}

\item{na_rm}{If TRUE, NA values will be removed for mean/min calculations}
}
\description{
(DEPRECATED) Add minimum and mean p-values to an extended solutions matrix
}
