% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind.R
\name{rbind.ext_solutions_df}
\alias{rbind.ext_solutions_df}
\title{Row-binding of extended solutions data frame class objects.}
\usage{
\method{rbind}{ext_solutions_df}(...)
}
\arguments{
\item{...}{An arbitrary number of \code{ext_solutions_df} class objects.}
}
\value{
An \code{ext_solutions_df} class object.
}
\description{
Row-binding of extended solutions data frame class objects.
}
