% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind.R
\name{rbind.solutions_df}
\alias{rbind.solutions_df}
\title{Row-binding of solutions data frame class objects.}
\usage{
\method{rbind}{solutions_df}(...)
}
\arguments{
\item{...}{An arbitrary number of \code{solutions_df} class objects.}
}
\value{
A \code{solutions_df} class object.
}
\description{
Row-binding of solutions data frame class objects.
}
