% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind.R
\name{rbind.t_solutions_df}
\alias{rbind.t_solutions_df}
\title{Row-binding of t_solutions_df class objects}
\usage{
\method{rbind}{t_solutions_df}(...)
}
\arguments{
\item{...}{An arbitrary number of \code{t_solutions_df} class objects.}
}
\value{
A \code{t_solutions_df} class object.
}
\description{
Vertically stack two or more \code{t_solutions_df} class objects.
}
