% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind.R
\name{rbind.weights_matrix}
\alias{rbind.weights_matrix}
\title{Row-bind weights matrices}
\usage{
\method{rbind}{weights_matrix}(...)
}
\arguments{
\item{...}{An arbitrary number of \code{weights_matrix} class objects.}
}
\value{
A \code{weights_matrix} class object.
}
\description{
Vertically stack two or more \code{weights_matrix} class objects.
}
