% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{reduce_dl_to_common}
\alias{reduce_dl_to_common}
\title{Reduce data_list to common subjects}
\usage{
reduce_dl_to_common(dl)
}
\arguments{
\item{dl}{A nested list of input data from \code{generate_data_list()}.}
}
\value{
reduced_dl The data_list object subsetted only to subjectssnf
shared across all nested dataframes
}
\description{
Given a \code{data_list} object, reduce each nested dataframe to contain only the
set of subjects that are shared by all nested dataframes
}
