% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{reduce_dll_to_common}
\alias{reduce_dll_to_common}
\title{Reduce data list-like object to common observations}
\usage{
reduce_dll_to_common(dll)
}
\arguments{
\item{dll}{A data list-like \code{list} class object.}
}
\value{
reduced_dl The data list object subsetted only to uids shared across
all nested dataframes.
}
\description{
Given a data list-like list object, reduce each nested dataframe to contain
only the set of UIDs that are shared by all nested dataframes.
}
\keyword{internal}
