% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_dd}
\alias{remove_dd}
\title{Remove the data dictionary from a dataset}
\usage{
remove_dd(abcd_df)
}
\arguments{
\item{abcd_df}{An ABCD dataframe containing a data dictionary}
}
\value{
abcd_df_no_dd The same dataframe without the data dictionary
}
\description{
Remove the data dictionary from a dataset
}
