% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{remove_dl_na}
\alias{remove_dl_na}
\title{Remove NAs from a data_list object}
\usage{
remove_dl_na(data_list)
}
\arguments{
\item{data_list}{A data_list}
}
\value{
data_list A data_list without NAs
}
\description{
Remove NAs from a data_list object
}
