% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{remove_dll_incomplete}
\alias{remove_dll_incomplete}
\title{Remove observations with incomplete data from a data list-like list object}
\usage{
remove_dll_incomplete(dll)
}
\arguments{
\item{dll}{A data list-like \code{list} class object.}
}
\value{
The provided dll with missing observations removed.
}
\description{
Helper function during \code{data_list} class initialization. First applies
\code{stats::na.omit()} to the data frames named "data" within a nested list.
Then removes any observations that are not present across all data frames.
}
\keyword{internal}
