% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{remove_dll_na}
\alias{remove_dll_na}
\title{Remove NA values from a data list-like list object}
\usage{
remove_dll_na(dll)
}
\arguments{
\item{dll}{A data list-like \code{list} class object. Should contain at least one
nested list with a data frame named "data".}
}
\value{
dll The provided data list-like object with missing observations
removed.
}
\description{
Helper function during \code{data_list} class initialization. Applies
'\code{stats::na.omit()} to the data frames named "data" within a nested list.
}
\keyword{internal}
