% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_otbi01.R
\name{rename_otbi01}
\alias{rename_otbi01}
\title{Rename ambiguous columns in otbi01 file.}
\usage{
rename_otbi01(otbi01)
}
\arguments{
\item{otbi01}{The complete file abcd_otbi01.txt}
}
\value{
renamed_otbi01 A modified form of otbi01 with clearer column names
}
\description{
Renames column names of ABCD's Ohio TBI Screen data to meaningful names that are easier to work with.
Use 'original_otbi_names()' to print out a conversion table of old and new column names.
}
\examples{
# Mock abcd_otbi01.txt
abcd_otbi01 <- data.frame(matrix(NA, nrow = 2, ncol = 49))

colnames(abcd_otbi01) <- c("collection_id", "abcd_otbi01_id", "dataset_id", 
    "subjectkey", "src_subject_id", "interview_date", "interview_age", "sex", 
    "eventname", "tbi_select_language___1", "tbi_1", "tbi_1b", "tbi_1c",
    "tbi_1d", "tbi_2", "tbi_2b", "tbi_2c", "tbi_2d", "tbi_3", "tbi_3b",
    "tbi_3c", "tbi_3d", "tbi_4", "tbi_4b", "tbi_4c", "tbi_4d", "tbi_5",
    "tbi_5b", "tbi_5c", "tbi_5d", "tbi_6o", "tbi_6p", "tbi_6q", "tbi_6r",
    "tbi_6s", "tbi_7a", "tbi_7c1", "tbl_7c2", "tbi_7e", "tbi_7f", "tbi_7g",
    "tbi_7i", "tbi_7k", "tbi_7l", "tbi_8g", "tbi_8i", "tbi_8k", "tbi_8l",
    "collection_title")

otbi01_renamed <- rename_otbi01(abcd_otbi01)
}
