% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{reorder_dl_subs}
\alias{reorder_dl_subs}
\title{Reorder the subjects in a data_list}
\usage{
reorder_dl_subs(data_list, ordered_subjects)
}
\arguments{
\item{data_list}{A nested list of input data from \code{generate_data_list()}.}

\item{ordered_subjects}{A vector of the subjectkey values in the data_list
in the desired order of the sorted data_list.}
}
\value{
A data list ("list"-class object) with reordered observations.
}
\description{
Reorder the subjects in a data_list
}
