% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{reorder_dl_uids}
\alias{reorder_dl_uids}
\title{Reorder the uids in a data list}
\usage{
reorder_dl_uids(dl, ordered_uids)
}
\arguments{
\item{dl}{A nested list of input data from \code{data_list()}.}

\item{ordered_uids}{A vector of the uid values in the data list
in the desired order of the sorted data list.}
}
\value{
A data list ("list"-class object) with reordered observations.
}
\description{
Reorder the uids in a data list
}
\keyword{internal}
