% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resample}
\alias{resample}
\title{Helper resample function found in ?sample}
\usage{
resample(x, ...)
}
\arguments{
\item{x}{Vector or single value to sample from}

\item{...}{Remaining arguments for base::sample function}
}
\value{
Numeric vector result of running base::sample.
}
\description{
Like sample, but when given a single value x, returns back that single
value instead of a random value from 1 to x.
}
