% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_snf.R
\name{run_snf}
\alias{run_snf}
\title{Run SNF}
\usage{
run_snf(i, dl, sc, return_sim_mats, sim_mats_dir, p)
}
\arguments{
\item{i}{Row of settings_df and weights_matrix within SNF config to use.}

\item{dl}{A nested list of input data from \code{data_list()}.}

\item{sc}{An \code{snf_config} class object which stores all sets of
hyperparameters used to transform data in dl into a cluster solutions. See
\code{?settings_df} or
https://branchlab.github.io/metasnf/articles/settings_df.html for more
details.}

\item{return_sim_mats}{If TRUE, function will return a list where
the first element is the solutions data frame and the second element is a list
of similarity matrices for each row in the sol_df. Default FALSE.}

\item{sim_mats_dir}{If specified, this directory will be used to
save all generated similarity matrices.}
}
\value{
A list containing a cluster solution (numeric vector) and a
the fused network used to create that cluster solution. The fused network
is NULL if return_sim_mats is FALSE.
}
\description{
Helper function for running a single SNF config pipeline.
}
\keyword{internal}
