% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{save_heatmap}
\alias{save_heatmap}
\title{Save a heatmap object to a file}
\usage{
save_heatmap(heatmap, path, width = 480, height = 480, res = 100)
}
\arguments{
\item{heatmap}{The heatmap object to save.}

\item{path}{The path to save the heatmap to.}

\item{width}{The width of the heatmap.}

\item{height}{The height of the heatmap.}

\item{res}{The resolution of the heatmap.}
}
\value{
Does not return any value. Saves heatmap to file.
}
\description{
Save a heatmap object to a file
}
