% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{save_plot}
\alias{save_plot}
\title{Save a plot to a png file}
\usage{
save_plot(plot, path, width, height, res)
}
\arguments{
\item{plot}{The heatmap object to save.}

\item{path}{The path to save the heatmap to.}

\item{width}{The width of the heatmap.}

\item{height}{The height of the heatmap.}

\item{res}{The resolution of the heatmap.}
}
\description{
Save a plot to a png file
}
