% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{scabcd_ggsave}
\alias{scabcd_ggsave}
\title{Wrapper for ggsave}
\usage{
scabcd_ggsave(path, plotname)
}
\arguments{
\item{path}{the external path location ending in trailing slash}

\item{plotname}{the name of the gg figure to be saved ending in .png}
}
\description{
Saves figures to one internal and one external location.
The internal path is set as "figures/abcd/"
}
