% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{scabcd_hist}
\alias{scabcd_hist}
\title{Wrapper for geom_hist}
\usage{
scabcd_hist(df, xval, xlabel = "", title = "", size = 20)
}
\arguments{
\item{df}{A dataframe}

\item{xval}{x-value for histogram passed as a string}

\item{xlabel}{x-axis label}

\item{title}{plot title}

\item{size}{x-axis label}
}
\value{
renamed_tbi A modified form of tbi_df with clearer column names
}
\description{
Wrapper for geom_hist
}
