% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{sdl}
\alias{sdl}
\title{Summarize data list}
\usage{
sdl(data_list)
}
\arguments{
\item{data_list}{nested list of input data generated by the function
\code{get_data_list()}}
}
\value{
dl_summary Summarized output
}
\description{
Summarize data list
}
