% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Select variables from an object}
\usage{
select(.data, ...)
}
\arguments{
\item{.data}{The object to select columns from.}

\item{...}{Additional arguments for selection.}
}
\value{
Object with selected columns.
}
\description{
Select variables from an object
}
