% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select.settings_df}
\alias{select.settings_df}
\title{Select method for class \code{settings_df}}
\usage{
select.settings_df(.data, ...)
}
\arguments{
\item{.data}{A \code{settings_df} class object.}

\item{...}{Other arguments passed to \code{select} (not used in this function)}
}
\value{
If the result of the selection is a valid settings data frame,
returns that result. Otherwise, returns that result as a non-settings_df.
}
\description{
Reduce a settings data frame to a subset of its columns.
}
