% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_characterization.R
\name{select_om}
\alias{select_om}
\title{Select specific row_ids from an output matrix}
\usage{
select_om(om, row_ids)
}
\arguments{
\item{om}{output matrix}

\item{row_ids}{vector of row_id values to be selected}
}
\value{
selected_om
}
\description{
Select specific row_ids from an output matrix
}
