% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mats_list.R
\name{sim_mats_list}
\alias{sim_mats_list}
\title{Create or extract a \code{sim_mats_list} class object}
\usage{
sim_mats_list(x)
}
\arguments{
\item{x}{The object to create or extract a \code{sim_mats_list} from.}
}
\value{
A \code{sim_mats_list} class object.
}
\description{
Create or extract a \code{sim_mats_list} class object
}
