% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_measures.R
\name{sim_to_dist}
\alias{sim_to_dist}
\title{Convert a similarity matrix to a dissimilarity matrix}
\usage{
sim_to_dist(sim_mat)
}
\arguments{
\item{sim_mat}{A similarity matrix}
}
\value{
A dissimilarity matrix
}
\description{
Convert a similarity matrix to a dissimilarity matrix
}
\keyword{internal}
