% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_plotting.R
\name{similarity_matrix_heatmap}
\alias{similarity_matrix_heatmap}
\title{Plot heatmap of similarity matrix}
\usage{
similarity_matrix_heatmap(
  similarity_matrix,
  order = NULL,
  cluster_solution = NULL,
  scale_diag = "mean",
  log_graph = TRUE,
  cluster_rows = FALSE,
  cluster_columns = FALSE,
  show_row_names = FALSE,
  show_column_names = FALSE,
  data_list = NULL,
  data = NULL,
  left_bar = NULL,
  right_bar = NULL,
  top_bar = NULL,
  bottom_bar = NULL,
  left_hm = NULL,
  right_hm = NULL,
  top_hm = NULL,
  bottom_hm = NULL,
  ...
)
}
\arguments{
\item{similarity_matrix}{A similarity matrix}

\item{cluster_solution}{Vector containing cluster assignments}

\item{scale_diag}{Method of rescaling matrix diagonals. Can be "none"
(don't change diagonals), "mean" (replace diagonals with average value of
off-diagonals), or "zero" (replace diagonals with 0).}

\item{log_graph}{If TRUE, log transforms the graph.}

\item{cluster_rows}{Parameter for ComplexHeatmap::Heatmap}

\item{cluster_columns}{Parameter for ComplexHeatmap::Heatmap}

\item{show_row_names}{Parameter for ComplexHeatmap::Heatmap}

\item{show_column_names}{Parameter for ComplexHeatmap::Heatmap}

\item{...}{Additional parameters passed into ComplexHeatmap::Heatmap}
}
\description{
Plot heatmap of similarity matrix
}
