% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{similarity_matrix_path}
\alias{similarity_matrix_path}
\title{Generate a complete path and filename to store an similarity matrix}
\usage{
similarity_matrix_path(similarity_matrix_dir, i)
}
\arguments{
\item{similarity_matrix_dir}{Directory to store similarity matrices}

\item{i}{Corresponding settings matrix row}
}
\value{
path Complete path and filename to store an similarity matrix
}
\description{
Generate a complete path and filename to store an similarity matrix
}
