% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_metrics.R
\name{sn_euclidean_distance}
\alias{sn_euclidean_distance}
\title{Distance metric: Standard normalization then Euclidean}
\usage{
sn_euclidean_distance(df)
}
\arguments{
\item{df}{Dataframe containing at least 1 data column.}
}
\value{
distance_matrix A distance matrix.
}
\description{
Distance metric: Standard normalization then Euclidean
}
