% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf_schemes.R
\name{snf_step}
\alias{snf_step}
\title{SNF a data_list}
\usage{
snf_step(
  data_list,
  scheme,
  k = 20,
  alpha = 0.5,
  t = 20,
  cont_dist_fn,
  disc_dist_fn,
  ord_dist_fn,
  cat_dist_fn,
  mix_dist_fn
)
}
\arguments{
\item{data_list}{nested list of input data generated by the function
\code{get_data_list()}}

\item{scheme}{Which SNF system to use to achieve the final fused network}

\item{k}{k hyperparameter}

\item{alpha}{alpha/eta/sigma hyperparameter}

\item{t}{SNF number of iterations hyperparameter}

\item{cont_dist_fn}{distance metric function for continuous data}

\item{disc_dist_fn}{distance metric function for discrete data}

\item{ord_dist_fn}{distance metric function for ordinal data}

\item{cat_dist_fn}{distance metric function for categorical data}

\item{mix_dist_fn}{distance metric function for mixed data}
}
\value{
fused_network The final fused network for clustering
}
\description{
SNF a data_list
}
