% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_list.R
\name{sol}
\alias{sol}
\title{Summarize an outcome list}
\usage{
sol(outcome_list)
}
\arguments{
\item{outcome_list}{an outcome_list}
}
\value{
ol_summary Summarized output
}
\description{
Summarize an outcome list
}
