% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions_df.R
\name{sol_df_col_order}
\alias{sol_df_col_order}
\title{Helper function for organizing solutions df-like column order}
\usage{
sol_df_col_order(x)
}
\arguments{
\item{x}{Object with columns "solution", "nclust", and "mc".}
}
\value{
x with column names reordered.
}
\description{
Reorders columns of a solutions data frame to "solution", "nclust", "mc",
then all other column names.
}
\keyword{internal}
