% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions_df.R
\name{solutions_df}
\alias{solutions_df}
\title{Constructor for \code{solutions_df} class object}
\usage{
solutions_df(sol_dfl, smll, sc, dl)
}
\arguments{
\item{run_snf_results}{}
}
\value{
RETURN
}
\description{
Constructor for \code{solutions_df} class object
}
