% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_clustering.R
\name{spectral_clustering}
\alias{spectral_clustering}
\title{Adaptation of SNFtool's spectral clustering function}
\usage{
spectral_clustering(affinity, K, type = 3)
}
\arguments{
\item{affinity}{A similarity matrix}

\item{K}{hyperparameter}

\item{type}{type of spectral clustering}
}
\value{
RETURN
}
\description{
Adaptation of SNFtool's spectral clustering function
}
