% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_algorithms.R
\name{spectral_eigen}
\alias{spectral_eigen}
\title{Clustering algorithm: Spectral clustering with eigen-gap heuristic}
\usage{
spectral_eigen(similarity_matrix)
}
\arguments{
\item{similarity_matrix}{A similarity matrix}
}
\value{
solution_data A list storing cluster assignments and the number of
clusters.
}
\description{
Applies spectral clustering to similarity matrix. Number of clusters is based
on the eigen-gap heuristic.
}
