% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_algorithms.R
\name{spectral_eigen_classic}
\alias{spectral_eigen_classic}
\title{Clustering algorithm: Spectral clustering with eigen-gap heuristic}
\usage{
spectral_eigen_classic(similarity_matrix)
}
\arguments{
\item{similarity_matrix}{A similarity matrix}
}
\value{
solution A vector indicating which cluster each patient was assigned
to
}
\description{
Applies spectral clustering to similarity matrix. Number of clusters is based
on the eigen-gap heuristic.
}
