% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_algorithms.R
\name{spectral_eight}
\alias{spectral_eight}
\title{Clustering algorithm: Spectral clustering for a eight cluster solution}
\usage{
spectral_eight(similarity_matrix)
}
\arguments{
\item{similarity_matrix}{An similarity matrix}
}
\value{
solution A vector indicating which cluster each patient was assigned
to
}
\description{
Applies spectral clustering to similarity matrix. Seeks eight clusters.
}
