% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_algorithms.R
\name{spectral_nine}
\alias{spectral_nine}
\title{Clustering algorithm: Spectral clustering for a nine cluster solution}
\usage{
spectral_nine(similarity_matrix)
}
\arguments{
\item{similarity_matrix}{A similarity matrix}
}
\value{
solution A vector indicating which cluster each patient was assigned
to
}
\description{
Applies spectral clustering to similarity matrix. Seeks nine clusters.
}
