% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_algorithms.R
\name{spectral_rot}
\alias{spectral_rot}
\title{Clustering algorithm: Spectral clustering with rotation cost heuristic}
\usage{
spectral_rot(similarity_matrix)
}
\arguments{
\item{similarity_matrix}{A similarity matrix}
}
\value{
solution_data A list storing cluster assignments and the number of
clusters.
}
\description{
Applies spectral clustering to similarity matrix. Number of clusters is based
on the rotation cost heuristic.
}
