% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_algorithms.R
\name{spectral_seven}
\alias{spectral_seven}
\title{Clustering algorithm: Spectral clustering for a seven cluster solution}
\usage{
spectral_seven(affinity_matrix)
}
\arguments{
\item{affinity_matrix}{An affinity matrix}
}
\value{
solution A vector indicating which cluster each patient was assigned
to
}
\description{
Applies spectral clustering to affinity matrix. Seeks seven clusters.
}
