% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_algorithms.R
\name{spectral_six}
\alias{spectral_six}
\title{Clustering algorithm: Spectral clustering for a six cluster solution}
\usage{
spectral_six(affinity_matrix)
}
\arguments{
\item{affinity_matrix}{An affinity matrix}
}
\value{
solution A vector indicating which cluster each patient was assigned
to
}
\description{
Applies spectral clustering to affinity matrix. Seeks six clusters.
}
