% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_algorithms.R
\name{spectral_ten}
\alias{spectral_ten}
\title{Clustering algorithm: Spectral clustering for a ten cluster solution}
\usage{
spectral_ten(similarity_matrix)
}
\arguments{
\item{similarity_matrix}{A similarity matrix}
}
\value{
solution A vector indicating which cluster each patient was assigned
to
}
\description{
Applies spectral clustering to similarity matrix. Seeks ten clusters.
}
