% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{split_parser}
\alias{split_parser}
\title{Helper function to determine which row and columns to split on}
\usage{
split_parser(
  row_split_vector = NULL,
  column_split_vector = NULL,
  row_split = NULL,
  column_split = NULL,
  n_rows,
  n_columns
)
}
\arguments{
\item{row_split_vector}{A vector of row indices to split on.}

\item{column_split_vector}{A vector of column indices to split on.}

\item{row_split}{Standard parameter of \code{ComplexHeatmap::Heatmap}.}

\item{column_split}{Standard parameter of \code{ComplexHeatmap::Heatmap}.}

\item{n_rows}{The number of rows in the data.}

\item{n_columns}{The number of columns in the data.}
}
\value{
"list"-class object containing row_split and column_split character
vectors to pass into ComplexHeatmap::Heatmap.
}
\description{
Helper function to determine which row and columns to split on
}
