% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_metrics.R
\name{sq_euclidean_distance}
\alias{sq_euclidean_distance}
\title{Squared Euclidean distance}
\usage{
sq_euclidean_distance(df, weights_row)
}
\arguments{
\item{df}{Dataframe containing at least 1 data column.}

\item{weights_row}{Single-row dataframe where the column names contain the
column names in df and the row contains the corresponding weights.}
}
\value{
distance_matrix A distance matrix.
}
\description{
Squared Euclidean distance
}
