% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str.ari_matrix}
\alias{str.ari_matrix}
\title{Structure of a \code{ari_matrix} object}
\usage{
\method{str}{ari_matrix}(object, ...)
}
\arguments{
\item{object}{A \code{ari_matrix} class object.}

\item{...}{Additional arguments (not used).}
}
\value{
Does not return an object; outputs object structure to console.
}
\description{
Structure of a \code{ari_matrix} object
}
