% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str.clust_fns_list}
\alias{str.clust_fns_list}
\title{Structure of a \code{clust_fns_list} object}
\usage{
\method{str}{clust_fns_list}(object, ...)
}
\arguments{
\item{object}{A \code{clust_fns_list} class object.}

\item{...}{Additional arguments (not used).}
}
\value{
Does not return an object; outputs object structure to console.
}
\description{
Structure of a \code{clust_fns_list} object
}
