% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str.data_list}
\alias{str.data_list}
\title{Structure of a \code{data_list} object}
\usage{
\method{str}{data_list}(object, ...)
}
\arguments{
\item{object}{A \code{data_list} class object.}

\item{...}{Additional arguments (not used).}
}
\value{
Does not return an object; outputs object structure to console.
}
\description{
Structure of a \code{data_list} object
}
