% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str.ext_solutions_df}
\alias{str.ext_solutions_df}
\title{Structure of a \code{ext_solutions_df} object}
\usage{
\method{str}{ext_solutions_df}(object, ...)
}
\arguments{
\item{object}{A \code{ext_solutions_df} class object.}

\item{...}{Additional arguments (not used).}
}
\value{
Does not return an object; outputs object structure to console.
}
\description{
Structure of a \code{ext_solutions_df} object
}
