% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str.snf_config}
\alias{str.snf_config}
\title{Structure of a \code{snf_config} object}
\usage{
\method{str}{snf_config}(object, ...)
}
\arguments{
\item{object}{A \code{snf_config} class object.}

\item{...}{Additional arguments (not used).}
}
\value{
Does not return an object; outputs object structure to console.
}
\description{
Structure of a \code{snf_config} object
}
