% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str.weights_matrix}
\alias{str.weights_matrix}
\title{Structure of a \code{weights_matrix} object}
\usage{
\method{str}{weights_matrix}(object, ...)
}
\arguments{
\item{object}{A \code{weights_matrix} class object.}

\item{...}{Additional arguments (not used).}
}
\value{
Does not return an object; outputs object structure to console.
}
\description{
Structure of a \code{weights_matrix} object
}
