% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitive_extraction.R
\name{[.data_list}
\alias{[.data_list}
\title{Extraction operator for data lists}
\usage{
\method{[}{data_list}(x, i, ...)
}
\arguments{
\item{x}{A \code{data_list} class object.}

\item{i}{Indices for component extraction.}

\item{...}{Additional parameters (invalid for data list extraction).}
}
\value{
\code{data_list} class object of extracted components.
}
\description{
Enables usage of \code{[} extraction operator on \code{data_list} class objects. Given
a \code{numeric} or \code{character} vector of indices, returns a subsetted data list.
Data lists only use single-dimension extraction (component-wise).
}
