% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{[.settings_df}
\alias{[.settings_df}
\title{Extraction operator for settings data frames}
\usage{
\method{[}{settings_df}(x, i, j, ...)
}
\arguments{
\item{x}{A \code{settings_df} class object.}

\item{i}{Indices for component extraction.}

\item{j}{Indices for component extraction.}

\item{...}{Additional parameters (invalid for settings df extraction).}
}
\value{
\code{settings_df} class object of extracted components.
}
\description{
Enables usage of \code{[} extraction operator on \code{settings_df} class objects.
}
