% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{[.snf_config}
\alias{[.snf_config}
\title{Extraction operator for SNF config}
\usage{
\method{[}{snf_config}(x, i, ...)
}
\arguments{
\item{x}{A \code{snf_config} class object.}

\item{i}{Indices for component extraction.}

\item{...}{Additional parameters (invalid for snf_config extraction).}
}
\value{
\code{snf_config} class object of extracted components.
}
\description{
Enables usage of \code{[} extraction operator on \code{snf_config} class objects.
}
