% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{[.solutions_df}
\alias{[.solutions_df}
\title{Extraction operator for solutions data frame}
\usage{
\method{[}{solutions_df}(x, i, j, ...)
}
\arguments{
\item{x}{A \code{solutions_df} class object.}

\item{i}{Indices for component extraction.}

\item{j}{Indices for component extraction.}

\item{...}{Additional parameters (invalid for solutions_df extraction).}
}
\value{
\code{solutions_df} class object of extracted components.
}
\description{
Enables usage of \code{[} extraction operator on \code{solutions_df} class objects.
}
