% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{[.weights_matrix}
\alias{[.weights_matrix}
\title{Extraction operator for weights matrix}
\usage{
\method{[}{weights_matrix}(x, i, j, ...)
}
\arguments{
\item{x}{A \code{weights_matrix} class object.}

\item{i}{Indices for component extraction.}

\item{j}{Indices for component extraction.}

\item{...}{Additional parameters (invalid for weights_matrix extraction).}
}
\value{
\code{weights_matrix} class object of extracted components.
}
\description{
Enables usage of \code{[} extraction operator on \code{weights_matrix} class objects.
}
