% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf.R
\name{subdomains}
\alias{subdomains}
\title{Return subdomains of a data list}
\usage{
subdomains(data_list)
}
\arguments{
\item{data_list}{nested list of input data generated by the function
\code{get_data_list()}}
}
\value{
subdomain_list list of subdomains
}
\description{
Return subdomains of a data list
}
