% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_characterization.R
\name{subs}
\alias{subs}
\title{Select only output_matrix columns containing subjects}
\usage{
subs(df)
}
\arguments{
\item{df}{Datframe}
}
\value{
df_subs Dataframe without subjects
}
\description{
Return a dataframe with only subject columns
}
