% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_measures.R
\name{subsample_data_list}
\alias{subsample_data_list}
\title{Create subsamples of a data_list}
\usage{
subsample_data_list(
  data_list,
  n_subsamples,
  subsample_fraction = NULL,
  n_subjects = NULL
)
}
\arguments{
\item{data_list}{A data_list.}

\item{n_subsamples}{Number of subsamples to create.}

\item{subsample_fraction}{Percentage of patients to include per subsample.}

\item{n_subjects}{Number of patients to include per subsample.}
}
\value{
data_list_subsamples A list of subsampled data_list objects.
}
\description{
Create subsamples of a data_list
}
