% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coclustering.R
\name{subsample_dl}
\alias{subsample_dl}
\title{Create subsamples of a data list}
\usage{
subsample_dl(dl, n_subsamples, subsample_fraction = NULL, n_subjects = NULL)
}
\arguments{
\item{dl}{A nested list of input data from \code{data_list()}.}

\item{n_subsamples}{Number of subsamples to create.}

\item{subsample_fraction}{Percentage of patients to include per subsample.}

\item{n_subjects}{Number of patients to include per subsample.}
}
\value{
A "list" class object containing \code{n_subsamples} number of
data lists. Each of those data lists contains a random \code{subsample_fraction}
fraction of the observations of the provided data list.
}
\description{
Given a data list, return a list of smaller data lists that are generated
through random sampling (without replacement).
}
\examples{
# my_dl <- data_list(
#     list(subc_v, "subcortical_volume", "neuroimaging", "continuous"),
#     list(income, "household_income", "demographics", "continuous"),
#     list(pubertal, "pubertal_status", "demographics", "continuous"),
#     uid = "unique_id"
# )
# 
# sc <- snf_config(my_dl, n_solutions = 5, max_k = 40)
# 
# sol_df <- batch_snf(my_dl, sc)
# 
# my_dl_subsamples <- subsample_dl(
#     my_dl,
#     n_subsamples = 20,
#     subsample_fraction = 0.85
# )
# 
# batch_subsample_results <- batch_snf_subsamples(
#     my_dl_subsamples,
#     sc,
#     verbose = TRUE
# )
# 
# subsample_cluster_solutions <- batch_subsample_results[["cluster_solutions"]]
# 
# pairwise_aris <- subsample_pairwise_aris(
#     subsample_cluster_solutions,
#     return_raw_aris = TRUE,
#     verbose = TRUE
# )
# 
# coclustering_results <- calculate_coclustering(
#     subsample_cluster_solutions,
#     sol_df,
#     verbose = TRUE
# )
# 
# coclustering_results$"cocluster_summary"
# 
# cocluster_dfs <- coclustering_results$"cocluster_dfs"
# 
# cocluster_density(cocluster_dfs[[1]])
# 
# cocluster_heatmap(
#     cocluster_dfs[[1]],
#     dl = my_dl,
#     top_hm = list(
#         "Income" = "household_income",
#         "Pubertal Status" = "pubertal_status"
#     ),
#     annotation_colours = list(
#         "Pubertal Status" = colour_scale(
#             c(1, 4),
#             min_colour = "black",
#             max_colour = "purple"
#         ),
#         "Income" = colour_scale(
#             c(0, 4),
#             min_colour = "black",
#             max_colour = "red"
#         )
#     )
# )
}
