% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitive_assignment.R
\name{[<-.data_list}
\alias{[<-.data_list}
\title{Assignment operator for data lists}
\usage{
\method{[}{data_list}(x, i) <- value
}
\arguments{
\item{x}{A \code{data_list} class object.}

\item{i}{Index for component assignment.}

\item{value}{A list structured as a valid component, i.e. a named list of
items "data", "name", "domain", and "type" respectively of classes
'\code{data.frame}, \code{character}, \code{charater}, \code{character}. UIDs in value must
match those in the data list.}
}
\value{
A data list with the assigned component.
}
\description{
Enables usage of \verb{[<-} assignment operator on \code{data_list} class objects.
Given a \code{numeric} or \code{character} index, returns a subsetted data list.
Data lists only use single-dimension extraction (component-wise).
}
