% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{[<-.settings_df}
\alias{[<-.settings_df}
\title{Assignment operator for settings data frames}
\usage{
\method{[}{settings_df}(x, i, j) <- value
}
\arguments{
\item{x}{A \code{settings_df} class object.}

\item{i}{Index for component assignment.}

\item{j}{Index for component assignment.}

\item{value}{Value to pass into x.}
}
\value{
A settings data frame with the assigned value(s).
}
\description{
Enables usage of \verb{[<-} assignment operator on \code{settings_df} class objects.
Given a \code{numeric} or \code{character} index, returns an new settings data frame.
}
