% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{[<-.weights_matrix}
\alias{[<-.weights_matrix}
\title{Assignment operator for weights matrix}
\usage{
\method{[}{weights_matrix}(x, i, j) <- value
}
\arguments{
\item{x}{A \code{weights_matrix} class object.}

\item{i}{Index for component assignment.}

\item{j}{Index for component assignment.}

\item{value}{Value to pass into x.}
}
\value{
A weights matrix with the assigned value(s).
}
\description{
Enables usage of \verb{[<-} assignment operator on \code{weights_matrix} objects.
Given a \code{numeric} or \code{character} index, returns an new settings data frame.
}
