% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_algorithms.R
\name{summarize_clust_algs_list}
\alias{summarize_clust_algs_list}
\title{Summarize a clust_algs_list object}
\usage{
summarize_clust_algs_list(clust_algs_list)
}
\arguments{
\item{clust_algs_list}{A clust_algs_list object}
}
\value{
summary_df Dataframe containing the name of each algorithm and the
number of each algorithm (the value in the settings_matrix that will be
used to access that algorithm)
}
\description{
Summarize a clust_algs_list object
}
