% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{summarize_clust_fns_list}
\alias{summarize_clust_fns_list}
\title{Summarize a clust_fns_list object}
\usage{
summarize_clust_fns_list(cfl)
}
\arguments{
\item{cfl}{A \code{clust_fns_list} class object.}
}
\value{
summary_df "data.frame" class object containing the name and index
of each clustering algorithm in the provided \code{clust_fns_list}.
}
\description{
Summarize a clust_fns_list object
}
\keyword{internal}
