% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_fns_list.R
\name{summarize_dfl}
\alias{summarize_dfl}
\title{Summarize metrics contained in a dist_fns_list}
\usage{
summarize_dfl(dist_fns_list)
}
\arguments{
\item{dist_fns_list}{A dist_fns_list.}
}
\value{
"data.frame"-class object summarizing items in a distance metrics
list.
}
\description{
Summarize metrics contained in a dist_fns_list
}
