% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{summarize_distance_metrics_list}
\alias{summarize_distance_metrics_list}
\title{Summarize metrics contained in a distance_metrics_list}
\usage{
summarize_distance_metrics_list(distance_metrics_list)
}
\arguments{
\item{distance_metrics_list}{A distance_metrics_list.}
}
\description{
Summarize metrics contained in a distance_metrics_list
}
