% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{summarize_dl}
\alias{summarize_dl}
\title{Summarize data list}
\usage{
summarize_dl(data_list)
}
\arguments{
\item{data_list}{A nested list of input data from \code{generate_data_list()}.}
}
\value{
dl_summary Summarized output
}
\description{
Summarize data list
}
