% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_list.R
\name{summarize_ol}
\alias{summarize_ol}
\title{Summarize outcome list}
\usage{
summarize_ol(outcome_list)
}
\arguments{
\item{outcome_list}{nested list of input data generated by the function
\code{get_outcome_list()}}
}
\value{
ol_summary Summarized output
}
\description{
Summarize outcome list
}
