% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_list.R
\name{summarize_target_list}
\alias{summarize_target_list}
\title{Summarize target list}
\usage{
summarize_target_list(target_list)
}
\arguments{
\item{target_list}{nested list of input data generated by the function
\code{get_target_list()}}
}
\value{
ol_summary Summarized output
}
\description{
Summarize target list
}
