% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.ari_matrix}
\alias{summary.ari_matrix}
\title{Summary method for class \code{ari_matrix}}
\usage{
\method{summary}{ari_matrix}(object, ...)
}
\arguments{
\item{object}{A \code{ari_matrix} class object.}

\item{...}{Other arguments passed to \code{summary} (not used in this function).}
}
\value{
A named list containing the number of solutions and the distribution
of ARI values.
}
\description{
Provides a summary of the \code{ari_matrix} class object, including the
distribution of the adjusted Rand index (ARI) values and the number of
solutions.
}
