% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.clust_fns_list}
\alias{summary.clust_fns_list}
\title{Summary method for class \code{clust_fns_list}}
\usage{
\method{summary}{clust_fns_list}(object, ...)
}
\arguments{
\item{object}{A \code{clust_fns_list} class object.}

\item{...}{Other arguments passed to \code{summary} (not used in this function).}
}
\value{
Returns no value. Outputs a message to the console.
}
\description{
This summary function simply returns to the console the number of functions
contained in the \code{clust_fns_list} object.
}
