% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.data_list}
\alias{summary.data_list}
\title{Summary method for class \code{data_list}}
\usage{
\method{summary}{data_list}(object, scope = "component", ...)
}
\arguments{
\item{object}{A \code{data_list} class object.}

\item{scope}{The level of detail for the summary. By default, this is set to
"component", which returns a summary of the data list at the component
level. Can also be set to "feature", resulting in a summary at the feature
level.}

\item{...}{Other arguments passed to \code{summary} (not used in this function)}
}
\value{
A \code{data.frame} class object. If \code{scope} is "component", each row
shows the name, variable type, domain, and dimensions of each component. If
\code{scope} is "feature", each row shows the name, variable type, and domain of
each feature.
}
\description{
Returns a data list summary (\code{data.frame} class object) containing
information on components, features, variable types, domains, and component
dimensions.
}
