% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.settings_df}
\alias{summary.settings_df}
\title{Summary method for class \code{settings_df}}
\usage{
\method{summary}{settings_df}(object, ...)
}
\arguments{
\item{object}{A \code{settings_df} class object.}

\item{...}{Other arguments passed to \code{summary} (not used in this function).}
}
\value{
A named list containing summary information of the settings data
frame.
}
\description{
This summary function provides a summary of the \code{settings_df} class
object, including the number of settings, the distribution of alpha values,
the distribution of k values, and the distribution of clustering functions.
}
