% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.snf_config}
\alias{summary.snf_config}
\title{Summary method for class \code{snf_config}}
\usage{
\method{summary}{snf_config}(object, ...)
}
\arguments{
\item{object}{A \code{snf_config} class object.}

\item{...}{Other arguments passed to \code{summary} (not used in this function).}
}
\value{
A named list containing the summaries of objects within the config.
}
\description{
This summary function provides a summary of the \code{snf_config} class object,
including the settings data frame, clustering functions list, distance
functions list, and weights matrix.
}
