% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.solutions_df}
\alias{summary.solutions_df}
\title{Summary method for class \code{solutions_df}}
\usage{
\method{summary}{solutions_df}(object, ...)
}
\arguments{
\item{object}{A \code{ext_solutions_df} class object.}

\item{...}{Other arguments passed to \code{summary} (not used in this function).}
}
\value{
A named list containing the number of solutions, the distribution of
the number of clusters, and the number of observations.
}
\description{
This summary function provides a summary of the \code{solutions_df} class
object, including the number of solutions, the distribution of the number of
clusters, and the number of observations.
}
