% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.t_solutions_df}
\alias{summary.t_solutions_df}
\title{Summary method for class \code{t_solutions_df}}
\usage{
\method{summary}{t_solutions_df}(object, ...)
}
\arguments{
\item{object}{A \code{t_solutions_df} class object.}

\item{...}{Other arguments passed to \code{summary} (not used in this function).}
}
\value{
A named list containing the number of solutions, the distribution of
the number of clusters, the number of features, the number of observations,
and the distribution of p-values.
}
\description{
This summary function provides a summary of the \code{t_solutions_df} class
object, including the number of solutions, the distribution of the number of
clusters, the number of features, the number of observations, and the
distribution of p-values.
}
