% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.weights_matrix}
\alias{summary.weights_matrix}
\title{Summary method for class \code{weights_matrix}}
\usage{
\method{summary}{weights_matrix}(object, ...)
}
\arguments{
\item{object}{A \code{weights_matrix} class object.}

\item{...}{Other arguments passed to \code{summary} (not used in this function).}
}
\value{
A named list containing the summary statistics of the weights matrix,
the number of solutions, and the number of features.
}
\description{
This summary function provides a summary of the \code{weights_matrix} class
object, including the minimum, maximum, mean, and standard deviation of the
feature weights.
}
