% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_features.R
\name{summary_features}
\alias{summary_features}
\title{Pull features used to calculate summary p-values from an object}
\usage{
summary_features(x)
}
\arguments{
\item{x}{The object to extract summary features from.}
}
\value{
A character vector of summary features.
}
\description{
Pull features used to calculate summary p-values from an object
}
\keyword{internal}
