% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.R
\name{t.ext_solutions_df}
\alias{t.ext_solutions_df}
\title{Transpose a solutions data frame}
\usage{
\method{t}{ext_solutions_df}(x)
}
\arguments{
\item{x}{A \code{solutions_df} class object.}
}
\value{
A transposed \code{solutions_df} class object (class data.frame).
}
\description{
Transpose a solutions data frame
}
