% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterization.R
\name{top_om_per_cluster}
\alias{top_om_per_cluster}
\title{Select the top output matrix rows for each cluster}
\usage{
top_om_per_cluster(om)
}
\arguments{
\item{om}{an output matrix}
}
\value{
top_clusts_df dataframe with top om rows
}
\description{
Given an output matrix, returns a dataframe containing the row with the
lowest mean p-value and lowest min p-value for cluster sizes 2-5
}
