% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf.R
\name{two_step_merge}
\alias{two_step_merge}
\title{Two step SNF}
\usage{
two_step_merge(data_list)
}
\arguments{
\item{data_list}{nested list of input data generated by the function
\code{get_data_list()}}
}
\value{
fused_network The final fused network for clustering
}
\description{
Individual dataframes into individual similarity matrices into one fused
network per domain into one final fused network.
}
