% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{uids}
\alias{uids}
\title{Extract observations from a metasnf object}
\usage{
uids(x)
}
\arguments{
\item{x}{The object to extract observations from.}
}
\value{
Character vector of the UIDs of the observations stored in x.
}
\description{
This function returns a character vector of the UIDs of the observations
stored within an object from the metasnf package.
}
