% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ari_matrix.R
\name{validate_ari_matrix}
\alias{validate_ari_matrix}
\title{Validator for \code{ari_matrix} class object}
\usage{
validate_ari_matrix(aml)
}
\arguments{
\item{aml}{An ari_matrix-like matrix object to be validated.}
}
\value{
If aml has a valid structure for a \code{ari_matrix} class
object, returns the input unchanged. Otherwise, raises an error.
}
\description{
Validator for \code{ari_matrix} class object
}
\keyword{internal}
