% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_fns_list.R
\name{validate_clust_fns_list}
\alias{validate_clust_fns_list}
\title{Validator for \code{clust_fns_list} class object}
\usage{
validate_clust_fns_list(cfll)
}
\arguments{
\item{cfll}{A clust_fns_list-like \code{list} class object.}
}
\value{
If cfll has a valid structure for a \code{clust_fns_list} class object,
returns the input unchanged. Otherwise, raises an error.
}
\description{
Validator for \code{clust_fns_list} class object
}
\keyword{internal}
