% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_solutions_df.R
\name{validate_ext_solutions_df}
\alias{validate_ext_solutions_df}
\title{Validator for \code{ext_solutions_df} class object}
\usage{
validate_ext_solutions_df(ext_sol_dfl)
}
\arguments{
\item{ext_sol_dfl}{An extended solutions data frame-like object.}
}
\value{
If ext_sol_dfl has a valid structure for an object of class
ext_solutions_df, returns ext_sol_dfl. Otherwise, raises an error.
}
\description{
Validator for \code{ext_solutions_df} class object
}
\keyword{internal}
