% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_df.R
\name{validate_settings_df}
\alias{validate_settings_df}
\title{Validator for \code{settings_df} class object}
\usage{
validate_settings_df(sdfl)
}
\arguments{
\item{sdfl}{A settings data frame-like matrix object to be validated.}
}
\value{
If sdfl has a valid structure for a \code{settings_df} class object,
returns the input unchanged. Otherwise, raises an error.
}
\description{
Validator for \code{settings_df} class object
}
\keyword{internal}
