% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mats_list.R
\name{validate_sim_mats_list}
\alias{validate_sim_mats_list}
\title{Validator for \code{similarity_matrix_list} class object}
\usage{
validate_sim_mats_list(smll)
}
\value{
RETURN
}
\description{
Validator for \code{similarity_matrix_list} class object
}
