% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf_config.R
\name{validate_snf_config}
\alias{validate_snf_config}
\title{Validator for snf_config class object}
\usage{
validate_snf_config(scl)
}
\arguments{
\item{scl}{An SNF config-like \code{list} class object.}
}
\value{
If dll has a valid structure for a \code{data_list} class object,
returns input unchanged. Otherwise, raises an error.
}
\description{
Validator for snf_config class object
}
\keyword{internal}
