% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf_results.R
\name{validate_snf_results}
\alias{validate_snf_results}
\title{Validator for \code{snf_results} class object}
\usage{
validate_snf_results(srl)
}
\arguments{
\item{srl}{A snf_results-like list object to be validated.}
}
\value{
If srl has a valid structure for a \code{snf_results} class
object, returns the input unchanged. Otherwise, raises an error.
}
\description{
Validator for \code{snf_results} class object
}
\keyword{internal}
