% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_type.R
\name{validate_try_catch}
\alias{validate_try_catch}
\title{Helper function for is_valid tryCatch statement}
\usage{
validate_try_catch(x, validator_fn)
}
\arguments{
\item{x}{The object to check for validity}

\item{validator_fn}{Helper function containing tryCatch statement}
}
\value{
RETURN
}
\description{
Helper function for is_valid tryCatch statement
}
\keyword{internal}
