% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_matrix.R
\name{validate_weights_matrix}
\alias{validate_weights_matrix}
\title{Validator for \code{weights_matrix} class object}
\usage{
validate_weights_matrix(wml)
}
\arguments{
\item{wml}{A weights_matrix-like matrix object to be validated.}
}
\value{
If wml has a valid structure for a \code{weights_matrix} class
object, returns the input unchanged. Otherwise, raises an error.
}
\description{
Validator for \code{weights_matrix} class object
}
\keyword{internal}
