% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_plot.R
\name{var_manhattan_plot}
\alias{var_manhattan_plot}
\title{Manhattan plot of variable-variable associaiton p-values}
\usage{
var_manhattan_plot(
  data_list,
  key_var,
  neg_log_pval_thresh = 5,
  point_size = 5,
  text_size = 20,
  threshold = NULL,
  plot_title = NULL,
  hide_x_labels = FALSE,
  bonferroni_line = FALSE
)
}
\arguments{
\item{data_list}{List of dataframes containing data information.}

\item{key_var}{Variable for which the association p-values of all other
variables are plotted.}

\item{neg_log_pval_thresh}{Threshold for negative log p-values.}

\item{point_size}{Size of points in the plot.}

\item{text_size}{Size of text in the plot.}

\item{threshold}{p-value threshold to plot dashed line at.}

\item{plot_title}{Title of the plot.}

\item{hide_x_labels}{If TRUE, hides x-axis labels.}

\item{bonferroni_line}{If TRUE, plots a dashed black line at the
Bonferroni-corrected equivalent of the p-value threshold.}
}
\description{
Manhattan plot of variable-variable associaiton p-values
}
