% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_metrics.R
\name{weighted_euclidean_distance}
\alias{weighted_euclidean_distance}
\title{Distance metric: Weighted Euclidean distance}
\usage{
weighted_euclidean_distance(df, weights)
}
\arguments{
\item{df}{Dataframe containing one subjectkey column in the first column and
at least 1 continuous data column. All feature data should be continuous.}

\item{weights}{Dataframe with 1 column containing weights for each feature per
row in the same order as the order of feature columns start}
}
\value{
weighted_distance_matrix A distance matrix.
}
\description{
Distance metric: Weighted Euclidean distance
}
