% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_matrix.R
\name{weights_matrix}
\alias{weights_matrix}
\title{Generate a matrix to store feature weights}
\usage{
weights_matrix(dl = NULL, n_solutions = NULL, weights_fill = "ones")
}
\arguments{
\item{dl}{A nested list of input data from \code{data_list()}.}

\item{n_solutions}{Number of rows to generate the template weights matrix for.}

\item{weights_fill}{String indicating what to populate generate rows with.
Can be "ones" (default; fill matrix with 1), "uniform" (fill matrix with
uniformly distributed random values), or "exponential" (fill matrix with
exponentially distributed random values).}
}
\value{
wm A properly formatted matrix containing columns for
all the features that require weights and rows.
}
\description{
Generate a matrix to store feature weights
}
