% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_metrics.R
\name{wt_euclidean}
\alias{wt_euclidean}
\title{Weighted Euclidean distance}
\usage{
wt_euclidean(df, weights)
}
\arguments{
\item{df}{Dataframe containing at least 1 data column.}

\item{weights}{Vector of weights to apply to the columns}
}
\value{
distance_matrix A distance matrix.
}
\description{
Scales columns of a matrix by a vector of weights before calculating
Euclidean distance.
}
