% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_columns}
\alias{add_columns}
\title{Add columns to a dataframe}
\usage{
add_columns(df, newcols, fill = NA)
}
\arguments{
\item{df}{The dataframe to extend}

\item{newcols}{The vector containing new column names}

\item{fill}{The values of the elements of the newly added columns. NA by
default.}
}
\value{
extended_df The dataframe containing the added columns
}
\description{
Add new columns to a dataframe by providing a character vector of column
names (param \code{newcols}) and a value to occupy each row of the new columns
(param \code{fill}, NA by default).
}
