% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf.R
\name{add_design_matrix_rows}
\alias{add_design_matrix_rows}
\title{Add design matrix rows}
\usage{
add_design_matrix_rows(design_matrix, nrows)
}
\arguments{
\item{design_matrix}{The existing design matrix}

\item{nrows}{The number of rows to be added to the design matrix}
}
\value{
design_matrix New design matrix containing additional rows
}
\description{
Add design matrix rows
}
