% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_matrix.R
\name{add_settings_matrix_rows}
\alias{add_settings_matrix_rows}
\title{Add settings matrix rows}
\usage{
add_settings_matrix_rows(
  settings_matrix,
  nrows,
  min_removed = NULL,
  max_removed = NULL,
  possible_snf_schemes = c(1, 2, 3),
  retry_limit = 10
)
}
\arguments{
\item{settings_matrix}{The existing settings matrix}

\item{nrows}{The number of rows to be added to the settings matrix}

\item{min_removed}{The smallest number of elements that may be removed}

\item{max_removed}{The largest number of elements that may be removed}

\item{possible_snf_schemes}{A vector containing the possible snf_schemes to
uniformly randomly select from. By default, the vector contains all
possible schemes: c(1, 2, 3)}

\item{retry_limit}{The maximum number of attempts to generate a novel row}
}
\value{
settings_matrix New settings matrix containing additional rows
}
\description{
Add settings matrix rows
}
