% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_cluster.R
\name{ari_heatmap}
\alias{ari_heatmap}
\title{Heatmap meta-clustering results}
\usage{
ari_heatmap(
  output_matrix_aris,
  title = "",
  save = NULL,
  cluster_cols = TRUE,
  cluster_rows = TRUE,
  hide_columns = FALSE,
  hide_rows = FALSE,
  ...
)
}
\arguments{
\item{output_matrix_aris}{results from meta_cluster function}

\item{title}{plot title}

\item{save}{optional path to save figure to}

\item{cluster_cols}{boolean indicating if columns shold be clustered}

\item{cluster_rows}{boolean indicating if rows shold be clustered}

\item{hide_columns}{boolean indicating if column names should be hidden}

\item{hide_rows}{boolean indicating if row names should be hidden}

\item{...}{additional parameters to pass into pheatmap}
}
\description{
Heatmap meta-clustering results
}
