% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\name{arrange}
\alias{arrange}
\title{Arrange rows in an object}
\usage{
arrange(.data, ...)
}
\arguments{
\item{.data}{The object to arrange columns from.}

\item{...}{Additional arguments for arranging.}
}
\value{
Object with arrangeed columns.
}
\description{
Arrange rows in an object
}
