% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{arrange_dl}
\alias{arrange_dl}
\title{Given a data list object, sort data elements by uid}
\usage{
arrange_dl(dl)
}
\arguments{
\item{dl}{A nested list of input data from \code{data_list()}.}
}
\value{
arranged_dl The arranged data list object
}
\description{
Given a data list object, sort data elements by uid
}
