% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{arrange_dl}
\alias{arrange_dl}
\title{Given a data_list object, sort data elements by subjectkey}
\usage{
arrange_dl(data_list)
}
\arguments{
\item{data_list}{A nested list of input data from \code{generate_data_list()}.}
}
\value{
arranged_data_list The arranged data_list object
}
\description{
Given a data_list object, sort data elements by subjectkey
}
