% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_weights_matrix.R
\name{as_weights_matrix}
\alias{as_weights_matrix}
\title{Convert an object to a weights matrix}
\usage{
as_weights_matrix(x)
}
\arguments{
\item{x}{The object to convert into a data list.}
}
\value{
A \code{weights_matrix} class object.
}
\description{
This function converts non-\code{weights_matrix} objects into \code{weights_matrix}
class objects.
}
