% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{assemble_data}
\alias{assemble_data}
\title{Collapse a dataframe and/or a data_list into a single dataframe}
\usage{
assemble_data(data, data_list)
}
\arguments{
\item{data}{A dataframe.}

\item{data_list}{A nested list of input data from \code{generate_data_list()}.}
}
\value{
A class "data.frame" object containing all the features of the
provided data frame and/or data list.
}
\description{
Collapse a dataframe and/or a data_list into a single dataframe
}
