% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{assigned_clust_scatter}
\alias{assigned_clust_scatter}
\title{Scatter plot of cluster membership across an solutions matrix}
\usage{
assigned_clust_scatter(om, save = NULL)
}
\arguments{
\item{om}{solutions matrix}

\item{save}{optional path to save figure to}
}
\description{
Scatter plot of cluster membership across an solutions matrix
}
