% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_snf.R
\name{batch_snf}
\alias{batch_snf}
\title{Run variations of SNF as described by a design matrix}
\usage{
batch_snf(data_list, design_matrix, processes = 1)
}
\arguments{
\item{data_list}{nested list of input data generated by the function
\code{get_data_list()}}

\item{design_matrix}{matrix indicating parameters to iterate SNF through}

\item{processes}{Specify number of processes used to complete SNF iterations
* \code{1} (default) Sequential processing: function will iterate through the
\code{design_matrix} one row at a time with a for loop. This option will
not make use of multiple CPU cores, but will show a progress bar.
* \code{2} or higher: Parallel processing will use the
\code{future.apply::future_apply} to distribute the SNF iterations across
the specified number of CPU cores. If higher than the number of
available cores, a warning will be printed and the maximum number of
cores will be used.
* \code{max}: All available cores will be used.}
}
\value{
populated_design_matrix design matrix with filled columns related to
subtype membership
}
\description{
Run variations of SNF as described by a design matrix
}
