% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf.R
\name{build_design_matrix}
\alias{build_design_matrix}
\title{Build a design matrix}
\usage{
build_design_matrix(nrow, seed = NULL)
}
\arguments{
\item{nrow}{number of design matrix rows}

\item{seed}{set a seed for the random matrix generation. Note that this
affects the global seed.}
}
\value{
design_matrix A design matrix
}
\description{
Build a design matrix
}
