% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf.R
\name{build_design_matrix_base}
\alias{build_design_matrix_base}
\title{Construct the base of the design matrix}
\usage{
build_design_matrix_base()
}
\value{
design_matrix A skeleton dataframe to build a design matrix out of
}
\description{
Construct the base of the design matrix
}
