% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions_matrix.R
\name{calc_assoc_pval_matrix}
\alias{calc_assoc_pval_matrix}
\title{Calculate p-values for all pairwise associations of features in a data_list}
\usage{
calc_assoc_pval_matrix(data_list, verbose = FALSE, cat_test = "chi_squared")
}
\arguments{
\item{data_list}{A nested list of input data from \code{generate_data_list()}.}

\item{verbose}{If TRUE, prints new line everytime a p-value is being
calculated.}

\item{cat_test}{String indicating which statistical test will be used to
associate cluster with a categorical feature. Options are "chi_squared" for
the Chi-squared test and "fisher_exact" for Fisher's exact test.}
}
\value{
A "matrix" class object containing pairwise association p-values
between the features in the provided data list.
}
\description{
Calculate p-values for all pairwise associations of features in a data_list
}
