% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_measures.R
\name{calculate_db_indices}
\alias{calculate_db_indices}
\title{Calculate Davies-Bouldin indices}
\usage{
calculate_db_indices(sol_df)
}
\arguments{
\item{sol_df}{A sol_df (see ?batch_snf)}
}
\value{
davies_bouldin_indices A vector of Davies-Bouldin indices for each
cluster solution.
}
\description{
Given a sol_df and a list of similarity_matrices (or a single
similarity_matrix if the sol_df has only 1 row), return a vector of
Davies-Bouldin indices
}
