% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_measures.R
\name{calculate_db_indices}
\alias{calculate_db_indices}
\title{Calculate Davies-Bouldin indices}
\usage{
calculate_db_indices(solutions_matrix, similarity_matrices)
}
\arguments{
\item{solutions_matrix}{A solutions_matrix (see ?batch_snf)}

\item{similarity_matrices}{A list of similarity matrices (see ?batch_snf)}
}
\value{
davies_bouldin_indices A vector of Davies-Bouldin indices for each
cluster solution.
}
\description{
Given a solutions_matrix and a list of similarity_matrices (or a single
similarity_matrix if the solutions_matrix has only 1 row), return a vector of
Davies-Bouldin indices
}
