% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_measures.R
\name{calculate_dunn_indices}
\alias{calculate_dunn_indices}
\title{Calculate Dunn indices}
\usage{
calculate_dunn_indices(solutions_matrix, similarity_matrices)
}
\arguments{
\item{solutions_matrix}{A solutions_matrix (see ?batch_snf)}

\item{similarity_matrices}{A list of similarity matrices (see ?batch_snf)}
}
\value{
dunn_indices A vector of Dunn indices for each cluster solution
}
\description{
Given a solutions_matrix and a list of similarity_matrices (or a single
similarity_matrix if the solutions_matrix has only 1 row), return vector of
Dunn indices
}
