% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_measures.R
\name{calculate_silhouettes}
\alias{calculate_silhouettes}
\title{Calculate silhouette scores}
\usage{
calculate_silhouettes(solutions_matrix, affinity_matrices)
}
\arguments{
\item{solutions_matrix}{A solutions_matrix (see ?batch_snf)}

\item{affinity_matrices}{A list of affinity matrices (see ?batch_snf)}
}
\value{
silhouette_scores A list of "silhouette" objects from the cluster
package.
}
\description{
Given a solutions_matrix and a list of affinity_matrices (or a single
affinity_matrix if the solutions_matrix has only 1 row), return a list
of 'silhouette' objects from the cluster package
}
\examples{

# load package
library(metasnf)

# generate data_list
data_list <- generate_data_list(
    list(abcd_cort_t, "cortical_thickness", "neuroimaging", "continuous"),
    list(abcd_cort_sa, "cortical_surface_area", "neuroimaging", "continuous"),
    list(abcd_subc_v, "subcortical_volume", "neuroimaging", "continuous"),
    list(abcd_income, "household_income", "demographics", "continuous"),
    list(abcd_pubertal, "pubertal_status", "demographics", "continuous"),
    uid = "patient"
)

# build settings_matrix
settings_matrix <- generate_settings_matrix(data_list, nrow = 15, seed = 42)

# collect affinity matrices and solutions matrix from batch_snf
batch_snf_results <- batch_snf(
    data_list,
    settings_matrix,
    return_affinity_matrices = TRUE
)

solutions_matrix <- batch_snf_results$"solutions_matrix"
affinity_matrices <- batch_snf_results$"affinity_matrices"

# calculate silhouette scores
silhouette_scores <- calculate_silhouettes(
    solutions_matrix,
    affinity_matrices
)

# plot the silhouette scores of the first solutions
plot(silhouette_scores[[1]])

}
