% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterization.R
\name{cbcl_ord_reg_from_om}
\alias{cbcl_ord_reg_from_om}
\title{Calculate overall p-values for an om}
\usage{
cbcl_ord_reg_from_om(om, cbcl_list, bonferroni = FALSE)
}
\arguments{
\item{om}{an output matrix}

\item{cbcl_list}{a list of CBCL measures}

\item{bonferroni}{boolean for reporting bonferroni corrected p-values}
}
\description{
Calculate overall p-values for an om
}
