% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{char_to_fac}
\alias{char_to_fac}
\title{Convert character-type columns of a dataframe to factor-type}
\usage{
char_to_fac(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
df_converted The dataframe with factor-type columns instead of
char-type columns
}
\description{
Convert character-type columns of a dataframe to factor-type
}
