% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{char_to_fac}
\alias{char_to_fac}
\title{Convert character-type columns of a data frame to factor-type}
\usage{
char_to_fac(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
The data frame with factor columns instead of char columns.
}
\description{
Convert character-type columns of a data frame to factor-type
}
\keyword{internal}
