% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_plotting.R
\name{clusterToOutcomeManhattan}
\alias{clusterToOutcomeManhattan}
\title{Display cluster assignment to outcome correlation as Manhattan plot}
\usage{
clusterToOutcomeManhattan(target_pvals, levels = NULL)
}
\arguments{
\item{target_pvals}{short for Correlation of Clusters vs Outcomes (cco). It is a dataframe with columns:
datatype: data type being integrated and clustered. Think of this as predictor
outcomes: outcome variables computed against datatype. Think of this as outcome
p_value: p_value from statistical testing on datatype clusters vs outcome
size:}

\item{levels}{optional argument to re-arrange outcome display on x-axis}
}
\value{
plot A Manhattan plot
}
\description{
Manhattan plot plots the correlation of SNF clustering to specified outcome, colored by data types, dot size represents sample size.
}
