% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coclustering.R
\name{cocluster_density}
\alias{cocluster_density}
\title{Density plot coclustering stability across subsampled data.}
\usage{
cocluster_density(cocluster_df)
}
\arguments{
\item{cocluster_df}{A dataframe containing coclustering data for a single
cluster solution. This object is generated by the \code{calculate_coclustering}
function.}
}
\value{
Density plot (class "gg", "ggplot") of the distribution of
coclustering across pairs and subsamples of the data.
}
\description{
This function creates a density plot that shows, for all pairs of
observations that originally clustered together, the distribution of the
the fractions that those pairs clustered together across subsampled data.
}
