% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{collapse_dl}
\alias{collapse_dl}
\title{Collapse a dl into a single dataframe}
\usage{
collapse_dl(dl)
}
\arguments{
\item{dl}{A nested list of input data from \code{data_list()}.}
}
\value{
A "data.frame"-formatted version of the provided data list.
}
\description{
Collapse a dl into a single dataframe
}
