% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{collapse_dl}
\alias{collapse_dl}
\title{Convert a data list into a data frame}
\usage{
collapse_dl(data_list)
}
\arguments{
\item{data_list}{A nested list of input data from \code{generate_data_list()}.}
}
\value{
A "data.frame"-formatted version of the provided data list.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Defunct function for converting a data list into a data frame. Please
use \code{as.data.frame()} instead.
}
\keyword{internal}
