% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{convert_uids}
\alias{convert_uids}
\title{Convert unique identifiers of data list to "uid"}
\usage{
convert_uids(dll, uid)
}
\arguments{
\item{dll}{A data list-like \code{list} class object.}

\item{uid}{(string) the name of the uid column currently used data}
}
\value{
The provided nested list with "uid" as UID.
}
\description{
Column name "uid" is reserved for the unique identifier of observations.
This function ensures all data frames have their UID set as "uid".
}
\keyword{internal}
