% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{corrHeatmap_legend}
\alias{corrHeatmap_legend}
\title{Generate correlation heatmap legend}
\usage{
corrHeatmap_legend(legend_name, labels, labels_color)
}
\arguments{
\item{legend_name}{graph path to be saved to}

\item{labels}{argument to specify outcome label names}

\item{labels_color}{argument to specify outcome label name colors}
}
\value{
pd a collection of heatmap legends saved to a file in current path
}
\description{
Generate legend for correlation heatmap.
}
