% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_plotting.r
\name{CorrManhattan}
\alias{CorrManhattan}
\title{Generate predictor-outcome correlations in Manhattan plot}
\usage{
CorrManhattan(df_export, outcome)
}
\arguments{
\item{df_export}{is a dataframe with features in rownames, and columns:
"p.value": from correlation test,
"n": number of samples,
"Group": datatype name,
"Group_index": sequence of datatypes to be displayed}

\item{outcome}{name the correlations were computed against. To be displayed in plot title}
}
\description{
Manhattan plot showing predictor correlations to an outcome
}
