% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_measures.R
\name{db_manual}
\alias{db_manual}
\title{Calculate Davies-Bouldin index}
\usage{
db_manual(dist_mat, clusters)
}
\arguments{
\item{dist_mat}{A dissimilarity matrix.}

\item{clusters}{An integer vector of cluster assignments for that matrix.}
}
\value{
Davies-Bouldin index (numeric).
}
\description{
Calculate Davies-Bouldin index
}
\keyword{internal}
