% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snftool_variations.R
\name{discretisation}
\alias{discretisation}
\title{Adaptation of SNFtool's discretisation function}
\usage{
discretisation(eigenvectors)
}
\arguments{
\item{eigenvectors}{Eigenvectors}
}
\description{
Adaptation of SNFtool's discretisation function
}
