% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nclust_estimation.R
\name{discretisation}
\alias{discretisation}
\title{Internal function for \code{estimate_nclust_given_graph}}
\usage{
discretisation(eigenvectors)
}
\arguments{
\item{eigenvectors}{Matrix of eigenvectors.}
}
\value{
"Matrix" class object, intermediate product in spectral clustering.
}
\description{
Internal function taken from \code{SNFtool} to use for number of cluster
estimation.
}
\keyword{internal}
