% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{discretisationeigenvectordata}
\alias{discretisationeigenvectordata}
\title{Adaptation of SNFtool's discretisationeigenvectordata}
\usage{
discretisationeigenvectordata(eigenvector)
}
\arguments{
\item{eigenvector}{Eigenvector}
}
\description{
Adaptation of SNFtool's discretisationeigenvectordata
}
