% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_plotting.r
\name{displayClustersHeatmap}
\alias{displayClustersHeatmap}
\title{Display SNF cluster output in Heatmap}
\usage{
displayClustersHeatmap(
  W,
  group_cluster,
  top_annotation = NULL,
  left_annotation = NULL,
  scale_method = "none"
)
}
\arguments{
\item{W}{similarity matrix outputted from SNFtool's SNF function, with rownames and colnames in sample IDs}

\item{group_cluster}{cluster assignment from spectral clustering}

\item{top_annotation}{annotation to be displayed above the heatmap output}

\item{left_annotation}{annotation to be displayed on the left of the heatmap output}

\item{scale_method}{Method of rescaling matrix diagonals. Can be "none"
(default - don't change diagonals), "mean" (replace diagonals with average
value of off-diagonals), or "zero" (replace diagonals with 0).}
}
\description{
Normalize SNF matrix and plot heatmap.
}
