% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_plotting.r
\name{displayClustersHeatmap}
\alias{displayClustersHeatmap}
\title{Display SNF cluster output in Heatmap}
\usage{
displayClustersHeatmap(
  W,
  group_cluster,
  top_annotation = NULL,
  left_annotation = NULL
)
}
\arguments{
\item{W}{similarity matrix outputted from SNFtool's SNF function, with rownames and colnames in sample IDs}

\item{group_cluster}{cluster assignment from spectral clustering}

\item{top_annotation}{annotation to be displayed above the heatmap output}

\item{left_annotation}{annotation to be displayed on the left of the heatmap output}
}
\description{
Normalize SNF matrix and plot heatmap.
}
