% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{dl_has_duplicates}
\alias{dl_has_duplicates}
\title{Check if data list contains any duplicate features}
\usage{
dl_has_duplicates(dl)
}
\arguments{
\item{dl}{A nested list of input data from \code{generate_data_list()}.}
}
\value{
Doesn't return any value. Raises warning if there are features
with duplicate names in a generated data list.
}
\description{
Check if data list contains any duplicate features
}
