% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{dl_uid_first_col}
\alias{dl_uid_first_col}
\title{Make the subjectkey UID columns of a data_list first}
\usage{
dl_uid_first_col(data_list)
}
\arguments{
\item{data_list}{A nested list of input data from \code{generate_data_list()}.}
}
\value{
A data list ("list"-class object) in which each data-subcomponent
has "subjectkey" positioned as its first column.
}
\description{
Make the subjectkey UID columns of a data_list first
}
