% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{dl_variable_summary}
\alias{dl_variable_summary}
\title{Variable-level summary of a data_list}
\usage{
dl_variable_summary(data_list)
}
\arguments{
\item{data_list}{A nested list of input data from \code{generate_data_list()}.}
}
\value{
variable_level_summary A dataframe containing the name, type, and
domain of every variable in a data_list.
}
\description{
Variable-level summary of a data_list
}
