% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{dl_variable_summary}
\alias{dl_variable_summary}
\title{Variable-level summary of a data list}
\usage{
dl_variable_summary(dl)
}
\arguments{
\item{dl}{A nested list of input data from \code{data_list()}.}
}
\value{
variable_level_summary A data frame containing the name, type, and
domain of every variable in a data list.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Defunct function to summarize a data list. Please use \code{summary()} with
argument \code{scope = "feature"} instead.
}
