% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{domain_merge}
\alias{domain_merge}
\title{Domain merge}
\usage{
domain_merge(data_list)
}
\arguments{
\item{data_list}{nested list of input data generated by the function
\code{get_data_list()}}
}
\value{
domain_dl
}
\description{
Given a data_list, returns a new data_list where all original data objects of
a particlar domain have been concatenated
}
