% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_snf.R
\name{domain_merge}
\alias{domain_merge}
\title{SNF scheme: Domain merge}
\usage{
domain_merge(
  data_list,
  cont_dist_fn,
  disc_dist_fn,
  ord_dist_fn,
  cat_dist_fn,
  mix_dist_fn,
  weights_row,
  k,
  alpha,
  t
)
}
\arguments{
\item{data_list}{A nested list of input data from \code{generate_data_list()}.}

\item{cont_dist_fn}{distance metric function for continuous data.}

\item{disc_dist_fn}{distance metric function for discrete data.}

\item{ord_dist_fn}{distance metric function for ordinal data.}

\item{cat_dist_fn}{distance metric function for categorical data.}

\item{mix_dist_fn}{distance metric function for mixed data.}

\item{weights_row}{dataframe row containing feature weights.}

\item{k}{k hyperparameter.}

\item{alpha}{alpha/eta/sigma hyperparameter.}

\item{t}{SNF number of iterations hyperparameter.}
}
\value{
fused_network The final fused network (class "matrix", "array")
generated by SNF.
}
\description{
Given a data_list, returns a new data_list where all data objects of
a particlar domain have been concatenated.
}
