% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{domains}
\alias{domains}
\title{Pull domains from a data list}
\usage{
domains(dl)
}
\arguments{
\item{dl}{A nested list of input data from \code{data_list()}.}
}
\value{
A character vector of domains.
}
\description{
Pull domains from a data list
}
\keyword{internal}
