% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_metrics_list.R
\name{euclidean_distance}
\alias{euclidean_distance}
\title{Distance metric: Euclidean distance}
\usage{
euclidean_distance(df, weights_row)
}
\arguments{
\item{df}{Dataframe containing at least 1 data column}

\item{weights_row}{Single-row dataframe where the column names contain the
column names in df and the row contains the corresponding weights_row.}
}
\value{
distance_matrix A distance matrix.
}
\description{
Distance metric: Euclidean distance
}
