% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf.R
\name{execute_design_matrix}
\alias{execute_design_matrix}
\title{Execute variations of SNF as described by a design matrix}
\usage{
execute_design_matrix(
  design_matrix,
  mtbi_loc = NULL,
  mtbi_mechanism = NULL,
  mtbi_mem_daze = NULL,
  income = NULL,
  interview_age = NULL,
  mtbi_age = NULL,
  pubertal_status = NULL,
  race_cd = NULL,
  race_ed = NULL,
  sex = NULL,
  wmndf = NULL,
  headaches = NULL,
  mtbi_count = NULL,
  gord_cor = NULL,
  gord_var = NULL,
  subc_cor = NULL,
  subc_var = NULL,
  cort_sa = NULL,
  cort_t = NULL,
  subc_v = NULL
)
}
\arguments{
\item{design_matrix}{Design matrix
* as built by fns \code{build_design_matrix_base} and \code{add_design_matrix_rows}}

\item{mtbi_loc}{mtbi_loc dataframe generated by \verb{get_*} fns}

\item{mtbi_mechanism}{mtbi_mechanisc dataframe generated by \verb{get_*} fns}

\item{mtbi_mem_daze}{mtbi_mem_dazc dataframe generated by \verb{get_*} fns}

\item{income}{incomc dataframe generated by \verb{get_*} fns}

\item{interview_age}{interview_agc dataframe generated by \verb{get_*} fns}

\item{mtbi_age}{mtbi_agc dataframe generated by \verb{get_*} fns}

\item{pubertal_status}{pubertal_statuc dataframe generated by \verb{get_*} fns}

\item{race_cd}{race_cc dataframe generated by \verb{get_*} fns}

\item{race_ed}{race_ec dataframe generated by \verb{get_*} fns}

\item{sex}{sec dataframe generated by \verb{get_*} fns}

\item{wmndf}{wmndc dataframe generated by \verb{get_*} fns}

\item{headaches}{headachec dataframe generated by \verb{get_*} fns}

\item{mtbi_count}{mtbi_counc dataframe generated by \verb{get_*} fns}

\item{gord_cor}{gord_coc dataframe generated by \verb{get_*} fns}

\item{gord_var}{gord_vac dataframe generated by \verb{get_*} fns}

\item{subc_cor}{subc_coc dataframe generated by \verb{get_*} fns}

\item{subc_var}{subc_vac dataframe generated by \verb{get_*} fns}

\item{cort_sa}{cort_sc dataframe generated by \verb{get_*} fns}

\item{cort_t}{cort_c dataframe generated by \verb{get_*} fns}

\item{subc_v}{subc_c dataframe generated by \verb{get_*} fns}
}
\value{
design_matrix_executed Design matrix with populated subtype
membership and target variable columns
}
\description{
Execute variations of SNF as described by a design matrix
}
