% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_snf.R
\name{execute_inclusion}
\alias{execute_inclusion}
\title{Execute inclusion}
\usage{
execute_inclusion(settings_matrix, data_list)
}
\arguments{
\item{settings_matrix}{matrix indicating parameters to iterate SNF through}

\item{data_list}{nested list of input data generated by the function
\code{get_data_list()}}
}
\value{
selected_data_list
}
\description{
Given a data list and a settings matrix row, returns a data list of selected
inputs
}
