% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{execute_inclusion}
\alias{execute_inclusion}
\title{Execute inclusion}
\usage{
execute_inclusion(data_list, design_matrix)
}
\arguments{
\item{data_list}{nested list of input data generated by the function
\code{get_data_list()}}

\item{design_matrix}{matrix indicating parameters to iterate SNF through}
}
\value{
selected_data_list
}
\description{
Given a data list and a design matrix row, returns a data list of selected
inputs
}
