% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_subjects}
\alias{filter_subjects}
\title{Subset a dataframe to a given subject list}
\usage{
filter_subjects(abcd_df, subjects = NULL)
}
\arguments{
\item{abcd_df}{An ABCD dataframe}

\item{subjects}{Dataframe containing list of required subjects}
}
\value{
filtered_df The subsetted dataframe
}
\description{
Subset a dataframe to a given subject list
}
