% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{generate_clust_algs_list}
\alias{generate_clust_algs_list}
\title{Generate a clustering algorithms list}
\usage{
generate_clust_algs_list(..., disable_base = FALSE)
}
\arguments{
\item{...}{An arbitrary number of named clustering functions}

\item{disable_base}{If TRUE, do not prepend the base clustering algorithms
(spectral_eigen and spectral_rot, which apply spectral clustering and use
the eigen-gap and rotation cost heuristics respectively for determining
the number of clusters in the graph.}
}
\value{
A list of clustering algorithm functions that can
be passed into the batch_snf and generate_settings_list functions.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Deprecated function for building a clustering algorithms list. Please use
\code{clust_fns_list()} (or better yet, \code{snf_config()}) instead.
}
