% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{generate_data_list}
\alias{generate_data_list}
\title{Generate data_list object - but softcoded}
\usage{
generate_data_list(
  ...,
  old_uid = NULL,
  test_subjects = NULL,
  train_subjects = NULL,
  assigned_splits = NULL
)
}
\arguments{
\item{...}{Lists formatted as (df, "df_name", "df_domain", "df_type")}

\item{old_uid}{(string) the name of the uid column currently used data}

\item{test_subjects}{character vector of test subjects (useful if building
a full data list for label propagation)}

\item{train_subjects}{character vector of train subjects (useful if building
a full data list for label propagation)}

\item{assigned_splits}{ouptut from assign_splits function - can be given
as an alternative to specifying the train/test subjects separately.}
}
\description{
This is the major data object that will be processed when iterating through
the design matrix. The full list contains one list per measurement type.
Within each measurement type's list, elements include the actual data
structure, the name, the domain, and the data 'type' (i.e, numeric or
categorical).
}
\details{
To-do: include checks to make sure format of data list is correct
}
\examples{
heart_rate_df <- data.frame(
    patient_id = c("1", "2", "3"),
    var1 = c(0.04, 0.1, 0.3),
    var2 = c(30, 2, 0.3)
)

personality_test_df <- data.frame(
    patient_id = c("1", "2", "3"),
    var3 = c(900, 1990, 373),
    var4 = c(509, 2209, 83)
)

dl <- generate_data_list(
    list(heart_rate_df, "data1", "domain1", "numeric"),
    list(personality_test_df, "data2", "domain2", "numeric"),
    old_uid = "patient_id"
)
}
