% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{generate_distance_metrics_list}
\alias{generate_distance_metrics_list}
\title{Generate a list of distance metrics}
\usage{
generate_distance_metrics_list(
  continuous_distances = NULL,
  discrete_distances = NULL,
  ordinal_distances = NULL,
  categorical_distances = NULL,
  mixed_distances = NULL,
  keep_defaults = TRUE
)
}
\arguments{
\item{continuous_distances}{A named list of distance metric functions}

\item{discrete_distances}{A named list of distance metric functions}

\item{ordinal_distances}{A named list of distance metric functions}

\item{categorical_distances}{A named list of distance metric functions}

\item{mixed_distances}{A named list of distance metric functions}

\item{keep_defaults}{If TRUE (default), prepend the base distance metrics
(euclidean and standard normalized euclidean)}
}
\value{
A nested and named list of distance metrics functions.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Deprecated function for building a distance metrics list. Please use
\code{dist_fns_list()} (or better yet, \code{snf_config()}) instead.
}
\keyword{internal}
