% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf.R
\name{generate_outcome_list}
\alias{generate_outcome_list}
\title{Generate outcome_list object}
\usage{
generate_outcome_list(
  cbcl_headaches = NULL,
  cbcl_nausea = NULL,
  cbcl_vomiting = NULL,
  cbcl_dizzy = NULL,
  cbcl_overtired = NULL,
  cbcl_sleeping_more = NULL,
  cbcl_sleeping_less = NULL,
  cbcl_depress_r = NULL,
  cbcl_anxiety_r = NULL,
  cbcl_attention_r = NULL,
  cbcl_aggressive_r = NULL,
  nihtbx_list_fc = NULL,
  nihtbx_pattern_fc = NULL,
  nihtbx_cardsort_fc = NULL,
  sds_sleep = NULL
)
}
\description{
The major object containing all outcome variables
}
