% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{generate_settings_matrix}
\alias{generate_settings_matrix}
\title{Build a settings data frame}
\usage{
generate_settings_matrix(...)
}
\arguments{
\item{...}{Arguments used to generate a settings matrix.}
}
\value{
Raises a deprecated error.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Deprecated function for building a settings matrix. Please use
\code{settings_df()} instead.
}
