% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{generate_target_list}
\alias{generate_target_list}
\title{(DEPRECATED) Generate target_list object}
\usage{
generate_target_list(...)
}
\arguments{
\item{...}{Function arguments}
}
\description{
(DEPRECATED) Generate target_list object
}
