% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_metrics.R
\name{generate_weights_matrix}
\alias{generate_weights_matrix}
\title{Generate a matrix to store variable weights}
\usage{
generate_weights_matrix(data = NULL, data_list = NULL, rows = 1, fill = "ones")
}
\arguments{
\item{data}{A dataframe with column names spanning all the variables that
may ever require weights. Cannot be provided at the same time as the
data_list parameter.}

\item{data_list}{A data_list with column names spanning all the variables that
may ever require weights. Cannot be provided at the same time as the
data parameter.}

\item{rows}{Number of rows to generate the template weights matrix for.}

\item{fill}{String indicating what to populate generate rows with. Can be
"ones" (default; fill matrix with 1), "uniform" (fill matrix with uniformly
distributed random values), or "exponential" (fill matrix with
exponentially distributed random values).}
}
\value{
weights_matrix A properly formatted matrix containing columns for
all the variables that require weights and rows
}
\description{
Generate a matrix to store variable weights
}
