% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_matrix.R
\name{generate_weights_matrix}
\alias{generate_weights_matrix}
\title{Generate a matrix to store feature weights}
\usage{
generate_weights_matrix(data_list = NULL, nrow = 1, fill = "ones")
}
\arguments{
\item{data_list}{A nested list of input data from \code{generate_data_list()}.}

\item{nrow}{Number of rows to generate the template weights matrix for.}

\item{fill}{String indicating what to populate generate rows with. Can be
"ones" (default; fill matrix with 1), "uniform" (fill matrix with uniformly
distributed random values), or "exponential" (fill matrix with
exponentially distributed random values).}
}
\value{
weights_matrix A properly formatted matrix containing columns for
all the features that require weights and rows.
}
\description{
Generate a matrix to store feature weights
}
