% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbcl.R
\name{get_cbcl_depress_r}
\alias{get_cbcl_depress_r}
\title{Get CBCL depression data}
\usage{
get_cbcl_depress_r(abcd_cbcls01, subjects = NULL)
}
\arguments{
\item{subjects}{list of subjects to receive data for}

\item{abcd_cbcl01}{NDA cbcl dataframe}
}
\value{
cbcl_depress_r depression data
}
\description{
Get CBCL depression data
}
