% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions_matrix.R
\name{get_cluster_pval}
\alias{get_cluster_pval}
\title{Get p-value}
\usage{
get_cluster_pval(
  assigned_subs,
  outcome_df,
  outcome_type,
  outcome_name,
  cat_test = "chi_squared"
)
}
\arguments{
\item{assigned_subs}{dataframe of subjects who were assigned to a cluster
and the cluster they were assigned to.}

\item{outcome_df}{dataframe containing subjectkey and outcome feautre column}

\item{outcome_type}{string indicating the outcome type (numeric or ordinal)}

\item{outcome_name}{string indicating the name of the feature}

\item{cat_test}{String indicating which statistical test will be used to
associate cluster with a categorical variable. Options are "chi_squared" for
the Chi-squared test and "fisher_exact" for Fisher's exact test.}
}
\value{
p_val the smallest p-value of interest
}
\description{
Depending on outcome measure, perform ordinal regression or linear regression
and return p-value as a benchmark measure of how well-separated clusters
are by the outcome measure.
}
