% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_extraction.R
\name{get_cluster_solutions}
\alias{get_cluster_solutions}
\title{Extract cluster membership information from a solutions_matrix}
\usage{
get_cluster_solutions(solutions_matrix)
}
\arguments{
\item{solutions_matrix}{A solutions_matrix.}
}
\value{
cluster_solutions A "data.frame" object where each row is an
observation and each column (apart from the subjectkey column) indicates
the cluster that observation was assigned to for the corresponding
solutions matrix row.
}
\description{
This function takes in a solutions matrix and returns a dataframe containing
the cluster assignments for each subjectkey. It is similar to
'\code{get_clusters()}, which takes one solutions matrix row and returns a vector
of cluster assignments' and \code{get_cluster_df()}, which takes a solutions
matrix with only one row and returns a dataframe with two columns: "cluster"
and "subjectkey" (the UID of the observation).
}
