% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_cluster_solutions}
\alias{get_cluster_solutions}
\title{Extract cluster membership information from a sol_df}
\usage{
get_cluster_solutions(sol_df)
}
\arguments{
\item{sol_df}{A sol_df.}
}
\value{
A "data.frame" object where each row is an
observation and each column (apart from the uid column) indicates
the cluster that observation was assigned to for the corresponding
solutions data frame row.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Deprecated function for building extracting cluster solutions from a
solutions data frame. Please use \code{t()} instead.

This function takes in a solutions data frame and returns a data frame containing
the cluster assignments for each uid. It is similar to
'\code{get_clusters()}, which takes one solutions data frame row and returns a vector
of cluster assignments' and \code{get_cluster_df()}, which takes a solutions
matrix with only one row and returns a data frame with two columns: "cluster"
and "uid" (the UID of the observation).
}
