% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions_matrix.R
\name{get_clustered_subs}
\alias{get_clustered_subs}
\title{Get clustered subjects}
\usage{
get_clustered_subs(solutions_matrix_row)
}
\arguments{
\item{solutions_matrix_row}{Output matrix row containing subtype membership}
}
\value{
clustered_subs Dataframe
}
\description{
Pull a dataframe of clustered subjects from an solutions matrix structure.
}
