% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_clusters}
\alias{get_clusters}
\title{Extract cluster membership vector from one solutions data frame row}
\usage{
get_clusters(sol_df_row)
}
\arguments{
\item{sol_df_row}{Output matrix row.}
}
\value{
clusters Vector of assigned clusters.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Deprecated function for building extracting cluster solutions from a
solutions data frame. Please use \code{t()} instead.

This function takes in a single row of a solutions data frame and returns a
vector containing the cluster assignments for each observation. It is
similar to \code{get_cluster_df()}, which takes a solutions data frame with only one
row and returns a data frame with two columns: "cluster" and "uid"
'(the UID of the observation) and \code{get_cluster_solutions()}, which takes a
solutions data frame with any number of rows and returns a data frame indicating
the cluster assignments for each of those rows.
}
\keyword{internal}
