% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroimaging.R
\name{get_cort_sa}
\alias{get_cort_sa}
\title{Extract cortical surface areas}
\usage{
get_cort_sa(mrisdp10201, subjects)
}
\arguments{
\item{mrisdp10201}{Data file containing cortical data}

\item{subjects}{Optional dataframe containing needed subjectkeys}
}
\value{
cort_t_df Dataframe of cortical surface areas
}
\description{
Extract cortical surface areas
}
