% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_metrics.R
\name{get_dist_matrix}
\alias{get_dist_matrix}
\title{Calculate distance matrices}
\usage{
get_dist_matrix(
  df,
  input_type,
  scale = FALSE,
  cont_dist_fn,
  disc_dist_fn,
  ord_dist_fn,
  cat_dist_fn,
  mix_dist_fn
)
}
\arguments{
\item{df}{Raw dataframe with subject IDs in column 1}

\item{input_type}{Either "numeric" (resulting in euclidean distances),
"categorical" (resulting in binary distances), or "mixed" (resulting in
gower distances)}

\item{scale}{Whether or not the data should be standard normalized prior to
distance calculations}

\item{cont_dist_fn}{distance metric function for continuous data}

\item{disc_dist_fn}{distance metric function for discrete data}

\item{ord_dist_fn}{distance metric function for ordinal data}

\item{cat_dist_fn}{distance metric function for categorical data}

\item{mix_dist_fn}{distance metric function for mixed data}
}
\value{
dist_matrix Matrix of inter-observation distances
}
\description{
Given a dataframe of numerical variables, return a euclidean distance matrix
}
