% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_dl_uids}
\alias{get_dl_uids}
\title{Extract UIDs from a data list}
\usage{
get_dl_uids(dl, prefix = FALSE)
}
\arguments{
\item{dl}{A nested list of input data from \code{data_list()}.}

\item{prefix}{If TRUE, preserves the "uid_" prefix added to UIDs when
creating a data list.}
}
\value{
A character vector of the UID labels contained in a data list.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Deprecated function for extracting UIDs from a data list.
Please use \code{uids()} instead.
}
\keyword{internal}
