% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{get_heatmap_order}
\alias{get_heatmap_order}
\title{Return the row or column ordering present in a heatmap}
\usage{
get_heatmap_order(heatmap, type = "rows")
}
\arguments{
\item{heatmap}{A heatmap object to collect ordering from.}

\item{type}{The type of ordering to return. Either "rows" or "columns".}
}
\value{
A numeric vector of the ordering used within the provided
ComplexHeatmap "Heatmap" object.
}
\description{
Return the row or column ordering present in a heatmap
}
