% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{get_income}
\alias{get_income}
\title{Returns combined household incomes split into low, medium, and high groups}
\usage{
get_income(parent_demographics)
}
\arguments{
\item{parent_demographics}{x}
}
\value{
income_df x
}
\description{
Low: $0 - $50k, Medium: $50k - $100k, High: > $100k
}
