% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions_df.R
\name{get_mean_pval}
\alias{get_mean_pval}
\title{Get mean p-value}
\usage{
get_mean_pval(sol_df_row)
}
\arguments{
\item{sol_df_row}{row of sol_df object}
}
\value{
mean_pval mean p-value
}
\description{
Given an solutions data frame row containing evaluated p-values, returns mean.
}
\keyword{internal}
