% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions_matrix.R
\name{get_min_pval}
\alias{get_min_pval}
\title{Get minimum p-value}
\usage{
get_min_pval(solutions_matrix_row)
}
\arguments{
\item{solutions_matrix_row}{row of solutions_matrix object}
}
\value{
min_pval minimum p-value
}
\description{
Given an solutions matrix row containing evaluated p-values, returns min.
}
