% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subjects.R
\name{get_mtbi_subjects}
\alias{get_mtbi_subjects}
\title{Extract mTBI subjects with a minimum time-since-last-mtbi threshold}
\usage{
get_mtbi_subjects(tbi_df, min_latest_mtbi_mpi)
}
\arguments{
\item{tbi_df}{A TBI dataframe}

\item{min_latest_mtbi_mpi}{The minimum time-since-last-mtbi to be selected}
}
\value{
subjects subjectkey values of selected subjects
}
\description{
Extract mTBI subjects with a minimum time-since-last-mtbi threshold
}
