% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{get_pubertal_status}
\alias{get_pubertal_status}
\title{Return dataframe containing pubertal status of specified subjects}
\usage{
get_pubertal_status(ssphp01, ssphy01, subjects)
}
\arguments{
\item{ssphp01}{Dataframe containing parent pubertal status report}

\item{ssphy01}{Dataframe containing youth pubertal status report}

\item{subjects}{Dataframe containing list of required subjects}
}
\value{
pubertal_status Dataframe containing average pubertal status
}
\description{
Return dataframe containing pubertal status of specified subjects
}
