% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_representative_solutions.R
\name{get_representative_solutions}
\alias{get_representative_solutions}
\title{Extract representative solutions from a matrix of ARIs}
\usage{
get_representative_solutions(
  aris,
  split_vector,
  order = NULL,
  sol_df,
  filter_fn = NULL
)
}
\arguments{
\item{aris}{Matrix of adjusted rand indices from \code{calc_aris()}}

\item{split_vector}{A vector of partition indices.}

\item{order}{Numeric vector indicating row ordering of solutions data frame.}

\item{sol_df}{Output of \code{batch_snf} containing cluster solutions.}

\item{filter_fn}{Optional function to filter the meta-cluster by prior to
maximum average ARI determination. This can be useful if you are explicitly
trying to select a solution that meets a certain condition, such as only
picking from the 4 cluster solutions within a meta cluster. An example
valid function could be \code{fn <- function(x) x[x$"nclust" == 4, ]}.}
}
\value{
The provided solutions data frame reduced to just one row per meta
cluster defined by the split vector.
}
\description{
Following clustering with \code{batch_snf}, a matrix of pairwise ARIs that show
how related each cluster solution is to each other can be generated by the
\code{calc_aris} function. Partitioning of the ARI matrix can be done by
visual inspection of \code{meta_cluster_heatmap()} results or by
\code{shiny_annotator}. Given the indices of meta cluster boundaries, this
function will return a single representative solution from each meta cluster
based on maximum average ARI to all other solutions within that meta
cluster.
}
