% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{get_screen_time}
\alias{get_screen_time}
\title{Extract screen time}
\usage{
get_screen_time(stq01, abcd_stq01, subjects)
}
\arguments{
\item{stq01}{ABCD Parent Screen Time Survey}

\item{abcd_stq01}{ABCD Youth Screen Time Survey}

\item{subjects}{Dataframe containing list of required subjects}
}
\value{
screen_time
}
\description{
Extract screen time
}
