% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_metrics_list.R
\name{gower_distance}
\alias{gower_distance}
\title{Distance metric: Gower distance}
\usage{
gower_distance(df, weights_row)
}
\arguments{
\item{df}{Dataframe containing at least 1 data column.}

\item{weights_row}{For compatibility - function does not accept weights.}
}
\value{
distance_matrix A distance matrix.
}
\description{
Distance metric: Gower distance
}
