% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{keep_split}
\alias{keep_split}
\title{Filter data to training or testing subjects only}
\usage{
keep_split(df, assigned_df, split)
}
\arguments{
\item{df}{Dataframe to be subsetted into training or testing split}

\item{assigned_df}{Dataframe containing "subjectkey" and "split" cols}

\item{split}{Split to keep ("train" or "test")}
}
\value{
split_df
}
\description{
Given a dataframe, assigned_df object (from train_test_split()), and split,
return just the data for subjects that were assigned the specified split
}
