% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_remove}
\alias{list_remove}
\title{Remove items from a data list}
\usage{
list_remove(dl, ...)
}
\arguments{
\item{dl}{The data list containing components to be removed}

\item{...}{Any number of components to remove from the list object, passed as
strings}
}
\value{
A "list"-class object in which any specified elements have been
removed.
}
\description{
Removes specified elements from a provided data list
}
