% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{lp_row}
\alias{lp_row}
\title{(DEPRECATED) Function for label propagating results in a solutions_matrix}
\usage{
lp_row(
  solutions_matrix,
  full_data_list,
  clust_algs_list = NULL,
  distance_metrics_list = NULL,
  weights_matrix = NULL
)
}
\arguments{
\item{solutions_matrix}{A solutions_matrix (training set only)}

\item{full_data_list}{A data_list containing training and testing subjects}

\item{clust_algs_list}{The clustering algorithms list used to create the
original solutions matrix (if any was used)}

\item{distance_metrics_list}{The distance metrics list used to create the
original solutions matrix (if any was used)}

\item{weights_matrix}{The weights matrix used to create the original
solutions matrix (if any was used)}
}
\description{
(DEPRECATED) Function for label propagating results in a solutions_matrix
}
