% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_plot.R
\name{manhattan_plot}
\alias{manhattan_plot}
\title{Manhattan plot of p-values}
\usage{
manhattan_plot(pvals_df)
}
\arguments{
\item{pvals_df}{Output of the pval_select() function}
}
\description{
Manhattan plot of p-values
}
