% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_cluster.R
\name{mc_heatmap}
\alias{mc_heatmap}
\title{Outdated adjusted_rand_index_heatmap}
\usage{
mc_heatmap(mc_results, save = NULL)
}
\arguments{
\item{mc_results}{outdated}

\item{save}{outdated}
}
\description{
Outdated adjusted_rand_index_heatmap
}
