% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_df_list}
\alias{merge_df_list}
\title{Merge list of data frames into a single data frame}
\usage{
merge_df_list(df_list, join = "inner", uid = "uid", no_na = FALSE)
}
\arguments{
\item{df_list}{list of data frames.}

\item{join}{String indicating if join should be "inner" or "full".}

\item{uid}{Column name to join on. Default is "uid".}

\item{no_na}{Whether to remove NA values from the merged data frame.}
}
\value{
Inner join of all data frames in list.
}
\description{
This helper function combines all data frames in a single-level list into a
single data frame.
}
\examples{
merge_df_list(list(income, pubertal), uid = "unique_id")
}
