% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{metasnf_warning}
\alias{metasnf_warning}
\title{Helper function for raising warnings}
\usage{
metasnf_warning(..., env = 1)
}
\arguments{
\item{...}{Arbitrary number of strings to be pasted together into warning
message.}

\item{env}{Environment to evaluate expressions in.}
}
\value{
Returns no value. Raises a warning through cli::cli_warn
}
\description{
Helper function for raising warnings
}
\keyword{internal}
