% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_fns_list.R
\name{new_clust_fns_list}
\alias{new_clust_fns_list}
\title{Constructor for \code{clust_fns_list} class object}
\usage{
new_clust_fns_list(cfll)
}
\arguments{
\item{cfll}{A clust_fns_list-like \code{list} class object.}
}
\value{
A \code{clust_fns_list} class object.
}
\description{
Constructor for \code{clust_fns_list} class object
}
\keyword{internal}
