% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mats_list.R
\name{new_sim_mats_list}
\alias{new_sim_mats_list}
\title{Constructor for \code{similarity_matrix_list} class object}
\usage{
new_sim_mats_list(smll)
}
\arguments{
\item{smll}{A similarity matrix list-like object.}
}
\value{
A \code{similarity_matrix_list} class object.
}
\description{
Constructor for \code{similarity_matrix_list} class object
}
\keyword{internal}
