% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions_df.R
\name{new_solutions_df}
\alias{new_solutions_df}
\title{Constructor for \code{solutions_df} class object}
\usage{
new_solutions_df(sol_dfl)
}
\arguments{
\item{sol_dfl}{A solutions data frame-like object to be validated and
converted into a solutions data frame.}
}
\value{
A \code{solutions_df} class object.
}
\description{
Constructor for \code{solutions_df} class object
}
