% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{no_subs}
\alias{no_subs}
\title{Select all columns of a dataframe not starting with a given string prefix.}
\usage{
no_subs(df)
}
\arguments{
\item{df}{Datframe}
}
\value{
df_no_subs Dataframe without subjects
}
\description{
Originally constructed to return a dataframe without any subject columns.
Returns a dataframe excluding all columns with a specified string prefix.
}
