% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf.R
\name{no_subs}
\alias{no_subs}
\title{No subjects}
\usage{
no_subs(df)
}
\arguments{
\item{df}{Datframe}
}
\value{
df_no_subs Dataframe without subjects
}
\description{
Return a dataframe without any subject columns
}
