% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{no_subs}
\alias{no_subs}
\title{Select all columns of a dataframe not starting with the 'subject_' prefix.}
\usage{
no_subs(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
df_no_subs Dataframe without subjects
}
\description{
Removes the 'subject_' prefixed columns from a dataframe. Useful for printing
solutions_matrix structures to the console.
}
