% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{numcol_to_numeric}
\alias{numcol_to_numeric}
\title{Convert dataframe columns to numeric type}
\usage{
numcol_to_numeric(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
df The dataframe with all possible columns converted to type numeric
}
\description{
Converts all columns in a dataframe that can be converted to numeric type to
numeric type.
}
