% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archived_functions.R
\name{om_scatter}
\alias{om_scatter}
\title{Scatter plot solutions matrix results by minimum and mean p-values}
\usage{
om_scatter(
  om,
  title = NULL,
  pointsize = 5,
  labelsize = 5,
  fontsize = 20,
  cluster_colours = 6,
  xlim = NULL,
  ylim = NULL,
  save = NULL
)
}
\arguments{
\item{om}{solutions matrix}

\item{title}{plot title}

\item{pointsize}{size of scatter plot point}

\item{labelsize}{size of label text}

\item{fontsize}{plot fontsize}

\item{cluster_colours}{use a colour palette built for the specified number
of clusters. If generating multiple graphs with varying numbers of clusters,
specifying this parameter will ensure the cluster count and colour pairs
stay constant.}

\item{xlim}{x-axis limits for log scale}

\item{ylim}{y-axis limits for log scale}

\item{save}{optional path to save figure to}
}
\description{
Scatter plot solutions matrix results by minimum and mean p-values
}
