% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{om_scatter}
\alias{om_scatter}
\title{Scatter plot alpha and k hyperparameter results by minimum and mean p-values}
\usage{
om_scatter(om, title = NULL, fontsize = 20, save = NULL)
}
\arguments{
\item{om}{output matrix}

\item{title}{plot title}

\item{fontsize}{plot fontsize}

\item{save}{optional path to save figure to}
}
\description{
Scatter plot alpha and k hyperparameter results by minimum and mean p-values
}
