% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions_df.R
\name{ord_reg_pval}
\alias{ord_reg_pval}
\title{Ordinal regression p-value}
\usage{
ord_reg_pval(predictor, response)
}
\arguments{
\item{predictor}{A categorical or numeric feature.}

\item{response}{A numeric feature.}
}
\value{
pval A p-value (class "numeric").
}
\description{
Returns the overall p-value of an ordinal regression on a categorical
predictor and response vetors. If the ordinal response
}
\keyword{internal}
