% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{p_val_select}
\alias{p_val_select}
\title{(DEPRECATED) Select p-values from solutions matrix
Replaced with "pval_select'}
\usage{
p_val_select(solutions_matrix)
}
\arguments{
\item{solutions_matrix}{The output of batch_snf}
}
\value{
p_val_matrix P-values ready for heatmap plotting
}
\description{
(DEPRECATED) Select p-values from solutions matrix
Replaced with "pval_select'
}
