% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_matrix.R
\name{p_val_select}
\alias{p_val_select}
\title{Select p-values from output matrix}
\usage{
p_val_select(output_matrix)
}
\arguments{
\item{output_matrix}{The output of batch_snf}
}
\value{
p_val_matrix P-values ready for heatmap plotting
}
\description{
Select p-values from output matrix
}
