% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_snf.R
\name{parallel_batch_snf}
\alias{parallel_batch_snf}
\title{Parallel processing form of batch_snf}
\usage{
parallel_batch_snf(data_list, design_matrix, processes)
}
\arguments{
\item{data_list}{nested list of input data generated by the function
\code{get_data_list()}}

\item{design_matrix}{matrix indicating parameters to iterate SNF through}

\item{processes}{Number of parallel processes used when executing SNF}
}
\value{
populated_design_matrix design matrix with filled columns related to
subtype membership
}
\description{
Parallel processing form of batch_snf
}
