% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archived_functions.R
\name{plot_all_cbcl}
\alias{plot_all_cbcl}
\title{Plot clusters across all CBCL measures}
\usage{
plot_all_cbcl(
  om,
  cbcl_list,
  fig_path_fn,
  save_prefix = NULL,
  save_suffix = NULL,
  include = NULL,
  exclude = NULL,
  w = 25,
  h = 20
)
}
\arguments{
\item{om}{An solutions matrix-like structure}

\item{cbcl_list}{List containing all CBCL dataframes}

\item{fig_path_fn}{Closure specifying the location to save the final figure}

\item{save_prefix}{prefix to add to each file saved}

\item{save_suffix}{suffix to add to each file saved}

\item{include}{a string or list of strings specifying which CBCL measures
should be included. All other CBCL measures will be excluded.}

\item{exclude}{a string or list of strings specifying which CBCL measures
should be excluded. All other CBCL measures will be excluded.}

\item{w}{the width of the final saved plots}

\item{h}{the height of the final saved plots}
}
\description{
Arrange and optionally save a grid of bar charts with cluster on the x-axis
and each CBCL measure on the y-axis.
}
