% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_all_cbcl}
\alias{plot_all_cbcl}
\title{Plot clusters across all CBCL measures}
\usage{
plot_all_cbcl(cluster_df, sig, cbcl_list, save = NULL)
}
\arguments{
\item{cluster_df}{a dataframe containing cluster and subjectkey}

\item{sig}{a name for the current set of plots}

\item{cbcl_list}{List containing all CBCL dataframes}

\item{save}{optional path to save figure to}
}
\description{
Arrange and optionally save a grid of bar charts with cluster on the x-axis
and each CBCL measure on the y-axis.
}
