% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{pooled_cocluster_heatmap}
\alias{pooled_cocluster_heatmap}
\title{(DEPRECATED) Coclustering heatmap}
\usage{
pooled_cocluster_heatmap(...)
}
\arguments{
\item{...}{Function arguments}
}
\description{
(DEPRECATED) Coclustering heatmap
}
