% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{prefix_dl_sk}
\alias{prefix_dl_sk}
\title{Remove NAs from a data_list object}
\usage{
prefix_dl_sk(data_list)
}
\arguments{
\item{data_list}{A data_list}
}
\value{
data_list A data_list without NAs
}
\description{
Remove NAs from a data_list object
}
