% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{prefix_dl_sk}
\alias{prefix_dl_sk}
\title{Add "subject_" prefix to all UID values in subjectkey column}
\usage{
prefix_dl_sk(dl)
}
\arguments{
\item{dl}{A nested list of input data from \code{generate_data_list()}.}
}
\value{
dl A data list with UIDs prefixed with the string "subject_"
}
\description{
Add "subject_" prefix to all UID values in subjectkey column
}
