% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{prefix_dl_sk}
\alias{prefix_dl_sk}
\title{Add "subject_" prefix to all UID values in subjectkey column}
\usage{
prefix_dl_sk(data_list)
}
\arguments{
\item{data_list}{A nested list of input data from \code{generate_data_list()}.}
}
\value{
data_list A data_list without NAs
}
\description{
Add "subject_" prefix to all UID values in subjectkey column
}
