% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.solutions_df}
\alias{print.solutions_df}
\title{Print method for class \code{solutions_df}}
\usage{
\method{print}{solutions_df}(x, n = NULL, tips = TRUE, ...)
}
\arguments{
\item{x}{A \code{weights_matrix} class object.}

\item{n}{Number of rows to print, passed into \code{tibble::print.tbl_df()}.}

\item{tips}{If TRUE, include lines on how to print more rows / transposed.}

\item{...}{Other arguments passed to \code{print} (not used in this function).}
}
\value{
Function prints to console but does not return any value.
}
\description{
Custom formatted print for weights matrices that outputs
information about feature weights functions to the console.
}
