% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf.R
\name{print_data_list_dims}
\alias{print_data_list_dims}
\title{Print data_list nested dims}
\usage{
print_data_list_dims(data_list)
}
\arguments{
\item{data_list}{The data_list object to be reduced}
}
\description{
Print data_list nested dims
}
