% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions_matrix.R
\name{pval_summaries}
\alias{pval_summaries}
\title{Add minimum and mean p-values to an extended solutions matrix}
\usage{
pval_summaries(solutions_matrix)
}
\arguments{
\item{solutions_matrix}{A solutions_matrix object that already has some
p-value columns included.}
}
\description{
Add minimum and mean p-values to an extended solutions matrix
}
