% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{pvals_pheatmap}
\alias{pvals_pheatmap}
\title{Pheatmap a matrix of p-values}
\usage{
pvals_pheatmap(pvals, save = NULL, reverse_colours = FALSE)
}
\arguments{
\item{pvals}{a matrix of p-values}

\item{save}{optional path to save figure to}

\item{reverse_colours}{boolean to invert colours}
}
\description{
Pheatmap a matrix of p-values
}
