% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{pvals_pheatmap}
\alias{pvals_pheatmap}
\title{Pheatmap a matrix of p-values}
\usage{
pvals_pheatmap(
  pvals,
  order = NULL,
  cluster_cols = TRUE,
  show_rownames = FALSE,
  save = NULL,
  reverse_colours = FALSE
)
}
\arguments{
\item{pvals}{a matrix of p-values}

\item{order}{numeric vector indicating row ordering of design matrix}

\item{cluster_cols}{if TRUE, pheatmap will cluster (and rearrange) columns}

\item{show_rownames}{If TRUE (default), rownames are shown on heatmap}

\item{save}{optional path to save figure to}

\item{reverse_colours}{boolean to invert colours}
}
\description{
Pheatmap a matrix of p-values
}
