% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf.R
\name{random_removal}
\alias{random_removal}
\title{Generate random sequence for column exclusions component of design matrix}
\usage{
random_removal(num_cols)
}
\arguments{
\item{num_cols}{Number of feature columns in consideration for exclusion}
}
\value{
shuffled_removals Binary vector sequence indicating if a column
should be included (1) or excluded (0)
}
\description{
Helper function to contribute to rows within the design matrix.
Number of columns removed follows the exponential probability distribution
to typically keep all or most columns.
}
