% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_matrix.R
\name{random_removal}
\alias{random_removal}
\title{Generate random removal sequence}
\usage{
random_removal(num_cols, min_removed, max_removed)
}
\arguments{
\item{num_cols}{Number of feature elements in consideration for exclusion}

\item{min_removed}{The smallest number of elements that may be removed}

\item{max_removed}{The largest number of elements that may be removed}
}
\value{
shuffled_removals Binary vector sequence indicating if a column
should be included (1) or excluded (0)
}
\description{
Helper function to contribute to rows within the design matrix.
Number of columns removed follows the exponential probability distribution
to typically keep all or most columns.
}
