% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf.R
\name{reduce_dl_to_common}
\alias{reduce_dl_to_common}
\title{Given a \code{data_list} object, reduce each nested dataframe to contain only the
set of subjects that are shared by all nested dataframes}
\usage{
reduce_dl_to_common(data_list)
}
\arguments{
\item{data_list}{The data_list object to be reduced}
}
\value{
reduced_data_list The data_list object subsetted only to subjectssnf
shared across all nested dataframes
}
\description{
Given a \code{data_list} object, reduce each nested dataframe to contain only the
set of subjects that are shared by all nested dataframes
}
