% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{remove_dl_na}
\alias{remove_dl_na}
\title{Remove NAs from a data list object}
\usage{
remove_dl_na(dl)
}
\arguments{
\item{dl}{A nested list of input data from \code{data_list()}.}
}
\value{
dl A data list without NAs
}
\description{
Remove NAs from a data list object
}
\keyword{internal}
