% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_diagonals}
\alias{scale_diagonals}
\title{Adjust the diagonals of a matrix}
\usage{
scale_diagonals(matrix, method = "mean")
}
\arguments{
\item{matrix}{Matrix to rescale.}

\item{method}{Method of rescaling. Can be:
\itemize{
\item "mean" (replace diagonals with average value of off-diagonals)
\item "zero" (replace diagonals with 0)
\item "min" (replace diagonals with min value of off-diagonals)
\item "max" (replace diagonals with max value of off-diagonals)
}}
}
\value{
A "matrix" class object with rescaled diagonals.
}
\description{
Adjust the diagonals of a matrix to reduce contrast with off-diagonals
during plotting.
}
