% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{settings_matrix_heatmap}
\alias{settings_matrix_heatmap}
\title{Heatmap settings matrix based on meta-clustering results}
\usage{
settings_matrix_heatmap(
  settings_matrix,
  order = NULL,
  show_rownames = TRUE,
  save = NULL,
  hide_ids = FALSE
)
}
\arguments{
\item{settings_matrix}{matrix indicating parameters to iterate SNF through}

\item{order}{numeric vector indicating row ordering of settings matrix}

\item{show_rownames}{If TRUE (default), rownames are shown on heatmap}

\item{save}{optional path to save figure to}

\item{hide_ids}{boolean indicating if row_id numbers should be hidden}
}
\description{
Normalizes the settings matrix and plots as a heatmap. Rows are reordered to
match the row-clustering present within a provided meta-clustering result.
}
