% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{settings_matrix_row_fn}
\alias{settings_matrix_row_fn}
\title{Apply-based function for batch_snf}
\usage{
settings_matrix_row_fn(settings_matrix_row, dl)
}
\arguments{
\item{settings_matrix_row}{a row of a settings matrix}

\item{dl}{a data list}
}
\value{
solutions_matrix_row the corresponding solutions_matrix row
}
\description{
Apply-based function for batch_snf
}
