% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{shiny_annotator}
\alias{shiny_annotator}
\title{Launch shiny app to identify meta cluster boundaries}
\usage{
shiny_annotator(ari_heatmap)
}
\arguments{
\item{ari_heatmap}{Heatmap of ARIs to divide into meta clusters.}
}
\value{
Does not return any value. Launches interactive shiny applet.
}
\description{
Launch shiny app to identify meta cluster boundaries
}
