% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf_config.R
\name{snf_config}
\alias{snf_config}
\title{Define configuration for generating a set of SNF-based cluster solutions}
\usage{
snf_config(nrows)
}
\arguments{
\item{nrows}{Number of rows to generate for the settings matrix.}
}
\value{
An \code{snf_config} class object.
}
\description{
\code{snf_config()} constructs an SNF config object which inherits from classes
\code{snf_config} and \code{list}. This object is used to store all settings
required to transform data stored in a \code{data_list} class object into a
space of cluster solutions by SNF. The SNF config object contains the
following components:
1. A settings data frame (inherits from \code{settings_df} and \code{data.frame}).
Data frame that stores SNF-specific hyperparameters and information
about feature selection and weighting, SNF schemes, clustering
algorithms, and distance metrics. Each row of the settings data frame
corresponds to a distinct cluster solution.
2. A clustering algorithms list (inherits from \code{clust_algs_list} and
\code{list}), which stores all clustering algorithms that the settings
data frame can point to.
3. A distance metrics list (inherits from \code{dist_metrics_list} and
\code{list}), which stores all distance metrics that the settings data
frame can point to.
4. A weights matrix (inherits from \code{weights_matrix}, \code{matrix}, and
\code{array}'), which stores the feature weights to use prior to distance
calculations. Each column of the weights matrix corresponds to a
different feature in the data list and each row corresponds to a
different row in the settings data frame.
}
