% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf.R
\name{snf_step}
\alias{snf_step}
\title{Given 'data_list' object and scheme, return SNF results}
\usage{
snf_step(data_list, scheme)
}
\arguments{
\item{data_list}{nested list of input data generated by the function
\code{get_data_list()}}

\item{scheme}{Which SNF system to use to achieve the final fused network}
}
\value{
df_list This will need to become the SNF's results
}
\description{
Given 'data_list' object and scheme, return SNF results
}
