% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf_schemes.R
\name{snf_step}
\alias{snf_step}
\title{SNF a data_list}
\usage{
snf_step(data_list, scheme, K = 20, alpha = 0.5, t = 20)
}
\arguments{
\item{data_list}{nested list of input data generated by the function
\code{get_data_list()}}

\item{scheme}{Which SNF system to use to achieve the final fused network}

\item{K}{K hyperparameter}

\item{alpha}{alpha/eta/sigma hyperparameter}

\item{t}{SNF number of iterations hyperparameter}
}
\value{
fused_network The final fused network for clustering
}
\description{
SNF a data_list
}
