% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_snf.R
\name{snf_step}
\alias{snf_step}
\title{Convert a data list to a similarity matrix through a variety of SNF schemes}
\usage{
snf_step(
  data_list,
  scheme,
  k = 20,
  alpha = 0.5,
  t = 20,
  cont_dist_fn,
  disc_dist_fn,
  ord_dist_fn,
  cat_dist_fn,
  mix_dist_fn,
  weights_row
)
}
\arguments{
\item{data_list}{A nested list of input data from \code{generate_data_list()}.}

\item{scheme}{Which SNF system to use to achieve the final fused network.}

\item{k}{k hyperparameter.}

\item{alpha}{alpha/eta/sigma hyperparameter.}

\item{t}{SNF number of iterations hyperparameter.}

\item{cont_dist_fn}{distance metric function for continuous data.}

\item{disc_dist_fn}{distance metric function for discrete data.}

\item{ord_dist_fn}{distance metric function for ordinal data.}

\item{cat_dist_fn}{distance metric function for categorical data.}

\item{mix_dist_fn}{distance metric function for mixed data.}

\item{weights_row}{dataframe row containing feature weights.}
}
\value{
fused_network The final fused network (class "matrix", "array")
generated by SNF.
}
\description{
Convert a data list to a similarity matrix through a variety of SNF schemes
}
