% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions_df.R
\name{solutions_df}
\alias{solutions_df}
\title{Constructor for \code{solutions_df} class object}
\usage{
solutions_df(sol_dfl, smll, sc, dl)
}
\arguments{
\item{sol_dfl}{A solutions data frame-like object to be validated and
converted into a solutions data frame.}

\item{smll}{A similarity matrix list-like object to be validated and used
to construct a solutions data frame.}

\item{sc}{An \code{snf_config} object used to construct a solutions data frame.}

\item{dl}{An \code{data_list} object used to construct a solutions data frame.}
}
\value{
A solutions data frame (\code{solutions_df} class object).
}
\description{
Constructor for \code{solutions_df} class object
}
\keyword{internal}
