% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_algorithms.R
\name{spectral_eigen}
\alias{spectral_eigen}
\title{Clustering algorithm: Spectral clustering with eigen-gap heuristic}
\usage{
spectral_eigen(affinity_matrix)
}
\arguments{
\item{affinity_matrix}{An affinity matrix}
}
\value{
solution A vector indicating which cluster each patient was assigned
to
}
\description{
Applies spectral clustering to affinity matrix. Number of clusters is based
on the eigen-gap heuristic.
}
