% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_algs_list.R
\name{spectral_eigen_classic}
\alias{spectral_eigen_classic}
\title{Clustering algorithm: Spectral clustering with eigen-gap heuristic}
\usage{
spectral_eigen_classic(similarity_matrix)
}
\arguments{
\item{similarity_matrix}{A similarity matrix.}
}
\value{
solution_data A list storing cluster assignments ("solution") and
the number of clusters ("nclust").
}
\description{
Applies spectral clustering to similarity matrix. Number of clusters is based
on the eigen-gap heuristic. Range of possible cluster solutions is fixed
between 2 and 5 inclusive.
}
