% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_algs_list.R
\name{spectral_four}
\alias{spectral_four}
\title{Clustering algorithm: Spectral clustering for a four cluster solution}
\usage{
spectral_four(similarity_matrix)
}
\arguments{
\item{similarity_matrix}{A similarity matrix.}
}
\value{
solution_data A list storing cluster assignments ("solution") and
the number of clusters ("nclust").
}
\description{
Applies spectral clustering to similarity matrix. Seeks four clusters.
}
