% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_algorithms.R
\name{spectral_seven}
\alias{spectral_seven}
\title{Clustering algorithm: Spectral clustering for a seven cluster solution}
\usage{
spectral_seven(similarity_matrix)
}
\arguments{
\item{similarity_matrix}{A similarity matrix}
}
\value{
solution_data A list storing cluster assignments and the number of
clusters.
}
\description{
Applies spectral clustering to similarity matrix. Seeks seven clusters.
}
