% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subs}
\alias{subs}
\title{Select all columns of a dataframe starting with a given string prefix.}
\usage{
subs(df)
}
\arguments{
\item{df}{Dataframe}
}
\value{
df_subs Dataframe without subjects
}
\description{
Originally constructed to return a dataframe with only subject columns.
Returns a dataframe including only columns with a specified string prefix.
}
