% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coclustering.R
\name{subsample_data_list}
\alias{subsample_data_list}
\title{Create subsamples of a data_list}
\usage{
subsample_data_list(
  data_list,
  n_subsamples,
  subsample_fraction = NULL,
  n_subjects = NULL
)
}
\arguments{
\item{data_list}{A nested list of input data from \code{generate_data_list()}.}

\item{n_subsamples}{Number of subsamples to create.}

\item{subsample_fraction}{Percentage of patients to include per subsample.}

\item{n_subjects}{Number of patients to include per subsample.}
}
\value{
A "list" class object containing \code{n_subsamples} number of
data lists. Each of those data lists contains a random \code{subsample_fraction}
fraction of the observations of the provided data list.
}
\description{
Given a data list, return a list of smaller data lists that are generated
through random sampling (without replacement).
}
