% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_measures.R
\name{subsample_pairwise_aris}
\alias{subsample_pairwise_aris}
\title{Calculate pairwise adjusted Rand indices across subsamples of data}
\usage{
subsample_pairwise_aris(data_list_subsamples, settings_matrix)
}
\arguments{
\item{data_list_subsamples}{A list of subsets of a larger data_list. See
?subsample_data_list to obtain this object.}

\item{settings_matrix}{A settings_matrix.}
}
\value{
pairwise_ari_df Dataframe containing a column for row (which row of
the original settings_matrix the rest of the information corresponds to),
mean_ari (the average adjusted Rand Index across all subsamples), and
ari_sd (the standard deviation of the adjusted Rand Indices).
}
\description{
Calculate pairwise adjusted Rand indices across subsamples of data
}
