% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coclustering.R
\name{subsample_pairwise_aris}
\alias{subsample_pairwise_aris}
\title{Calculate pairwise adjusted Rand indices across subsamples of data}
\usage{
subsample_pairwise_aris(
  subsample_solutions,
  return_raw_aris = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{subsample_solutions}{A list of containing cluster solutions from
distinct subsamples of the data. This object is generated by the function
\code{batch_snf_subsamples()}.}

\item{return_raw_aris}{Whether the ARI matrix used to calculate the average
ARI across subsamples should be returned.}

\item{verbose}{If TRUE, print time remaining estimates to console.}
}
\value{
If return_raw_aris is FALSE, this function will return
}
\description{
Calculate pairwise adjusted Rand indices across subsamples of data
}
