% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_algs_list.R
\name{summarize_clust_algs_list}
\alias{summarize_clust_algs_list}
\title{Summarize a clust_algs_list object}
\usage{
summarize_clust_algs_list(clust_algs_list)
}
\arguments{
\item{clust_algs_list}{A clust_algs_list object}
}
\value{
summary_df "data.frame" class object containing the name and index
of each clustering algorithm in te provided \code{clust_algs_list}.
}
\description{
Summarize a clust_algs_list object
}
