% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{summarize_dfl}
\alias{summarize_dfl}
\title{Summarize a distance functions list}
\usage{
summarize_dfl(dfl)
}
\arguments{
\item{dfl}{A dist_fns_list.}
}
\value{
"data.frame"-class object summarizing items in a distance metrics
list.
}
\description{
Summarize a distance functions list
}
\keyword{internal}
