% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{summarize_dl}
\alias{summarize_dl}
\title{Summarize a data list}
\usage{
summarize_dl(data_list, scope = "component")
}
\arguments{
\item{data_list}{A nested list of input data from \code{data_list()}.}

\item{scope}{The level of detail for the summary. Options are:
\itemize{
\item "component" (default): One row per component (data frame) in the data list.
\item "feature": One row for each feature in the data list.
}}
}
\value{
data.frame class object summarizing all components (or features if
scope == "component").
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Defunct function for summarizing a data list. Please
use \code{summary()} instead.
}
