% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_metrics_list.R
\name{summarize_dml}
\alias{summarize_dml}
\title{Summarize metrics contained in a distance_metrics_list}
\usage{
summarize_dml(distance_metrics_list)
}
\arguments{
\item{distance_metrics_list}{A distance_metrics_list.}
}
\value{
"data.frame"-class object summarizing items in a distance metrics
list.
}
\description{
Summarize metrics contained in a distance_metrics_list
}
