% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions_df.R
\name{summarize_pvals}
\alias{summarize_pvals}
\title{Summarize p-value columns of an extended solutions data frame}
\usage{
summarize_pvals(ext_sol_df)
}
\arguments{
\item{ext_sol_df}{Result of \code{extend_solutions}}
}
\value{
The provided extended solutions data frame along with columns for
the min, mean, and maximum across p-values for each row.
}
\description{
Summarize p-value columns of an extended solutions data frame
}
\keyword{internal}
