% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{summarize_target_list}
\alias{summarize_target_list}
\title{(DEPRECATED) Summarize target list}
\usage{
summarize_target_list(...)
}
\arguments{
\item{...}{Function arguments}
}
\description{
(DEPRECATED) Summarize target list
}
