% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf_schemes.R
\name{two_step_merge}
\alias{two_step_merge}
\title{Two step SNF}
\usage{
two_step_merge(data_list, k = 20, alpha = 0.5, t = 20)
}
\arguments{
\item{data_list}{nested list of input data generated by the function
\code{get_data_list()}}

\item{k}{k hyperparameter}

\item{alpha}{alpha/eta/sigma hyperparameter}

\item{t}{SNF number of iterations hyperparameter}
}
\value{
fused_network The final fused network for clustering
}
\description{
Individual dataframes into individual similarity matrices into one fused
network per domain into one final fused network.
}
