% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{validate_data_list}
\alias{validate_data_list}
\title{Validator for data_list class object}
\usage{
validate_data_list(dll)
}
\arguments{
\item{dll}{A data list-like \code{list} class object.}
}
\value{
If dll has a valid structure for a \code{data_list} class object,
returns the input unchanged. Otherwise, raises an error.
}
\description{
Validator for data_list class object
}
\keyword{internal}
