% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_fns_list.R
\name{validate_dist_fns_list}
\alias{validate_dist_fns_list}
\title{Validator for dist_fns_list class object}
\usage{
validate_dist_fns_list(dfll)
}
\arguments{
\item{dfll}{A distance metrics list-like list object to be validated.}
}
\value{
If dfll has a valid structure for a \code{dist_fns_list} class
object, returns the input unchanged. Otherwise, raises an error.
}
\description{
Validator for dist_fns_list class object
}
\keyword{internal}
