% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions_df.R
\name{validate_solutions_df}
\alias{validate_solutions_df}
\title{Validator for \code{solutions_df} class object}
\usage{
validate_solutions_df(sol_dfl)
}
\arguments{
\item{sol_dfl}{A solutions data frame-like object to be validated and
converted into a solutions data frame.}
}
\value{
If sol_dfl has a valid structure for a \code{solutions_df} class object,
returns the input unchanged. Otherwise, raises an error.
}
\description{
Validator for \code{solutions_df} class object
}
